Add a Layer of Data (internal use)

This function adds an additional layer of data along the third dimension to an existing 3D array (or 2D matrix/vector) by stacking a new layer of data. It ensures that the dimensions of the new layer match the first two dimensions of the existing array or matrix. If there is a mismatch in row or column names and the force parameter is set to TRUE, the function will align the data by filling missing values with NA. It is used internally and not intended for routine user use.

Usage

add_layer(current, layer, layer_name, force = FALSE)

Arguments

current

A vector, matrix, or 3D array representing the current stack of data.

layer

A matrix or array that represents the new layer of data to be added. It should match the dimensions of the first two dimensions of current.

layer_name

A character string specifying the name of the new dimension for the 3rd axis. This can be used to annotate the new data layer.

force

A logical value indicating whether to force the join and create NA values where row or column names do not match between current and layer. Default is FALSE.

Value

A 3D array with the added layer in the third dimension.

Examples

# Example: Adding a new data layer to an existing 3D array
current_stack <- array(1:12,
                       dim = c(3, 4, 1),
                       dimnames = list(c(1:3), c(1:4), c(1)))
new_layer <- matrix(13:16, nrow = 3, ncol = 4)
rownames(new_layer) <- 1:3
colnames(new_layer) <- 1:4
updated_stack <- add_layer(current    = current_stack,
                           layer      = new_layer,
                           layer_name = "Layer2",
                           force      = FALSE)
dim(updated_stack)
#> [1] 3 4 2

# Example with mismatched row/column names (force = TRUE)
rownames(current_stack) <- c("A", "B", "C")
rownames(new_layer) <- c("B", "C", "D")
updated_stack_with_force <- add_layer(current    = current_stack,
                                      layer      = new_layer,
                                      layer_name = "Layer2",
                                      force      = TRUE)
dim(updated_stack_with_force)
#> [1] 4 4 2