Annotate Features with Common IDs

This function annotates a data.table by mapping feature identifiers (id_col) to standardized annotations. Optionally, pathway and platform information can be included.

Usage

annotate_features(features, id_col, pathway_col = NULL, platform_col = NULL)

Arguments

features

data.table The dataset containing features to be annotated.

id_col

character The name of the column containing feature IDs. This can be a column with identifiers such as comp_ids, chem_ids, or identifiers from Nightingale Excel or flat text files.

pathway_col

character, optional The name of the column containing pathway information. If provided, source pathway data from features will be included in the annotation with standardised column name pathway.

platform_col

character, optional The name of the column specifying the platform type. If supplied, source platform data from features will be included in the annotation with standardised column name platform.

Value

data.table A modified version of features with annotations added based on the provided IDs.

Examples

# Example dataset
features <- data.table::data.table(comp_ids = c("63042", "62919", "62921"))

# Annotate features using 'comp_ids'
annotated_features <- annotate_features(features, id_col = "comp_ids")

# Annotate with additional pathway and platform information
annotated_features <- annotate_features(features,
                                        id_col = "comp_ids",
                                        pathway_col = "pathway",
                                        platform_col = "platform")