Export Data from a Metabolites Object

Exports all data layers from a Metabolites object to a structured directory format. For each data layer, the function creates a subdirectory containing:

the primary data matrix (data.tsv),

associated feature and sample metadata (features.tsv, samples.tsv),

feature and sample summaries (if present, feature_summary.tsv, sample_summary.tsv),

a serialized feature tree (if present),

and a config.yml file with additional metadata and processing parameters.

Usage

export_data(metabolites, directory)

Arguments

metabolites

A Metabolites object containing the data to be exported.

directory

A character string specifying the path to the directory where the data should be written.

Value

This function does not return a value. It writes the data to the specified directory as a file (e.g., CSV).

Examples

file <- system.file("extdata", "metabolon_v1_example.xlsx", package = "metaboprep2")
metabolites <- Metabolites(project_name = "MYPROJECT",
                           format       = "metabolon_v1",
                           filepath     = file)
metabolites <- import_data(metabolites)
#> Using cached MetaboAnalyst DB.
#> Attempting join [features$hmdb] ~ [annotation$hmdb_id] 
#> Attempting join [features$kegg] ~ [annotation$kegg_id] 
#> Attempting join [features$metabolite_id] ~ [annotation$name] 
#> Attempting join [features$comp_id] ~ [annotation$comp_id] 
#> Warning: 100 unannotated metabolites remain after merging with metabolomics database
if (FALSE) { # \dontrun{
# Specify the directory to export the data
export_data(metabolites, directory = "path/to/directory")
} # }