Summary Statistics for Features

This function allows you to 'describe' metabolite features using the describe() function from the psych package, as well as estimate variance, a dispersion index, the coeficent of variation, and shapiro's W-statistic.

Usage

feature_describe(data)

Arguments

data

matrix, the metabolite data matrix. Samples in row, metabolites in columns

Value

a data frame of summary statistics for features (columns) of a matrix

Examples

ex_data = sapply(1:20, function(x){ rnorm(250, 40, 5) })
feature_describe(ex_data)
#>       n     mean       sd   median      min      max    range        skew
#> X1  250 39.64876 4.938880 39.70918 25.17094 52.21342 27.04248 -0.14990620
#> X2  250 40.39411 4.821959 40.18995 21.85827 52.92200 31.06373 -0.19713772
#> X3  250 40.36539 5.063169 40.75910 25.28200 52.82567 27.54367 -0.21904779
#> X4  250 39.75367 5.221677 40.17799 27.28989 55.48867 28.19878  0.01283709
#> X5  250 40.62097 5.224598 40.59280 26.28915 57.07218 30.78302  0.01276969
#> X6  250 39.52311 5.211264 39.61068 25.59214 54.30704 28.71489  0.16919978
#> X7  250 40.07369 4.965153 39.92106 29.08322 55.78830 26.70508  0.06558329
#> X8  250 39.92190 5.486999 39.78043 26.73585 54.68577 27.94992  0.11059507
#> X9  250 40.35127 4.830219 40.32990 27.04147 53.00468 25.96321  0.00628513
#> X10 250 40.31935 5.004767 40.13017 25.35711 53.12274 27.76563 -0.09750201
#> X11 250 39.71798 4.985978 39.70652 25.31409 51.26692 25.95284  0.06927148
#> X12 250 40.07022 4.803268 40.28551 26.23457 51.67681 25.44224 -0.11286502
#> X13 250 39.99684 4.851344 39.85722 24.96378 56.14068 31.17690  0.22974772
#> X14 250 39.67074 4.872844 39.96478 25.28539 54.46565 29.18026  0.08328983
#> X15 250 39.78146 5.253056 39.35775 25.97090 52.60410 26.63320  0.16528402
#> X16 250 39.65063 5.071943 39.47765 26.86954 53.95281 27.08328  0.17998355
#> X17 250 39.76762 5.113374 39.76227 23.36306 55.54407 32.18101  0.01078787
#> X18 250 40.06673 5.170566 39.94118 27.19727 55.75799 28.56073  0.32061703
#> X19 250 39.44491 4.916655 39.54003 22.17858 57.42081 35.24223  0.15561333
#> X20 250 40.08912 4.858911 39.95204 24.80277 53.07027 28.26750  0.07396956
#>         kurtosis        se missing      var disp_index coef_variance         W
#> X1   0.014242570 0.3123622       0 24.39253  0.6152155     0.1245658 0.9943386
#> X2   0.273842723 0.3049675       0 23.25129  0.5756108     0.1193728 0.9934845
#> X3  -0.002986971 0.3202229       0 25.63568  0.6350906     0.1254334 0.9934655
#> X4  -0.209069432 0.3302478       0 27.26591  0.6858715     0.1313508 0.9926534
#> X5  -0.344111069 0.3304326       0 27.29643  0.6719787     0.1286183 0.9959922
#> X6  -0.415101180 0.3295893       0 27.15727  0.6871238     0.1318536 0.9908994
#> X7  -0.045641696 0.3140239       0 24.65275  0.6151853     0.1239006 0.9909699
#> X8  -0.309623017 0.3470283       0 30.10715  0.7541514     0.1374433 0.9938472
#> X9  -0.445998790 0.3054899       0 23.33102  0.5781979     0.1197043 0.9958790
#> X10 -0.178343363 0.3165293       0 25.04769  0.6212325     0.1241282 0.9972881
#> X11 -0.225776380 0.3153409       0 24.85997  0.6259123     0.1255345 0.9944407
#> X12  0.106000099 0.3037853       0 23.07138  0.5757737     0.1198712 0.9922703
#> X13  0.453396287 0.3068260       0 23.53554  0.5884350     0.1212932 0.9938213
#> X14  0.026561453 0.3081857       0 23.74461  0.5985422     0.1228322 0.9956664
#> X15 -0.196355673 0.3322324       0 27.59460  0.6936546     0.1320478 0.9904347
#> X16 -0.221787438 0.3207778       0 25.72460  0.6487817     0.1279158 0.9939448
#> X17  0.113561030 0.3233982       0 26.14660  0.6574846     0.1285813 0.9964317
#> X18  0.354683111 0.3270153       0 26.73476  0.6672557     0.1290489 0.9892488
#> X19  0.858268637 0.3109566       0 24.17350  0.6128420     0.1246461 0.9900463
#> X20 -0.078379963 0.3073045       0 23.60902  0.5889134     0.1212027 0.9942207
#>       log10_W
#> X1  0.9782113
#> X2  0.9719287
#> X3  0.9753268
#> X4  0.9856352
#> X5  0.9902629
#> X6  0.9915144
#> X7  0.9856918
#> X8  0.9911076
#> X9  0.9912503
#> X10 0.9861779
#> X11 0.9909671
#> X12 0.9793475
#> X13 0.9937727
#> X14 0.9912225
#> X15 0.9904886
#> X16 0.9945025
#> X17 0.9859046
#> X18 0.9944190
#> X19 0.9820222
#> X20 0.9895078