Internal Data Extraction Helper

A helper function for internal use that facilitates data extraction from a Metabolites object. This function is utilized by several exposed data getter functions, including: get_data, get_features, get_samples, get_feature_summary, get_sample_summary, get_feature_tree, get_pcs, get_var_exp, get_prob_pcs, get_acceleration_factor, and get_n_parallel.

Usage

get_(
  metabolites,
  slot,
  layer,
  feature_ids = NULL,
  sample_ids = NULL,
  drop = FALSE,
  as_df = FALSE
)

Arguments

metabolites

An object of class Metabolites.

slot

the name of the metabolites object slot to extract

layer

Character. The layer of data to extract.

feature_ids

Optional. A vector of row names to extract. Defaults to NULL, meaning all rows are extracted.

sample_ids

Optional. A vector of column names to extract. Defaults to NULL, meaning all columns are extracted.

drop

Logical. Whether to drop all NA rows or columns (excluded samples or features)

as_df

Logical. Whether to return the result as a data.frame. Defaults to FALSE (returns a matrix).

Value

A matrix or a data.frame containing the extracted data.