Retrieve Feature Summary Data

Extracts feature-level summary statistics from a Metabolites object for a specified data layer, with optional exclusions and subsetting.

Usage

get_feature_summary(
  metabolites,
  layer,
  feature_ids = NULL,
  drop = FALSE,
  as_df = FALSE
)

Arguments

metabolites

A Metabolites object containing feature summary data.

layer

A character string specifying the data layer from which to extract feature summary data.

feature_ids

Optional. A vector of row names (features) to extract. Defaults to NULL, meaning all features are included.

drop

Logical. Whether to drop all NA (excluded features)

as_df

Logical. Whether to return the result as a data.frame. Defaults to FALSE, meaning the result is returned as a matrix.

Value

A matrix or data.frame (if as_df = TRUE) containing the feature summary data for the specified layer.

Examples

file <- system.file("extdata", "metabolon_v1_example.xlsx", package = "metaboprep2")
metabolites <- Metabolites(project_name = "MYPROJECT",
                           format       = "metabolon_v1",
                           filepath     = file)
metabolites <- import_data(metabolites)
#> Using cached MetaboAnalyst DB.
#> Attempting join [features$hmdb] ~ [annotation$hmdb_id] 
#> Attempting join [features$kegg] ~ [annotation$kegg_id] 
#> Attempting join [features$metabolite_id] ~ [annotation$name] 
#> Attempting join [features$comp_id] ~ [annotation$comp_id] 
#> Warning: 100 unannotated metabolites remain after merging with metabolomics database
metabolites <- metabolite_qc(metabolites, source="raw")
#> 
#> ── Starting Metabolite QC Process ──────────────────────────────────────────────
#> ℹ Validating input parameters...
#> ✔ Input validation complete.
#> ℹ Batch normalising raw data...
#> ✔ Normalised raw data.
#> ℹ Applying exclusion of xenobiotic features...
#> ✔ Xenobiotic features exclusions assessment complete. (3 features excluded)
#> ℹ Analysing exclusions...
#> ℹ Applying exclusion of derived features...
#> ✔ Derived features exclusions assessment complete. (0 features excluded)
#> ℹ Sample & Feature Summary Statistics for raw data...
#> ✔ Sample & Feature Summary Statistics for raw data...
#> ℹ Copying batch_normalised data to new 'qcing' data layer...
#> ✔ qcing data ready for QC.
#> ℹ Assessing for extreme sample missingness >=80%...
#> ✔ Extreme sample missingness exclusions assessment complete. (0 samples excluded)
#> ℹ Assessing for total feature missingness (100%)
#> ✔ Total feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Assessing for extreme feature missingness >=80%...
#> ✔ Extreme feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Assessing for sample missingness at specified level of >=20%...
#> ✔ Sample missingness exclusions assessment complete. (2 samples excluded)
#> ℹ Assessing for feature missingness at specified level of >=20%...
#> ✔ Feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Calculating total peak abundance outliers at +/- 5 Sdev...
#> ✔ Total peak abundance outlier exclusions assessment complete. (0 samples excluded)
#> ℹ Running data outlier analysis at +/- 5 Sdev...
#> ℹ Data outlier analysis complete.
#> ℹ Re-identify feature independence and PC outliers...
#> ℹ Re-running sample summary using current exclusions...
#> ℹ QC sample summary complete.
#> ℹ PCA outlier analysis 2 complete.
#> ℹ Creating final QC dataset...
#> ℹ Rerunning Sample & Feature Summary Statistics for QC data...
#> ✔ Sample & Feature Summary Statistics for QC data complete.
#> 
#> ── Metabolite QC Process Completed ─────────────────────────────────────────────
get_feature_summary(metabolites, layer = "post_qc")
#>                             comp_id_123 comp_id_124 comp_id_125 comp_id_126
#> missingness                  0.18556701  0.11340206  0.15463918  0.00000000
#> missingness_w_exclusions             NA          NA          NA          NA
#> outlier_count                0.00000000  0.00000000  0.00000000  0.00000000
#> n                           79.00000000 86.00000000 82.00000000 97.00000000
#> mean                         0.98320638  0.99022435  1.00077528  1.00621836
#> sd                           0.20618622  0.19963856  0.19402993  0.20161603
#> median                       0.99724848  1.00000000  1.00000000  1.00129940
#> min                          0.36947037  0.51237746  0.47409525  0.61880519
#> max                          1.35588440  1.49570502  1.43560122  1.53521264
#> range                        0.98641403  0.98332756  0.96150597  0.91640745
#> skew                        -0.52562888 -0.02645627 -0.04202353  0.22930318
#> kurtosis                     0.02542520 -0.13151797 -0.30372245 -0.26206083
#> se                           0.02319776  0.02152758  0.02142702  0.02047101
#> missing                     18.00000000 11.00000000 15.00000000  0.00000000
#> var                          0.04251276  0.03985555  0.03764761  0.04064902
#> disp_index                   0.04323889  0.04024901  0.03761845  0.04039782
#> coef_variance                0.20970797  0.20160942  0.19387962  0.20037006
#> W                            0.97558137  0.99014728  0.99191220  0.98652201
#> log10_W                      0.90774346  0.96432412  0.96971772  0.98355882
#> k                            1.00000000  2.00000000  3.00000000  4.00000000
#> independent_features_binary  1.00000000  1.00000000  1.00000000  1.00000000
#>                             comp_id_127 comp_id_128 comp_id_129 comp_id_130
#> missingness                  0.00000000  0.00000000  0.00000000  0.00000000
#> missingness_w_exclusions             NA          NA          NA          NA
#> outlier_count                0.00000000  0.00000000  0.00000000  0.00000000
#> n                           97.00000000 97.00000000 97.00000000 97.00000000
#> mean                         0.99853093  0.99881770  1.01649047  1.01004910
#> sd                           0.19500766  0.19263588  0.16215103  0.19234856
#> median                       1.00000000  0.99586373  1.00032296  1.00114610
#> min                          0.43833075  0.45018546  0.63565163  0.60333160
#> max                          1.46808329  1.39524782  1.55224712  1.56908034
#> range                        1.02975254  0.94506236  0.91659549  0.96574874
#> skew                        -0.25290379 -0.23463362  0.54003490  0.22654782
#> kurtosis                     0.20824021 -0.12717019  0.85509486 -0.38994300
#> se                           0.01980003  0.01955921  0.01646394  0.01953004
#> missing                      0.00000000  0.00000000  0.00000000  0.00000000
#> var                          0.03802799  0.03710858  0.02629296  0.03699797
#> disp_index                   0.03808394  0.03715251  0.02586641  0.03662987
#> coef_variance                0.19529456  0.19286390  0.15952046  0.19043486
#> W                            0.98834629  0.98856427  0.97232869  0.98972875
#> log10_W                      0.93722164  0.95015645  0.98543582  0.98990318
#> k                            5.00000000  3.00000000  6.00000000  7.00000000
#> independent_features_binary  1.00000000  0.00000000  1.00000000  1.00000000
#>                             comp_id_131 comp_id_132 comp_id_133 comp_id_134
#> missingness                  0.00000000  0.00000000  0.00000000  0.00000000
#> missingness_w_exclusions             NA          NA          NA          NA
#> outlier_count                0.00000000  0.00000000  0.00000000  0.00000000
#> n                           97.00000000 97.00000000 97.00000000 97.00000000
#> mean                         1.01385312  0.99781296  0.98346833  0.99961304
#> sd                           0.19237500  0.20346581  0.17789309  0.20870252
#> median                       1.00753929  1.00137671  1.00000000  1.00000000
#> min                          0.62570337  0.35899797  0.50348249  0.48152620
#> max                          1.50849441  1.49907819  1.45966706  1.50759649
#> range                        0.88279104  1.14008022  0.95618458  1.02607029
#> skew                        -0.03393611 -0.08226735 -0.09715115  0.14542207
#> kurtosis                    -0.75791099  0.36774326 -0.02335749 -0.42786124
#> se                           0.01953272  0.02065882  0.01806231  0.02119053
#> missing                      0.00000000  0.00000000  0.00000000  0.00000000
#> var                          0.03700814  0.04139833  0.03164595  0.04355674
#> disp_index                   0.03650247  0.04148907  0.03217790  0.04357360
#> coef_variance                0.18974643  0.20391177  0.18088339  0.20878331
#> W                            0.98169479  0.99188412  0.99546779  0.99225167
#> log10_W                      0.96957502  0.94125321  0.96945058  0.98493746
#> k                            8.00000000  9.00000000 10.00000000 11.00000000
#> independent_features_binary  1.00000000  1.00000000  1.00000000  1.00000000
#>                             comp_id_135 comp_id_136 comp_id_137 comp_id_138
#> missingness                  0.00000000  0.00000000  0.00000000  0.00000000
#> missingness_w_exclusions             NA          NA          NA          NA
#> outlier_count                0.00000000  0.00000000  0.00000000  0.00000000
#> n                           97.00000000 97.00000000 97.00000000 97.00000000
#> mean                         1.00259686  1.00857637  1.04597478  1.00792655
#> sd                           0.17914232  0.17814468  0.19831688  0.18078250
#> median                       0.99972399  1.00414087  1.00164120  1.00000000
#> min                          0.52311637  0.41114581  0.63223371  0.65674191
#> max                          1.42479328  1.38339950  1.62790432  1.38542531
#> range                        0.90167691  0.97225369  0.99567062  0.72868340
#> skew                        -0.05349572 -0.32504125  0.44381469  0.13377645
#> kurtosis                    -0.20303173  0.34099224 -0.04349913 -0.55350610
#> se                           0.01818915  0.01808785  0.02013603  0.01835568
#> missing                      0.00000000  0.00000000  0.00000000  0.00000000
#> var                          0.03209197  0.03173553  0.03932959  0.03268231
#> disp_index                   0.03200885  0.03146566  0.03760089  0.03242529
#> coef_variance                0.17867832  0.17662983  0.18960006  0.17936079
#> W                            0.99356826  0.98677257  0.97506679  0.98291624
#> log10_W                      0.97103800  0.93260528  0.98528424  0.98033682
#> k                           12.00000000 13.00000000 14.00000000 15.00000000
#> independent_features_binary  1.00000000  1.00000000  1.00000000  1.00000000
#>                             comp_id_139 comp_id_140  comp_id_141  comp_id_142
#> missingness                  0.00000000  0.00000000  0.000000000  0.000000000
#> missingness_w_exclusions             NA          NA           NA           NA
#> outlier_count                0.00000000  0.00000000  0.000000000  0.000000000
#> n                           97.00000000 97.00000000 97.000000000 97.000000000
#> mean                         0.98330749  0.99955657  0.995748889  0.996159052
#> sd                           0.19357383  0.18752380  0.169342034  0.189550330
#> median                       1.00164161  0.99726902  1.002738572  0.997763762
#> min                          0.43723029  0.46371539  0.629247480  0.549969288
#> max                          1.44579580  1.42424604  1.436828556  1.451617518
#> range                        1.00856551  0.96053065  0.807581076  0.901648230
#> skew                        -0.25965740 -0.09480308 -0.005370907  0.009562064
#> kurtosis                     0.42568340  0.17574885 -0.491518860 -0.400709151
#> se                           0.01965444  0.01904016  0.017194078  0.019245920
#> missing                      0.00000000  0.00000000  0.000000000  0.000000000
#> var                          0.03747083  0.03516518  0.028676724  0.035929328
#> disp_index                   0.03810693  0.03518078  0.028799153  0.036067862
#> coef_variance                0.19685992  0.18760699  0.170064999  0.190281190
#> W                            0.98362476  0.99020429  0.991616847  0.994906308
#> log10_W                      0.93098518  0.95733003  0.980261557  0.980382230
#> k                           16.00000000 17.00000000 18.000000000 19.000000000
#> independent_features_binary  1.00000000  1.00000000  1.000000000  1.000000000
#>                             comp_id_143 comp_id_144 comp_id_145 comp_id_146
#> missingness                  0.00000000  0.00000000  0.00000000  0.00000000
#> missingness_w_exclusions             NA          NA          NA          NA
#> outlier_count                0.00000000  0.00000000  0.00000000  0.00000000
#> n                           97.00000000 97.00000000 97.00000000 97.00000000
#> mean                         0.98396328  1.00472702  1.01296103  1.00886892
#> sd                           0.19280081  0.20359819  0.20906508  0.19066748
#> median                       1.00000000  1.00286837  1.00372893  1.00270099
#> min                          0.50374289  0.55940752  0.46127729  0.45833624
#> max                          1.46160144  1.68684057  1.65814904  1.63611533
#> range                        0.95785854  1.12743305  1.19687175  1.17777909
#> skew                        -0.27794411  0.32245531  0.12371376  0.08617355
#> kurtosis                    -0.09541363  0.08276226  0.08090125  0.66666173
#> se                           0.01957596  0.02067226  0.02122734  0.01935935
#> missing                      0.00000000  0.00000000  0.00000000  0.00000000
#> var                          0.03717215  0.04145222  0.04370821  0.03635409
#> disp_index                   0.03777799  0.04125720  0.04314895  0.03603450
#> coef_variance                0.19594309  0.20264030  0.20639005  0.18899134
#> W                            0.98710846  0.98714672  0.99479908  0.98836255
#> log10_W                      0.94703916  0.99050665  0.97791392  0.96407596
#> k                           20.00000000 21.00000000 22.00000000 23.00000000
#> independent_features_binary  1.00000000  1.00000000  1.00000000  1.00000000
#>                             comp_id_147 comp_id_148 comp_id_149 comp_id_150
#> missingness                  0.00000000  0.00000000  0.00000000  0.00000000
#> missingness_w_exclusions             NA          NA          NA          NA
#> outlier_count                0.00000000  0.00000000  0.00000000  0.00000000
#> n                           97.00000000 97.00000000 97.00000000 97.00000000
#> mean                         1.00965721  0.99588323  1.00268666  1.00643917
#> sd                           0.16634610  0.20984013  0.18359655  0.20165418
#> median                       1.00764806  1.00000000  1.00000000  1.00000000
#> min                          0.57491341  0.47042824  0.55404004  0.52260205
#> max                          1.40413592  1.45020607  1.55878240  1.45741302
#> range                        0.82922251  0.97977783  1.00474235  0.93481097
#> skew                         0.01131511 -0.09260720  0.08723114 -0.05024686
#> kurtosis                    -0.19167377 -0.37018399  0.26390285 -0.49060428
#> se                           0.01688989  0.02130604  0.01864141  0.02047488
#> missing                      0.00000000  0.00000000  0.00000000  0.00000000
#> var                          0.02767102  0.04403288  0.03370769  0.04066441
#> disp_index                   0.02740636  0.04421490  0.03361737  0.04040424
#> coef_variance                0.16475503  0.21070756  0.18310461  0.20036400
#> W                            0.99290948  0.99143190  0.98827456  0.99112413
#> log10_W                      0.97946823  0.96567505  0.97338340  0.97151594
#> k                           24.00000000 25.00000000 26.00000000 27.00000000
#> independent_features_binary  1.00000000  1.00000000  1.00000000  1.00000000
#>                              comp_id_151 comp_id_152 comp_id_153 comp_id_154
#> missingness                  0.000000000  0.00000000  0.00000000  0.00000000
#> missingness_w_exclusions              NA          NA          NA          NA
#> outlier_count                0.000000000  0.00000000  0.00000000  0.00000000
#> n                           97.000000000 97.00000000 97.00000000 97.00000000
#> mean                         0.972398347  1.01390035  1.00544071  0.98217864
#> sd                           0.194517995  0.17631480  0.20445864  0.19153066
#> median                       0.998860299  1.00000000  1.00880484  1.00220955
#> min                          0.507615469  0.57840011  0.61305893  0.39110379
#> max                          1.456744120  1.54569763  1.78697566  1.44268869
#> range                        0.949128652  0.96729752  1.17391673  1.05158490
#> skew                         0.001418062  0.23290119  0.52477912 -0.35066260
#> kurtosis                    -0.386794862 -0.12292950  0.76042281  0.23930494
#> se                           0.019750310  0.01790206  0.02075963  0.01944699
#> missing                      0.000000000  0.00000000  0.00000000  0.00000000
#> var                          0.037837251  0.03108691  0.04180333  0.03668399
#> disp_index                   0.038911266  0.03066072  0.04157712  0.03734962
#> coef_variance                0.200039414  0.17389757  0.20335225  0.19500593
#> W                            0.994701245  0.99330701  0.96992147  0.98533644
#> log10_W                      0.977396372  0.99266025  0.98524573  0.93099896
#> k                           28.000000000 29.00000000 30.00000000  1.00000000
#> independent_features_binary  1.000000000  1.00000000  1.00000000  0.00000000
#>                             comp_id_155 comp_id_156 comp_id_157 comp_id_158
#> missingness                  0.00000000  0.00000000  0.00000000  0.00000000
#> missingness_w_exclusions             NA          NA          NA          NA
#> outlier_count                0.00000000  0.00000000  0.00000000  0.00000000
#> n                           97.00000000 97.00000000 97.00000000 97.00000000
#> mean                         1.01090639  0.99368596  0.99628193  1.00457905
#> sd                           0.21563021  0.19674492  0.17381852  0.20063209
#> median                       1.00097983  0.99949829  1.00190008  1.00000000
#> min                          0.31191193  0.47337549  0.53746737  0.51984616
#> max                          1.76942146  1.55602873  1.36624335  1.50513595
#> range                        1.45750952  1.08265324  0.82877599  0.98528979
#> skew                         0.01203956  0.01079798 -0.02658161  0.10805650
#> kurtosis                     1.25039064  0.29520528 -0.34288251 -0.12527648
#> se                           0.02189393  0.01997642  0.01764860  0.02037110
#> missing                      0.00000000  0.00000000  0.00000000  0.00000000
#> var                          0.04649639  0.03870856  0.03021288  0.04025323
#> disp_index                   0.04599475  0.03895452  0.03032563  0.04006975
#> coef_variance                0.21330383  0.19799507  0.17446720  0.19971757
#> W                            0.98082532  0.99210386  0.98862599  0.99216236
#> log10_W                      0.91684836  0.96074253  0.97084540  0.97874256
#> k                           31.00000000 32.00000000 33.00000000 34.00000000
#> independent_features_binary  1.00000000  1.00000000  1.00000000  1.00000000
#>                             comp_id_159 comp_id_160 comp_id_161 comp_id_162
#> missingness                  0.00000000  0.00000000  0.00000000  0.00000000
#> missingness_w_exclusions             NA          NA          NA          NA
#> outlier_count                0.00000000  0.00000000  0.00000000  0.00000000
#> n                           97.00000000 97.00000000 97.00000000 97.00000000
#> mean                         1.02400365  0.98287315  1.00841827  0.99605205
#> sd                           0.21939953  0.18650997  0.18725873  0.17084154
#> median                       1.00052737  1.00195175  1.00029080  0.99840704
#> min                          0.55924479  0.52927437  0.49515449  0.65325998
#> max                          1.65701135  1.38740681  1.50956392  1.58425157
#> range                        1.09776656  0.85813244  1.01440943  0.93099159
#> skew                         0.52417312 -0.31401663  0.19541682  0.24290090
#> kurtosis                     0.19161290 -0.58529416  0.52882353  0.21468486
#> se                           0.02227665  0.01893722  0.01901324  0.01734633
#> missing                      0.00000000  0.00000000  0.00000000  0.00000000
#> var                          0.04813615  0.03478597  0.03506583  0.02918683
#> disp_index                   0.04700779  0.03539213  0.03477310  0.02930252
#> coef_variance                0.21425659  0.18975996  0.18569550  0.17151869
#> W                            0.97919663  0.98093222  0.98282291  0.98509393
#> log10_W                      0.99535370  0.95095210  0.96537487  0.98438899
#> k                           35.00000000 36.00000000 37.00000000 38.00000000
#> independent_features_binary  1.00000000  1.00000000  1.00000000  1.00000000
#>                             comp_id_163 comp_id_164 comp_id_165 comp_id_166
#> missingness                  0.00000000  0.00000000  0.00000000  0.00000000
#> missingness_w_exclusions             NA          NA          NA          NA
#> outlier_count                0.00000000  0.00000000  0.00000000  0.00000000
#> n                           97.00000000 97.00000000 97.00000000 97.00000000
#> mean                         0.96421878  1.00023632  0.99664977  0.99504378
#> sd                           0.19039750  0.17990620  0.18408510  0.19014016
#> median                       1.00281956  1.00000000  1.00000000  0.99737465
#> min                          0.56764479  0.58723064  0.33620344  0.46286674
#> max                          1.30542341  1.41457264  1.40658306  1.44474884
#> range                        0.73777862  0.82734199  1.07037963  0.98188211
#> skew                        -0.20454313  0.04517994 -0.30914560 -0.13860268
#> kurtosis                    -1.08468635 -0.55642648  0.61307286 -0.02378502
#> se                           0.01933194  0.01826671  0.01869101  0.01930581
#> missing                      0.00000000  0.00000000  0.00000000  0.00000000
#> var                          0.03625121  0.03236624  0.03388732  0.03615328
#> disp_index                   0.03759645  0.03235859  0.03400124  0.03633336
#> coef_variance                0.19746296  0.17986369  0.18470390  0.19108723
#> W                            0.96153002  0.99252056  0.98431885  0.99123223
#> log10_W                      0.94643786  0.98465498  0.90730416  0.95823591
#> k                            4.00000000 39.00000000 40.00000000 41.00000000
#> independent_features_binary  0.00000000  1.00000000  1.00000000  1.00000000
#>                             comp_id_167  comp_id_168 comp_id_169 comp_id_170
#> missingness                  0.00000000  0.000000000  0.00000000  0.00000000
#> missingness_w_exclusions             NA           NA          NA          NA
#> outlier_count                0.00000000  0.000000000  0.00000000  0.00000000
#> n                           97.00000000 97.000000000 97.00000000 97.00000000
#> mean                         1.00769781  0.991355514  0.97609150  1.00574323
#> sd                           0.19625827  0.204313076  0.19764069  0.19604939
#> median                       0.99907670  1.000000000  1.00243194  1.00000000
#> min                          0.59160996  0.407247155  0.47576563  0.59139222
#> max                          1.54192138  1.536904217  1.40181359  1.57282507
#> range                        0.95031141  1.129657062  0.92604796  0.98143285
#> skew                         0.33343905 -0.001038002 -0.09747860  0.26671507
#> kurtosis                    -0.10020066  0.086865556 -0.31071805 -0.39792657
#> se                           0.01992701  0.020744850  0.02006737  0.01990580
#> missing                      0.00000000  0.000000000  0.00000000  0.00000000
#> var                          0.03851731  0.041743833  0.03906184  0.03843536
#> disp_index                   0.03822307  0.042107834  0.04001863  0.03821588
#> coef_variance                0.19475905  0.206094659  0.20248172  0.19492986
#> W                            0.98168419  0.993291164  0.99051966  0.98579167
#> log10_W                      0.98532325  0.963161109  0.96617158  0.98910397
#> k                           10.00000000 42.000000000 43.00000000 44.00000000
#> independent_features_binary  0.00000000  1.000000000  1.00000000  1.00000000
#>                             comp_id_171 comp_id_172 comp_id_173 comp_id_174
#> missingness                  0.00000000  0.00000000  0.00000000  0.00000000
#> missingness_w_exclusions             NA          NA          NA          NA
#> outlier_count                0.00000000  0.00000000  0.00000000  0.00000000
#> n                           97.00000000 97.00000000 97.00000000 97.00000000
#> mean                         1.00096980  1.00871031  0.99988188  1.01634012
#> sd                           0.18645445  0.18084928  0.21328164  0.20188028
#> median                       1.00019326  1.00000000  1.00000000  1.00000000
#> min                          0.58542266  0.62265041  0.54281572  0.54699808
#> max                          1.40987124  1.46888502  1.58046696  1.49340133
#> range                        0.82444859  0.84623461  1.03765124  0.94640326
#> skew                         0.16121799  0.09025017  0.07283816  0.01333828
#> kurtosis                    -0.54499282 -0.19978049  0.13351548 -0.29281790
#> se                           0.01893158  0.01836246  0.02165547  0.02049784
#> missing                      0.00000000  0.00000000  0.00000000  0.00000000
#> var                          0.03476526  0.03270646  0.04548906  0.04075565
#> disp_index                   0.03473158  0.03242404  0.04549443  0.04010040
#> coef_variance                0.18627381  0.17928762  0.21330684  0.19863457
#> W                            0.98540256  0.98823095  0.98134760  0.99367910
#> log10_W                      0.98505569  0.97743954  0.95741912  0.97524801
#> k                           45.00000000 46.00000000 47.00000000 48.00000000
#> independent_features_binary  1.00000000  1.00000000  1.00000000  1.00000000
#>                             comp_id_175  comp_id_176 comp_id_177 comp_id_178
#> missingness                  0.00000000  0.000000000  0.00000000  0.00000000
#> missingness_w_exclusions             NA           NA          NA          NA
#> outlier_count                0.00000000  0.000000000  0.00000000  0.00000000
#> n                           97.00000000 97.000000000 97.00000000 97.00000000
#> mean                         0.99726445  0.985738236  0.97829226  1.00040085
#> sd                           0.18052831  0.192442103  0.21170178  0.22603265
#> median                       1.00000000  1.000000000  1.00000000  0.99543689
#> min                          0.44129698  0.323857571  0.49382732  0.44395272
#> max                          1.81818570  1.614790619  1.47693182  1.76637358
#> range                        1.37688872  1.290933049  0.98310449  1.32242086
#> skew                         0.53484094  0.006160131 -0.13023233  0.17766421
#> kurtosis                     3.68145841  1.621046486 -0.64198264  0.36848429
#> se                           0.01832987  0.019539535  0.02149506  0.02295014
#> missing                      0.00000000  0.000000000  0.00000000  0.00000000
#> var                          0.03259047  0.037033963  0.04481764  0.05109076
#> disp_index                   0.03267987  0.037569774  0.04581212  0.05107029
#> coef_variance                0.18102351  0.195226376  0.21639932  0.22594208
#> W                            0.94571773  0.974976190  0.98836074  0.98592806
#> log10_W                      0.93923040  0.906236736  0.96468110  0.97160006
#> k                           49.00000000 50.000000000 51.00000000 52.00000000
#> independent_features_binary  1.00000000  1.000000000  1.00000000  1.00000000
#>                             comp_id_179 comp_id_180 comp_id_181 comp_id_182
#> missingness                  0.00000000  0.00000000  0.00000000  0.00000000
#> missingness_w_exclusions             NA          NA          NA          NA
#> outlier_count                0.00000000  0.00000000  0.00000000  0.00000000
#> n                           97.00000000 97.00000000 97.00000000 97.00000000
#> mean                         0.99322479  1.00406421  0.96392049  1.00005434
#> sd                           0.20243206  0.22086045  0.19859978  0.23424145
#> median                       1.00000000  1.00000000  1.00000000  1.00000000
#> min                          0.60743779  0.55670937  0.51146312  0.30474448
#> max                          1.65104741  1.79069731  1.52836789  1.86774552
#> range                        1.04360962  1.23398794  1.01690477  1.56300104
#> skew                         0.24006794  0.38346409 -0.15376790  0.22617918
#> kurtosis                     0.18067282  0.53190394  0.25754915  1.92234350
#> se                           0.02055386  0.02242498  0.02016475  0.02378362
#> missing                      0.00000000  0.00000000  0.00000000  0.00000000
#> var                          0.04097874  0.04877934  0.03944187  0.05486906
#> disp_index                   0.04125827  0.04858189  0.04091818  0.05486608
#> coef_variance                0.20381293  0.21996646  0.20603336  0.23422873
#> W                            0.98219872  0.98320747  0.96237589  0.96625423
#> log10_W                      0.97562538  0.98461480  0.93152845  0.89592818
#> k                           53.00000000 54.00000000 55.00000000 53.00000000
#> independent_features_binary  1.00000000  1.00000000  1.00000000  0.00000000
#>                             comp_id_183 comp_id_184 comp_id_185 comp_id_186
#> missingness                  0.00000000  0.00000000  0.00000000  0.00000000
#> missingness_w_exclusions             NA          NA          NA          NA
#> outlier_count                0.00000000  0.00000000  0.00000000  0.00000000
#> n                           97.00000000 97.00000000 97.00000000 97.00000000
#> mean                         1.00190555  1.00364693  0.99582403  0.97535077
#> sd                           0.22032132  0.20794712  0.21018739  0.19902919
#> median                       1.00000000  0.99917532  1.00000000  0.99718859
#> min                          0.38431970  0.54787018  0.25262544  0.37233868
#> max                          1.69032429  1.81195020  1.59545548  1.40070513
#> range                        1.30600459  1.26408002  1.34283005  1.02836646
#> skew                         0.18055536  0.68977839 -0.44576435 -0.32442043
#> kurtosis                     1.16585539  1.37631590  1.37842391 -0.03273280
#> se                           0.02237024  0.02111383  0.02134130  0.02020835
#> missing                      0.00000000  0.00000000  0.00000000  0.00000000
#> var                          0.04854148  0.04324201  0.04417874  0.03961262
#> disp_index                   0.04844916  0.04308488  0.04436400  0.04061371
#> coef_variance                0.21990228  0.20719151  0.21106881  0.20405909
#> W                            0.97506162  0.97076577  0.97553133  0.98635028
#> log10_W                      0.94241535  0.99298584  0.84301554  0.93344884
#> k                           56.00000000 57.00000000 58.00000000 59.00000000
#> independent_features_binary  1.00000000  1.00000000  1.00000000  1.00000000
#>                             comp_id_187 comp_id_188 comp_id_189 comp_id_190
#> missingness                  0.00000000  0.00000000  0.00000000  0.00000000
#> missingness_w_exclusions             NA          NA          NA          NA
#> outlier_count                0.00000000  0.00000000  0.00000000  0.00000000
#> n                           97.00000000 97.00000000 97.00000000 97.00000000
#> mean                         1.00674965  1.00314858  1.00844269  1.00335927
#> sd                           0.19255091  0.19989402  0.19321745  0.21538606
#> median                       1.00000000  1.00000000  0.99934571  1.00000000
#> min                          0.54561222  0.53165096  0.58073389  0.45247940
#> max                          1.38467366  1.48068819  1.47069773  1.52309223
#> range                        0.83906143  0.94903723  0.88996384  1.07061283
#> skew                        -0.15277173  0.15410266  0.24693833 -0.05007535
#> kurtosis                    -0.59439321 -0.49769926 -0.40421462 -0.21455074
#> se                           0.01955058  0.02029616  0.01961826  0.02186914
#> missing                      0.00000000  0.00000000  0.00000000  0.00000000
#> var                          0.03707585  0.03995762  0.03733298  0.04639116
#> disp_index                   0.03682728  0.03983221  0.03702043  0.04623584
#> coef_variance                0.19125998  0.19926662  0.19159984  0.21466494
#> W                            0.98601338  0.99052028  0.98559144  0.99236425
#> log10_W                      0.96528785  0.98606361  0.98675587  0.96027698
#> k                           60.00000000 61.00000000 62.00000000 63.00000000
#> independent_features_binary  1.00000000  1.00000000  1.00000000  1.00000000
#>                             comp_id_191 comp_id_192 comp_id_193 comp_id_194
#> missingness                  0.00000000  0.00000000  0.00000000  0.00000000
#> missingness_w_exclusions             NA          NA          NA          NA
#> outlier_count                0.00000000  0.00000000  0.00000000  0.00000000
#> n                           97.00000000 97.00000000 97.00000000 97.00000000
#> mean                         0.99972000  0.98366477  0.98437972  1.02772856
#> sd                           0.21212531  0.17819171  0.20395019  0.22319734
#> median                       0.99474722  1.00000000  0.99834990  1.00000000
#> min                          0.41243738  0.30210614  0.41800787  0.58609039
#> max                          1.86295319  1.55087285  1.52999345  1.52378643
#> range                        1.45051580  1.24876670  1.11198558  0.93769605
#> skew                         0.41404916 -0.50279182  0.04795437  0.44158170
#> kurtosis                     2.66126247  2.07341954  0.17329166 -0.46252318
#> se                           0.02153806  0.01809263  0.02070800  0.02266226
#> missing                      0.00000000  0.00000000  0.00000000  0.00000000
#> var                          0.04499715  0.03175229  0.04159568  0.04981705
#> disp_index                   0.04500975  0.03227958  0.04225572  0.04847297
#> coef_variance                0.21218472  0.18115085  0.20718650  0.21717538
#> W                            0.95304630  0.96317564  0.99049501  0.96781168
#> log10_W                      0.92876101  0.85439851  0.96458872  0.98541977
#> k                           64.00000000 65.00000000 51.00000000 66.00000000
#> independent_features_binary  1.00000000  1.00000000  0.00000000  1.00000000
#>                             comp_id_195 comp_id_196 comp_id_197 comp_id_198
#> missingness                  0.00000000  0.00000000  0.00000000  0.00000000
#> missingness_w_exclusions             NA          NA          NA          NA
#> outlier_count                0.00000000  0.00000000  0.00000000  0.00000000
#> n                           97.00000000 97.00000000 97.00000000 97.00000000
#> mean                         0.98693187  1.00640975  0.98887044  0.99421063
#> sd                           0.20397596  0.19572406  0.21198295  0.20578921
#> median                       1.00000000  1.00063063  0.99838393  0.99647598
#> min                          0.52161645  0.45054693  0.48734348  0.51039610
#> max                          1.60507678  1.76670924  1.58661712  1.47119316
#> range                        1.08346033  1.31616231  1.09927364  0.96079705
#> skew                         0.23017348  0.31476462 -0.09320208 -0.15005451
#> kurtosis                    -0.08896369  1.47420493 -0.01904159 -0.36019494
#> se                           0.02071062  0.01987277  0.02152361  0.02089473
#> missing                      0.00000000  0.00000000  0.00000000  0.00000000
#> var                          0.04160619  0.03830791  0.04493677  0.04234920
#> disp_index                   0.04215711  0.03806393  0.04544253  0.04259580
#> coef_variance                0.20667684  0.19447751  0.21436878  0.20698754
#> W                            0.99313479  0.97802898  0.99285935  0.98907344
#> log10_W                      0.98959330  0.96646076  0.95869234  0.95875300
#> k                           67.00000000 68.00000000 69.00000000 70.00000000
#> independent_features_binary  1.00000000  1.00000000  1.00000000  1.00000000
#>                             comp_id_199 comp_id_200 comp_id_201 comp_id_202
#> missingness                  0.00000000  0.00000000  0.00000000  0.00000000
#> missingness_w_exclusions             NA          NA          NA          NA
#> outlier_count                0.00000000  0.00000000  0.00000000  0.00000000
#> n                           97.00000000 97.00000000 97.00000000 97.00000000
#> mean                         1.00384826  1.00196416  1.01175452  1.00366523
#> sd                           0.18626338  0.21582671  0.23241981  0.19122114
#> median                       1.00000000  1.00000000  1.00000000  0.99896573
#> min                          0.64508255  0.50570922  0.34202367  0.62695366
#> max                          1.71156775  1.78900521  2.04646432  1.76603805
#> range                        1.06648520  1.28329599  1.70444065  1.13908438
#> skew                         0.57210350  0.49569793  0.64546166  0.69562393
#> kurtosis                     1.06241691  0.66723367  3.00889971  1.58544945
#> se                           0.01891218  0.02191388  0.02359866  0.01941556
#> missing                      0.00000000  0.00000000  0.00000000  0.00000000
#> var                          0.03469405  0.04658117  0.05401897  0.03656552
#> disp_index                   0.03456105  0.04648985  0.05339138  0.03643199
#> coef_variance                0.18554933  0.21540362  0.22971956  0.19052283
#> W                            0.96863058  0.97630645  0.95944814  0.96873169
#> log10_W                      0.98478699  0.98788205  0.94932743  0.99045128
#> k                           71.00000000 72.00000000 73.00000000 49.00000000
#> independent_features_binary  1.00000000  1.00000000  1.00000000  0.00000000
#>                             comp_id_203 comp_id_204 comp_id_205 comp_id_206
#> missingness                  0.00000000  0.00000000  0.00000000  0.00000000
#> missingness_w_exclusions             NA          NA          NA          NA
#> outlier_count                0.00000000  0.00000000  0.00000000  0.00000000
#> n                           97.00000000 97.00000000 97.00000000 97.00000000
#> mean                         1.00542828  1.01205840  0.99261519  1.00022973
#> sd                           0.18962613  0.20693664  0.19252469  0.21008124
#> median                       1.00000000  1.00000000  1.00000000  0.99989614
#> min                          0.50831236  0.33444719  0.53417072  0.49288196
#> max                          1.43523012  1.67539149  1.51443117  1.57044896
#> range                        0.92691776  1.34094430  0.98026045  1.07756700
#> skew                         0.09705390  0.22159342 -0.06971242  0.29931654
#> kurtosis                    -0.20549846  0.93853235  0.15169511 -0.02346308
#> se                           0.01925362  0.02101123  0.01954792  0.02133052
#> missing                      0.00000000  0.00000000  0.00000000  0.00000000
#> var                          0.03595807  0.04282277  0.03706576  0.04413413
#> disp_index                   0.03576393  0.04231255  0.03734152  0.04412399
#> coef_variance                0.18860234  0.20447105  0.19395703  0.21003299
#> W                            0.98771322  0.97972849  0.98480427  0.98731326
#> log10_W                      0.97512981  0.93230053  0.95358176  0.98638549
#> k                           74.00000000 50.00000000 75.00000000 76.00000000
#> independent_features_binary  1.00000000  0.00000000  1.00000000  1.00000000
#>                             comp_id_207 comp_id_208 comp_id_209 comp_id_210
#> missingness                  0.00000000  0.00000000  0.00000000  0.00000000
#> missingness_w_exclusions             NA          NA          NA          NA
#> outlier_count                0.00000000  0.00000000  0.00000000  1.00000000
#> n                           97.00000000 97.00000000 97.00000000 97.00000000
#> mean                         1.01129110  1.01669120  0.98854141  1.03090874
#> sd                           0.24254763  0.21419974  0.21255513  0.25770661
#> median                       1.00000000  1.00000000  0.99136977  1.00000000
#> min                          0.39403946  0.47734416  0.47065244  0.58572686
#> max                          1.79866892  1.52800050  1.49718123  2.57178844
#> range                        1.40462946  1.05065634  1.02652878  1.98606158
#> skew                         0.22306757 -0.03884212 -0.04161066  2.15750070
#> kurtosis                     0.24317194  0.08235053 -0.20311347 11.12421369
#> se                           0.02462698  0.02174869  0.02158170  0.02616614
#> missing                      0.00000000  0.00000000  0.00000000  0.00000000
#> var                          0.05882935  0.04588153  0.04517968  0.06641270
#> disp_index                   0.05817252  0.04512828  0.04570338  0.06442151
#> coef_variance                0.23983958  0.21068318  0.21501894  0.24998004
#> W                            0.99009433  0.98740356  0.99318090  0.85305682
#> log10_W                      0.97260449  0.95399368  0.96571826  0.96629722
#> k                           77.00000000 78.00000000 79.00000000 53.00000000
#> independent_features_binary  1.00000000  1.00000000  1.00000000  0.00000000
#>                             comp_id_211 comp_id_212  comp_id_213 comp_id_214
#> missingness                  0.00000000  0.00000000  0.000000000  0.00000000
#> missingness_w_exclusions             NA          NA           NA          NA
#> outlier_count                0.00000000  0.00000000  0.000000000  0.00000000
#> n                           97.00000000 97.00000000 97.000000000 97.00000000
#> mean                         1.02151149  0.97602232  0.992588513  0.98712206
#> sd                           0.21737905  0.19433121  0.208019425  0.22081850
#> median                       0.99728612  1.00000000  1.000000000  0.99996723
#> min                          0.54834398  0.30275098  0.506396810  0.48002715
#> max                          1.91788604  1.50135356  1.462443504  1.45780860
#> range                        1.36954206  1.19860258  0.956046694  0.97778146
#> skew                         0.84681832 -0.64150491 -0.009723337 -0.09010535
#> kurtosis                     2.10392516  1.15649688 -0.458292811 -0.56089139
#> se                           0.02207150  0.01973134  0.021121173  0.02242072
#> missing                      0.00000000  0.00000000  0.000000000  0.00000000
#> var                          0.04725365  0.03776462  0.043272081  0.04876081
#> disp_index                   0.04625856  0.03869237  0.043595187  0.04939694
#> coef_variance                0.21280138  0.19910529  0.209572671  0.22369929
#> W                            0.96000945  0.96897598  0.991648376  0.98965589
#> log10_W                      0.99026762  0.86480996  0.973564588  0.96495490
#> k                           60.00000000 80.00000000 81.000000000 69.00000000
#> independent_features_binary  0.00000000  1.00000000  1.000000000  0.00000000
#>                             comp_id_215 comp_id_216 comp_id_217 comp_id_218
#> missingness                  0.00000000  0.00000000  0.00000000  0.00000000
#> missingness_w_exclusions             NA          NA          NA          NA
#> outlier_count                0.00000000  0.00000000  0.00000000  0.00000000
#> n                           97.00000000 97.00000000 97.00000000 97.00000000
#> mean                         1.02160948  0.99844163  0.99921759  1.00671763
#> sd                           0.21212564  0.21650312  0.20488578  0.20485923
#> median                       1.00000000  1.00000000  1.00000000  1.00000000
#> min                          0.50997723  0.50681111  0.48758510  0.55470787
#> max                          1.70225187  2.08011964  1.71713029  1.48517557
#> range                        1.19227464  1.57330854  1.22954520  0.93046770
#> skew                         0.48419412  1.26436328  0.40228396  0.09583011
#> kurtosis                     0.72625905  5.10944056  1.24971336 -0.72436065
#> se                           0.02153810  0.02198256  0.02080300  0.02080030
#> missing                      0.00000000  0.00000000  0.00000000  0.00000000
#> var                          0.04499729  0.04687360  0.04197818  0.04196730
#> disp_index                   0.04404549  0.04694676  0.04201105  0.04168726
#> coef_variance                0.20763868  0.21684104  0.20504621  0.20349224
#> W                            0.97691182  0.92560391  0.96751225  0.98510853
#> log10_W                      0.98038962  0.97889631  0.96388948  0.97968532
#> k                           82.00000000 81.00000000 83.00000000 84.00000000
#> independent_features_binary  1.00000000  0.00000000  1.00000000  1.00000000
#>                             comp_id_219 comp_id_220 comp_id_221 comp_id_222
#> missingness                  0.00000000  0.00000000  0.00000000  0.00000000
#> missingness_w_exclusions             NA          NA          NA          NA
#> outlier_count                0.00000000  0.00000000  0.00000000  0.00000000
#> n                           97.00000000 97.00000000 97.00000000 97.00000000
#> mean                         1.03167435  1.00207369  1.01686370  0.98253249
#> sd                           0.20958361  0.22367786  0.18699744  0.21688631
#> median                       1.00000000  0.99810202  1.00000000  1.00000000
#> min                          0.62822786  0.46675204  0.56789247  0.44808521
#> max                          1.74200694  1.58847254  1.45494560  1.58098431
#> range                        1.11377908  1.12172050  0.88705313  1.13289910
#> skew                         0.65284823 -0.08786347  0.14706545 -0.19567967
#> kurtosis                     0.55948360 -0.08688682 -0.35538339 -0.16858123
#> se                           0.02127999  0.02271105  0.01898671  0.02202147
#> missing                      0.00000000  0.00000000  0.00000000  0.00000000
#> var                          0.04392529  0.05003179  0.03496804  0.04703967
#> disp_index                   0.04257670  0.04992825  0.03438813  0.04787594
#> coef_variance                0.20314900  0.22321498  0.18389627  0.22074212
#> W                            0.97349102  0.99195761  0.98667194  0.98928230
#> log10_W                      0.99663856  0.95606595  0.98166750  0.94981091
#> k                           85.00000000 86.00000000 87.00000000 77.00000000
#> independent_features_binary  1.00000000  1.00000000  1.00000000  0.00000000