Retrieve Feature Data

Extracts feature-level data from a Metabolites object for a specified data layer, with optional exclusions and subsetting.

Usage

get_features(metabolites, layer, feature_ids = NULL, as_df = FALSE)

Arguments

metabolites

A Metabolites object containing feature data.

layer

A character string specifying the data layer from which to extract features.

feature_ids

Optional. A vector of row names to extract. Defaults to NULL, meaning all rows are included.

as_df

Logical. Whether to return the result as a data.frame. Defaults to FALSE, returning a matrix.

Value

A data.table or data.frame (if as_df = TRUE) containing feature data for the specified layer.

Examples

file <- system.file("extdata", "metabolon_v1_example.xlsx", package = "metaboprep2")
metabolites <- Metabolites(project_name = "MYPROJECT",
                           format       = "metabolon_v1",
                           filepath     = file)
metabolites <- import_data(metabolites)
metabolites <- metabolite_qc(metabolites, source="raw")
#> 
#> ── Starting Metabolite QC Process ──────────────────────────────────────────────
#> ℹ Validating input parameters...
#> ✔ Input validation complete.
#> ℹ Normalising raw data...
#> ✔ Normalised raw data.
#> ℹ Sample & Feature Summary Statistics for raw data...
#> ✔ Sample & Feature Summary Statistics for raw data...
#> ℹ Copying batch_normalised data to new 'qcing' data layer...
#> ✔ qcing data ready for QC.
#> ℹ Analysing exclusions...
#> ℹ Applying exclusion of derived features...
#> ✔ Derived features exclusions assessment complete. (0 features excluded)
#> ℹ Applying exclusion of xenobiotic features...
#> ✔ Xenobiotic features exclusions assessment complete. (3 features excluded)
#> ℹ Assessing for extreme sample missingness >=80%...
#> ✔ Extreme sample missingness exclusions assessment complete. (0 samples excluded)
#> ℹ Assessing for extreme feature missingness >=80%...
#> ✔ Extreme feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Assessing for sample missingness at specified level of >=20%...
#> ✔ Sample missingness exclusions assessment complete. (2 samples excluded)
#> ℹ Assessing for feature missingness at specified level of >=20%...
#> ✔ Feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Calculating total peak abundance outliers at +/- 5 Sdev...
#> ✔ Total peak abundance outlier exclusions assessment complete. (0 samples excluded)
#> ℹ Running principal component outlier analysis at +/- 5 Sdev...
#> ℹ PCA outlier analysis complete.
#> ℹ Re-running sample summary using current exclusions...
#> ℹ QC sample summary complete.
#> ℹ Re-identify feature independence and PC outliers...
#> ℹ PCA outlier analysis 2 complete.
#> ℹ Creating final QC dataset...
#> ℹ Rerunning Sample & Feature Summary Statistics for QC data...
#> ✔ Sample & Feature Summary Statistics for QC data complete.
#> 
#> ── Metabolite QC Process Completed ─────────────────────────────────────────────
get_features(metabolites, layer = "post_qc")
#> Index: <derived_feature>
#>      feature_id platform                pathway derived_feature metabolite_id
#>          <char>   <char>                 <char>          <lgcl>        <char>
#>   1: compid_123      neg             xenobiotic           FALSE          met1
#>   2: compid_124      neg             xenobiotic           FALSE          met2
#>   3: compid_125      neg             xenobiotic           FALSE          met3
#>   4: compid_126      neg                  lipid           FALSE          met4
#>   5: compid_127      neg                  lipid           FALSE          met5
#>   6: compid_128      neg                  lipid           FALSE          met6
#>   7: compid_129      neg                  lipid           FALSE          met7
#>   8: compid_130      neg                  lipid           FALSE          met8
#>   9: compid_131      neg                  lipid           FALSE          met9
#>  10: compid_132      neg                  lipid           FALSE         met10
#>  11: compid_133      neg                  lipid           FALSE         met11
#>  12: compid_134      neg                  lipid           FALSE         met12
#>  13: compid_135      neg                  lipid           FALSE         met13
#>  14: compid_136      neg                  lipid           FALSE         met14
#>  15: compid_137      neg                  lipid           FALSE         met15
#>  16: compid_138      neg                  lipid           FALSE         met16
#>  17: compid_139      neg                  lipid           FALSE         met17
#>  18: compid_140      neg                  lipid           FALSE         met18
#>  19: compid_141      neg                  lipid           FALSE         met19
#>  20: compid_142      neg                  lipid           FALSE         met20
#>  21: compid_143      neg                  lipid           FALSE         met21
#>  22: compid_144      neg                  lipid           FALSE         met22
#>  23: compid_145      neg                  lipid           FALSE         met23
#>  24: compid_146      neg                  lipid           FALSE         met24
#>  25: compid_147      neg                  lipid           FALSE         met25
#>  26: compid_148      neg                  lipid           FALSE         met26
#>  27: compid_149      neg                  lipid           FALSE         met27
#>  28: compid_150      neg                  lipid           FALSE         met28
#>  29: compid_151      neg                  lipid           FALSE         met29
#>  30: compid_152      neg                  lipid           FALSE         met30
#>  31: compid_153      neg                  lipid           FALSE         met31
#>  32: compid_154      neg                  lipid           FALSE         met32
#>  33: compid_155      neg                  lipid           FALSE         met33
#>  34: compid_156      neg                  lipid           FALSE         met34
#>  35: compid_157      neg                  lipid           FALSE         met35
#>  36: compid_158      neg                  lipid           FALSE         met36
#>  37: compid_159      neg                  lipid           FALSE         met37
#>  38: compid_160      neg                  lipid           FALSE         met38
#>  39: compid_161      neg                  lipid           FALSE         met39
#>  40: compid_162      neg                  lipid           FALSE         met40
#>  41: compid_163      neg                  lipid           FALSE         met41
#>  42: compid_164      neg                  lipid           FALSE         met42
#>  43: compid_165      neg                  lipid           FALSE         met43
#>  44: compid_166      neg                  lipid           FALSE         met44
#>  45: compid_167      neg                  lipid           FALSE         met45
#>  46: compid_168      neg                  lipid           FALSE         met46
#>  47: compid_169      neg                  lipid           FALSE         met47
#>  48: compid_170      neg                  lipid           FALSE         met48
#>  49: compid_171      neg             amino acid           FALSE         met49
#>  50: compid_172      pos             amino acid           FALSE         met50
#>  51: compid_173      pos             amino acid           FALSE         met51
#>  52: compid_174      pos             amino acid           FALSE         met52
#>  53: compid_175      pos             amino acid           FALSE         met53
#>  54: compid_176      pos             amino acid           FALSE         met54
#>  55: compid_177      pos             amino acid           FALSE         met55
#>  56: compid_178      pos             amino acid           FALSE         met56
#>  57: compid_179      pos             amino acid           FALSE         met57
#>  58: compid_180      pos             amino acid           FALSE         met58
#>  59: compid_181      pos             amino acid           FALSE         met59
#>  60: compid_182      pos           carbohydrate           FALSE         met60
#>  61: compid_183      pos           carbohydrate           FALSE         met61
#>  62: compid_184      pos           carbohydrate           FALSE         met62
#>  63: compid_185      pos           carbohydrate           FALSE         met63
#>  64: compid_186      pos           carbohydrate           FALSE         met64
#>  65: compid_187      pos           carbohydrate           FALSE         met65
#>  66: compid_188      pos           carbohydrate           FALSE         met66
#>  67: compid_189      pos           carbohydrate           FALSE         met67
#>  68: compid_190      pos           carbohydrate           FALSE         met68
#>  69: compid_191      pos           carbohydrate           FALSE         met69
#>  70: compid_192      pos cofactors and vitamins           FALSE         met70
#>  71: compid_193      pos cofactors and vitamins           FALSE         met71
#>  72: compid_194      pos cofactors and vitamins           FALSE         met72
#>  73: compid_195      pos cofactors and vitamins           FALSE         met73
#>  74: compid_196      pos cofactors and vitamins           FALSE         met74
#>  75: compid_197      pos cofactors and vitamins           FALSE         met75
#>  76: compid_198      pos cofactors and vitamins           FALSE         met76
#>  77: compid_199      pos cofactors and vitamins           FALSE         met77
#>  78: compid_200      pos cofactors and vitamins           FALSE         met78
#>  79: compid_201      pos cofactors and vitamins           FALSE         met79
#>  80: compid_202      pos cofactors and vitamins           FALSE         met80
#>  81: compid_203      pos cofactors and vitamins           FALSE         met81
#>  82: compid_204      pos cofactors and vitamins           FALSE         met82
#>  83: compid_205      pos cofactors and vitamins           FALSE         met83
#>  84: compid_206      pos cofactors and vitamins           FALSE         met84
#>  85: compid_207      pos cofactors and vitamins           FALSE         met85
#>  86: compid_208      pos cofactors and vitamins           FALSE         met86
#>  87: compid_209      pos cofactors and vitamins           FALSE         met87
#>  88: compid_210      pos cofactors and vitamins           FALSE         met88
#>  89: compid_211      pos cofactors and vitamins           FALSE         met89
#>  90: compid_212      pos                 energy           FALSE         met90
#>  91: compid_213      pos                 energy           FALSE         met91
#>  92: compid_214      pos                 energy           FALSE         met92
#>  93: compid_215      pos                 energy           FALSE         met93
#>  94: compid_216      pos                 energy           FALSE         met94
#>  95: compid_217      pos                 energy           FALSE         met95
#>  96: compid_218      pos                 energy           FALSE         met96
#>  97: compid_219      pos                 energy           FALSE         met97
#>  98: compid_220      pos                 energy           FALSE         met98
#>  99: compid_221      pos                 energy           FALSE         met99
#> 100: compid_222      pos                 energy           FALSE        met100
#>      feature_id platform                pathway derived_feature metabolite_id
#>      comp_id   kegg    hmdb
#>       <char> <char>  <char>
#>   1:     123   K100   HMDB1
#>   2:     124   K100   HMDB2
#>   3:     125   K100   HMDB3
#>   4:     126   K200   HMDB4
#>   5:     127   K200   HMDB5
#>   6:     128   K200   HMDB6
#>   7:     129   K200   HMDB7
#>   8:     130   K200   HMDB8
#>   9:     131   K200   HMDB9
#>  10:     132   K200  HMDB10
#>  11:     133   K200  HMDB11
#>  12:     134   K200  HMDB12
#>  13:     135   K200  HMDB13
#>  14:     136   K200  HMDB14
#>  15:     137   K200  HMDB15
#>  16:     138   K200  HMDB16
#>  17:     139   K200  HMDB17
#>  18:     140   K200  HMDB18
#>  19:     141   K200  HMDB19
#>  20:     142   K200  HMDB20
#>  21:     143   K200  HMDB21
#>  22:     144   K200  HMDB22
#>  23:     145   K200  HMDB23
#>  24:     146   K200  HMDB24
#>  25:     147   K200  HMDB25
#>  26:     148   K200  HMDB26
#>  27:     149   K200  HMDB27
#>  28:     150   K200  HMDB28
#>  29:     151   K200  HMDB29
#>  30:     152   K200  HMDB30
#>  31:     153   K200  HMDB31
#>  32:     154   K200  HMDB32
#>  33:     155   K200  HMDB33
#>  34:     156   K200  HMDB34
#>  35:     157   K200  HMDB35
#>  36:     158   K200  HMDB36
#>  37:     159   K200  HMDB37
#>  38:     160   K200  HMDB38
#>  39:     161   K200  HMDB39
#>  40:     162   K200  HMDB40
#>  41:     163   K200  HMDB41
#>  42:     164   K200  HMDB42
#>  43:     165   K200  HMDB43
#>  44:     166   K200  HMDB44
#>  45:     167   K200  HMDB45
#>  46:     168   K200  HMDB46
#>  47:     169   K200  HMDB47
#>  48:     170   K200  HMDB48
#>  49:     171   K300  HMDB49
#>  50:     172   K300  HMDB50
#>  51:     173   K300  HMDB51
#>  52:     174   K300  HMDB52
#>  53:     175   K300  HMDB53
#>  54:     176   K300  HMDB54
#>  55:     177   K300  HMDB55
#>  56:     178   K300  HMDB56
#>  57:     179   K300  HMDB57
#>  58:     180   K300  HMDB58
#>  59:     181   K300  HMDB59
#>  60:     182   K200  HMDB60
#>  61:     183   K200  HMDB61
#>  62:     184   K200  HMDB62
#>  63:     185   K200  HMDB63
#>  64:     186   K200  HMDB64
#>  65:     187   K200  HMDB65
#>  66:     188   K200  HMDB66
#>  67:     189   K200  HMDB67
#>  68:     190   K200  HMDB68
#>  69:     191   K200  HMDB69
#>  70:     192   K400  HMDB70
#>  71:     193   K400  HMDB71
#>  72:     194   K400  HMDB72
#>  73:     195   K400  HMDB73
#>  74:     196   K400  HMDB74
#>  75:     197   K400  HMDB75
#>  76:     198   K400  HMDB76
#>  77:     199   K400  HMDB77
#>  78:     200   K400  HMDB78
#>  79:     201   K400  HMDB79
#>  80:     202   K400  HMDB80
#>  81:     203   K400  HMDB81
#>  82:     204   K400  HMDB82
#>  83:     205   K400  HMDB83
#>  84:     206   K400  HMDB84
#>  85:     207   K400  HMDB85
#>  86:     208   K400  HMDB86
#>  87:     209   K400  HMDB87
#>  88:     210   K400  HMDB88
#>  89:     211   K400  HMDB89
#>  90:     212   K400  HMDB90
#>  91:     213   K400  HMDB91
#>  92:     214   K400  HMDB92
#>  93:     215   K400  HMDB93
#>  94:     216   K400  HMDB94
#>  95:     217   K400  HMDB95
#>  96:     218   K400  HMDB96
#>  97:     219   K400  HMDB97
#>  98:     220   K400  HMDB98
#>  99:     221   K400  HMDB99
#> 100:     222   K400 HMDB100
#>      comp_id   kegg    hmdb