Retrieve Parallel Analysis No. PCs

Extracts the number of principal components estimated to be informative from the parallel analysis on the data in specified data layer from a Metabolites object.

Usage

get_n_parallel(metabolites, layer)

Arguments

metabolites

A Metabolites object containing the acceleration factor data.

layer

A character string specifying the data layer from which to extract the acceleration factor.

Value

An integer representing the result of the parallel analysis for the specified layer.

Examples

file <- system.file("extdata", "metabolon_v1_example.xlsx", package = "metaboprep2")
metabolites <- Metabolites(project_name = "MYPROJECT",
                           format       = "metabolon_v1",
                           filepath     = file)
metabolites <- import_data(metabolites)
#> Using cached MetaboAnalyst DB.
#> Attempting join [features$hmdb] ~ [annotation$hmdb_id] 
#> Attempting join [features$kegg] ~ [annotation$kegg_id] 
#> Attempting join [features$metabolite_id] ~ [annotation$name] 
#> Attempting join [features$comp_id] ~ [annotation$comp_id] 
#> Warning: 100 unannotated metabolites remain after merging with metabolomics database
metabolites <- metabolite_qc(metabolites, source="raw")
#> 
#> ── Starting Metabolite QC Process ──────────────────────────────────────────────
#> ℹ Validating input parameters...
#> ✔ Input validation complete.
#> ℹ Batch normalising raw data...
#> ✔ Normalised raw data.
#> ℹ Applying exclusion of xenobiotic features...
#> ✔ Xenobiotic features exclusions assessment complete. (3 features excluded)
#> ℹ Analysing exclusions...
#> ℹ Applying exclusion of derived features...
#> ✔ Derived features exclusions assessment complete. (0 features excluded)
#> ℹ Sample & Feature Summary Statistics for raw data...
#> ✔ Sample & Feature Summary Statistics for raw data...
#> ℹ Copying batch_normalised data to new 'qcing' data layer...
#> ✔ qcing data ready for QC.
#> ℹ Assessing for extreme sample missingness >=80%...
#> ✔ Extreme sample missingness exclusions assessment complete. (0 samples excluded)
#> ℹ Assessing for extreme feature missingness >=80%...
#> ✔ Extreme feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Assessing for sample missingness at specified level of >=20%...
#> ✔ Sample missingness exclusions assessment complete. (2 samples excluded)
#> ℹ Assessing for feature missingness at specified level of >=20%...
#> ✔ Feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Calculating total peak abundance outliers at +/- 5 Sdev...
#> ✔ Total peak abundance outlier exclusions assessment complete. (0 samples excluded)
#> ℹ Running data outlier analysis at +/- 5 Sdev...
#> ℹ Data outlier analysis complete.
#> ℹ Re-identify feature independence and PC outliers...
#> ℹ Re-running sample summary using current exclusions...
#> ℹ QC sample summary complete.
#> ℹ PCA outlier analysis 2 complete.
#> ℹ Creating final QC dataset...
#> ℹ Rerunning Sample & Feature Summary Statistics for QC data...
#> ✔ Sample & Feature Summary Statistics for QC data complete.
#> 
#> ── Metabolite QC Process Completed ─────────────────────────────────────────────
get_n_parallel(metabolites, layer = "post_qc")
#> [1] 15