Retrieve Principal Components

Extracts the principal components (PCs) from a Metabolites object for a specified data layer, with optional exclusions and subsetting.

Usage

get_pcs(metabolites, layer, sample_ids = NULL, drop = FALSE, as_df = FALSE)

Arguments

metabolites

A Metabolites object containing the principal component data.

layer

A character string specifying the data layer from which to extract the PCs.

sample_ids

Optional. A vector of column names to extract. Defaults to NULL, meaning all columns are included.

drop

Logical. Whether to drop completely NA rows (i.e. excluded)

as_df

Logical. Whether to return the result as a data.frame. Defaults to FALSE, returning a matrix.

Value

A matrix or data.frame (if as_df = TRUE) containing the principal components.

Examples

file <- system.file("extdata", "metabolon_v1_example.xlsx", package = "metaboprep2")
metabolites <- Metabolites(project_name = "MYPROJECT",
                           format       = "metabolon_v1",
                           filepath     = file)
metabolites <- import_data(metabolites)
#> Using cached MetaboAnalyst DB.
#> Attempting join [features$hmdb] ~ [annotation$hmdb_id] 
#> Attempting join [features$kegg] ~ [annotation$kegg_id] 
#> Attempting join [features$metabolite_id] ~ [annotation$name] 
#> Attempting join [features$comp_id] ~ [annotation$comp_id] 
#> Warning: 100 unannotated metabolites remain after merging with metabolomics database
metabolites <- metabolite_qc(metabolites, source="raw")
#> 
#> ── Starting Metabolite QC Process ──────────────────────────────────────────────
#> ℹ Validating input parameters...
#> ✔ Input validation complete.
#> ℹ Batch normalising raw data...
#> ✔ Normalised raw data.
#> ℹ Applying exclusion of xenobiotic features...
#> ✔ Xenobiotic features exclusions assessment complete. (3 features excluded)
#> ℹ Analysing exclusions...
#> ℹ Applying exclusion of derived features...
#> ✔ Derived features exclusions assessment complete. (0 features excluded)
#> ℹ Sample & Feature Summary Statistics for raw data...
#> ✔ Sample & Feature Summary Statistics for raw data...
#> ℹ Copying batch_normalised data to new 'qcing' data layer...
#> ✔ qcing data ready for QC.
#> ℹ Assessing for extreme sample missingness >=80%...
#> ✔ Extreme sample missingness exclusions assessment complete. (0 samples excluded)
#> ℹ Assessing for extreme feature missingness >=80%...
#> ✔ Extreme feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Assessing for total feature missingness (100%) of xenobiotics...
#> ✔ Total xenobiotic feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Assessing for sample missingness at specified level of >=20%...
#> ✔ Sample missingness exclusions assessment complete. (2 samples excluded)
#> ℹ Assessing for feature missingness at specified level of >=20%...
#> ✔ Feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Calculating total peak abundance outliers at +/- 5 Sdev...
#> ✔ Total peak abundance outlier exclusions assessment complete. (0 samples excluded)
#> ℹ Running data outlier analysis at +/- 5 Sdev...
#> ℹ Data outlier analysis complete.
#> ℹ Re-identify feature independence and PC outliers...
#> ℹ Re-running sample summary using current exclusions...
#> ℹ QC sample summary complete.
#> ℹ PCA outlier analysis 2 complete.
#> ℹ Creating final QC dataset...
#> ℹ Rerunning Sample & Feature Summary Statistics for QC data...
#> ✔ Sample & Feature Summary Statistics for QC data complete.
#> 
#> ── Metabolite QC Process Completed ─────────────────────────────────────────────
get_pcs(metabolites, layer = "post_qc")
#>                pc1         pc2          pc3          pc4          pc5
#> ind1            NA          NA           NA           NA           NA
#> ind2   13.77695409  4.17441527  1.420756729  1.310236342  3.974351268
#> ind3   -2.77229823  4.64932901  1.913890329  0.038806075  2.094638252
#> ind4    0.91057797  1.71531752  0.606661049  0.273289061 -3.156375238
#> ind5    2.07755969  1.10325047  3.272865434  2.235201867  1.315029543
#> ind6   -1.10391953  1.09920865 -1.823095754 -3.749951829  0.356030905
#> ind7    0.50196012 -1.20675199 -0.654625165 -0.453481234  1.140269743
#> ind8    1.14221020 -3.32723155  4.801397170  1.709428852 -0.005435586
#> ind9            NA          NA           NA           NA           NA
#> ind10   1.30448150 -3.24697804 -1.555439647  0.495639604 -1.602470831
#> ind11   1.76149574 -3.87341426 -2.660457693  1.060700107  0.791282663
#> ind12  -1.01292728 -1.73698812 -3.026889809  1.696228378  0.389987868
#> ind13  -3.98338285 -0.27000845  1.162297640 -3.591911753  0.167783034
#> ind14   1.21024232  1.19080380 -0.059629057  1.252748461 -0.847202189
#> ind15  -1.86511358  1.70518587 -1.401426279 -2.999202679 -0.670118786
#> ind16  -4.58742930  0.43384307 -4.595256615 -0.546781499 -1.340974656
#> ind17   1.36995390 -6.55573913 -1.746362321  2.597189736  2.714268925
#> ind18   1.83855719 -1.30593017 -1.837944300 -3.687315428 -0.883966141
#> ind19  -0.85198589  0.20537850  1.069387619 -2.951758248  0.967034089
#> ind20  -0.95914411 -0.88722714  0.885760777 -1.604887629 -1.178453638
#> ind21   3.85469520  0.47574598 -0.333187445  0.186196071 -4.327708584
#> ind22   0.80581884  0.15723068 -1.804907756 -1.817194026 -1.487715482
#> ind23   1.94480703  0.93592321 -0.034358195 -2.777914034 -0.273810944
#> ind24  -1.93676295 -1.75732154  0.848494846 -3.203755101 -1.134983763
#> ind25   3.23508121 -1.33274120 -2.376922281 -0.498615614 -1.108332789
#> ind26   0.25736296 -2.30968217 -3.247250292 -0.014922353 -0.366254137
#> ind27  -0.68741711 -0.29613822  4.228679359 -0.563857288 -1.269397001
#> ind28  -1.49057315  3.13713459 -0.207166033  0.412132750  0.703938728
#> ind29  -0.99390734 -1.60108203  0.962671865  1.059972737  2.321189187
#> ind30   4.13230228  0.17825565 -1.369783138  0.730582304 -1.274230824
#> ind31   0.78792094  2.42670928 -0.544654293  1.562630697  0.636038148
#> ind32  -2.06111418  1.43205277 -1.710356320  1.878034284  2.068158919
#> ind33  -1.05897814 -0.16868604  0.289875826  0.582810675  0.417152125
#> ind34   2.04221319 -1.81324856 -0.113228760 -0.550128328 -2.054979114
#> ind35  -0.06645826 -3.06855640 -3.691983826 -1.289097297  5.653787314
#> ind36   3.69104539 -0.62116830 -1.829186766 -1.007484783  1.069219037
#> ind37  -2.95969521  4.23566349 -0.742444284  1.292570741 -0.281258858
#> ind38  -4.26284912 -1.44721068  0.194580977 -1.674969827  3.118902073
#> ind39  -2.74654889 -0.72966557 -1.706632706  4.645907888 -4.404417571
#> ind40  -4.09308609 -1.23923772  2.600627442  3.400906139  1.730766370
#> ind41  -3.28581749 -2.18051368  0.596757977  0.223795055  5.677405641
#> ind42  -1.56361735  1.06242567  0.241384098 -1.838949981 -0.275695209
#> ind43  -0.73392265  2.01922834  4.463447347  2.674562017 -1.342614239
#> ind44   0.59990578  2.92036901 -0.684280831  0.275081388  0.185375218
#> ind45   2.30979632 -0.57148529 -1.851940621  0.584348473  1.681023719
#> ind46  -0.63499828 -1.24571719  4.702459513 -0.405472658  0.672671092
#> ind47   1.11270670  1.05845870  2.848345294 -0.103525765  1.378399775
#> ind48  -4.57232739  0.92175142 -1.609833600 -0.435588797 -2.660429179
#> ind49   0.65355158 -1.24706875  1.905982212  0.416533626 -0.752574686
#> ind50   1.68465148 -2.16529564  1.351793295  5.577721893 -3.595694971
#> ind51   2.65047405  1.25651152 -1.790917978 -0.879274321 -1.108268951
#> ind52  -1.57853972  2.34851781 -0.932496578  0.067970019  0.375622927
#> ind53   1.68134317 -3.78889566  0.282319287 -0.025611669 -0.510561748
#> ind54   2.19055518  1.12097305  0.282970570  0.820321249  0.593699357
#> ind55   1.64370724  3.01695913 -1.566493629  0.075311575  0.088304130
#> ind56  -1.06906433  2.15518492 -2.771945900 -0.389458404 -0.423390232
#> ind57  -1.04434080 -0.35954793  0.295708113 -0.259620516  3.825938809
#> ind58  -0.80521326  0.05961088  2.443928522 -2.071257899  0.250205703
#> ind59           NA          NA           NA           NA           NA
#> ind60  -1.98370208 -0.99587272 -2.361762803  0.267087563 -2.779371727
#> ind61   1.52644128 -2.42267660  0.144802145 -1.557933738  1.052293154
#> ind62  -1.32353317  2.55516228  2.830502353  1.184279281 -1.621529908
#> ind63   0.43991196 -0.44078304  0.003798733 -1.628543912  1.861496861
#> ind64  -0.95176587  0.25409050 -0.788782818  2.995090082  0.437696022
#> ind65  -1.61057957  0.67897232  0.469527749  0.303917773  0.068035653
#> ind66   2.86025245  4.26010608  1.911314925  0.762545627 -0.878060058
#> ind67  -0.04230756 -3.86551903 -0.583263489 -1.722675102 -1.752150337
#> ind68   2.72277130  1.85877247 -0.568051774 -2.788461503 -1.150111399
#> ind69  -1.78343539 -3.09132094  2.356948642 -0.853257058 -1.295384168
#> ind70   3.59303759 -0.55432201 -2.843471301  0.944930602  0.348489677
#> ind71   0.59964122 -1.52685999  2.338767603  1.871628255 -0.823299683
#> ind72   1.44377300  5.61492240 -1.099604559 -1.689001313 -0.210967442
#> ind73  -1.02619115  0.10736987 -1.715018671 -1.548638111 -3.428611840
#> ind74  -1.55987860 -2.16775276  0.598228227  0.418738862 -1.328522981
#> ind75  -0.34331374  3.27624335 -3.667635199  1.046453669  0.158181707
#> ind76  -1.48235532 -1.42713605 -2.979505366  0.179362022  2.080893182
#> ind77   2.15967533  1.77430113 -1.164187229  2.330925532  1.513563653
#> ind78  -0.50596588 -5.94348594  1.092264495 -0.519284940 -3.000787482
#> ind79   1.99554355  0.96509262 -0.766912038 -0.429530406  2.330923769
#> ind80  -1.45355416 -1.40288432  3.225200781 -3.950308575 -0.271159924
#> ind81   1.52356984  1.08460504  0.820483438  3.082002764 -0.799752169
#> ind82  -2.67409152  2.00736053  2.038284618 -0.353815455  0.966292266
#> ind83  -1.64114824  1.64035838  3.097251349 -2.104983980  0.087589329
#> ind84   0.75152665 -1.69909109 -1.964145914 -0.252131363  0.986804064
#> ind85   0.37620929  1.90216920  4.290346465 -4.416058376 -0.440182860
#> ind86   0.70029542  1.88363702 -1.495289937  0.367484257 -0.205754759
#> ind87  -2.09152899 -2.64952917 -0.247050223  3.346978363  2.899158992
#> ind88  -0.96548786 -1.55579343 -0.390882232 -0.671975941 -0.751827526
#> ind89   1.69904584 -1.97864310 -0.968186213 -1.961969805  1.221885164
#> ind90  -1.14984548 -0.37761597  2.548592788 -0.004902922  0.086236319
#> ind91   1.33502846 -0.88224558  2.284503139  0.346552307 -0.593140822
#> ind92  -5.78580939  4.40372687  0.023710885  5.219104968 -0.736690375
#> ind93  -4.34310064 -0.25969361 -1.539562195  2.974225907  1.867819883
#> ind94  -1.55833240  4.41364380 -1.936336925 -2.342611362 -0.369607860
#> ind95  -0.31605509  2.11440655 -0.775961305  0.492395973 -1.393058701
#> ind96   2.07329370  0.06574252  0.174474787  1.038628334 -1.470859594
#> ind97   1.43127863 -1.61748266  4.933548817  0.010132181  2.656833365
#> ind98   0.47458757 -3.00908127 -0.223216689 -0.188801748 -2.324275681
#> ind99  -1.90343743  2.22145366 -1.070024873 -0.127572345  1.999017801
#> ind100  2.39903769 -2.05625413  0.574354399  0.185124528 -3.019262777
#>                pc6         pc7         pc8         pc9         pc10
#> ind1            NA          NA          NA          NA           NA
#> ind2    1.41990568 -5.57575628  1.17719603  2.47471155 -0.322418913
#> ind3   -0.82562104 -0.98200828  0.01763559  4.56101270 -3.681412198
#> ind4   -1.39669814  1.91228487  0.04940041  2.75415759 -1.217368667
#> ind5   -1.48634416 -0.34715164 -0.24914696 -1.31539461 -1.489654316
#> ind6   -0.10345122 -0.25412442 -0.68409821  0.36244164 -0.545197470
#> ind7    0.95410202 -1.08446235 -1.57479291 -0.36443164 -2.258439223
#> ind8   -0.86521892  1.94970817 -2.76022782  0.57929007  0.992297278
#> ind9            NA          NA          NA          NA           NA
#> ind10   1.09797745  0.24008686  0.36023866 -1.62578830  1.106776788
#> ind11  -2.69541977 -0.69555730 -1.17860144 -0.11629818  1.800106707
#> ind12  -1.54345789  1.25686735  2.84703448  0.09030215 -1.325949904
#> ind13  -1.07911715 -1.36199857  0.16105833  2.36423609  0.183121865
#> ind14   0.80684039  0.81021927 -1.32238576 -0.73010575  1.427478022
#> ind15  -0.74301857 -0.40909292  0.93226868 -1.00405780 -2.067577740
#> ind16   2.82276077  1.51923972 -1.08491225  0.81166344 -1.515134426
#> ind17   1.29788847  0.57914167  3.03027529  1.77300841 -2.104396155
#> ind18   0.35566823 -2.60835395  1.22155119  1.26258060 -2.480642654
#> ind19   3.19387239  0.45366787  0.70874748 -1.07210837 -1.267247205
#> ind20  -0.02800927 -2.06881822 -0.28918395  0.72823309 -0.339793111
#> ind21   0.50965292  1.34285137 -2.78068832 -1.35522794  1.348854158
#> ind22  -3.06789022  0.97676281  1.40967908  0.30790739  2.041364372
#> ind23   2.75991568 -0.53371753  1.09485225  0.25200538 -0.617943024
#> ind24  -2.00763233 -1.44070538 -1.77668181  0.40913807  0.685096538
#> ind25  -1.63107953 -1.89379009 -3.68863896 -0.04496886 -0.693577433
#> ind26   1.59454018  0.90882141  2.22850841  0.74898511 -1.899259866
#> ind27  -1.34930076  3.23690083 -0.75280445 -0.97401393 -1.512736444
#> ind28  -1.38144913  0.92458097 -2.21212174  0.20119381  1.478924960
#> ind29   1.12836686  1.94544627  0.87887722 -2.33930660  0.034839743
#> ind30  -2.46855370 -0.57478370  0.57542963 -2.27296252 -3.510656738
#> ind31   0.24779696  1.01208357  2.22129225  0.31648246  2.406253364
#> ind32   0.75431843 -0.27813348  0.85873856  0.43544083 -0.165198200
#> ind33  -0.56258640  0.38815816 -3.27609138 -1.27289386 -0.738571147
#> ind34   0.58990891 -0.09669271 -2.06589570  0.04014818 -2.598454516
#> ind35   1.82116522 -3.41205350 -1.69598922  0.91533259  4.646694492
#> ind36   0.39551606  1.33543295  0.89524759  1.66951510  1.295491961
#> ind37  -0.64349176 -4.79530720  1.09965798 -1.38846453  1.238319420
#> ind38   0.68139016 -1.52187304  0.25598216  1.93160089  0.764388391
#> ind39  -0.56096929 -2.04113792 -0.50839131 -0.39173712  1.120013778
#> ind40   1.80532810  3.57881122 -1.25558631  0.53713856  1.415371186
#> ind41  -0.54524765  2.13279961 -1.83089970 -0.45229521 -3.002045792
#> ind42   0.01698993 -0.20997671  1.54476128  0.23182310  0.513443250
#> ind43  -1.33274868  0.52673406  2.58219390  2.83888661 -0.289252389
#> ind44  -2.66804029  1.79227296  0.55202788  1.85316368  1.143955631
#> ind45   3.41416389 -0.19867730 -0.19325325 -1.80857215 -0.901342932
#> ind46   1.57538661 -0.31055542  1.33687897  0.84239259 -2.456974278
#> ind47  -3.00890894 -0.13458380  0.60451425 -1.72071462 -0.212549335
#> ind48   0.79649283 -0.91964041  1.16416332  0.15644883  2.264218418
#> ind49  -1.75349972 -0.33536943  1.50214081 -3.07825975  0.043666383
#> ind50   5.89403147  2.11108245 -0.54895985 -0.15743784 -0.093795622
#> ind51   1.36241103  2.32152511 -2.94725338  1.13790100 -0.680662201
#> ind52  -1.32824505  0.28097966  0.76610583  0.33602629  0.158978641
#> ind53   1.92051339 -1.35980312  0.19560221 -0.39355637 -0.866244163
#> ind54  -0.53317930 -0.35992136 -1.93676717 -0.43448471 -0.405310181
#> ind55   2.65587226  3.13261457  0.58862116 -0.58867275  1.444366654
#> ind56   2.02348313 -0.59065680 -2.31765028  0.70002453  0.869927881
#> ind57  -1.22104003 -0.80025433  1.19170333  0.23214264  1.720200541
#> ind58  -0.10948260  2.01107505  0.94010408  1.20782098  0.899677831
#> ind59           NA          NA          NA          NA           NA
#> ind60  -0.03600544 -0.99795893 -0.30886860  1.99909221  0.828087980
#> ind61   2.43406892 -0.60456166 -1.57682748 -2.76359228  0.471744077
#> ind62  -1.96909448 -1.18081826 -3.14028900  0.72607185  0.359895815
#> ind63  -1.31834511  1.53060279 -2.37125953 -0.46142903 -1.118269988
#> ind64  -1.94131573 -1.74420643  0.97383274 -1.82765801  1.399732926
#> ind65  -2.19800949  1.53645071 -0.91935165 -0.58570040 -3.099690229
#> ind66   0.74642016 -0.08388101 -0.30455965 -0.16608275  1.664990879
#> ind67  -0.09499428 -0.87398951  0.20623492 -2.25882508 -1.873331299
#> ind68  -2.30803502 -0.40357551  0.49934514 -1.28062164 -1.167098747
#> ind69  -2.93336238  0.27287052 -0.26678192  0.87084623  1.157186094
#> ind70  -3.77986506  3.30991289 -0.33074021  2.43228960 -0.357738321
#> ind71  -0.12204629 -3.30314460  1.36813691  1.19276517 -1.387458131
#> ind72   0.28728547  2.37416541  1.73513979  0.06675096 -0.188270301
#> ind73   1.78891348  0.57064808  3.17563425 -1.37011159 -0.259440673
#> ind74   2.70725081 -0.23817970  0.64360883 -0.40173011 -1.515042553
#> ind75  -1.60851135  0.61397389  0.20282469  0.17421227 -2.515331515
#> ind76  -2.51897978  0.54572000  4.48146925 -4.18898396  0.288135198
#> ind77  -2.96713493  0.65150446 -1.91818325 -1.56491908  1.757196524
#> ind78  -1.51113592 -3.81185864 -0.88747291  0.52277466  0.288878763
#> ind79  -0.04624852  3.23890853 -1.83444861 -2.87208904  0.944485890
#> ind80  -1.70581793 -0.88164321  1.58760600 -1.89906100  1.535946792
#> ind81   0.97667536  0.30674706  3.26451875 -2.96950967 -1.561081138
#> ind82   1.02199013 -1.94097855 -2.70521427 -3.06070490  0.932101439
#> ind83   2.38834660 -2.06531660 -0.47565806 -1.92771016  0.716137884
#> ind84  -0.22662858  0.69997235  0.72708666  0.61951743  1.649304967
#> ind85   1.55203490  3.44821781  2.75214839  0.99246505  1.921215513
#> ind86   0.75016133 -0.89974216 -1.41118418  1.16145995  0.006089979
#> ind87  -1.85232045  0.50860442 -0.34864579  2.16651879  1.804600054
#> ind88  -0.68354148  0.51284148 -0.49855001 -2.50355102  1.480236267
#> ind89  -0.55289101  0.83066643  1.05623688  1.94774640  1.148393554
#> ind90   2.60096471 -1.38962918 -2.82652756  2.20306942 -1.277958060
#> ind91   2.90634456  1.04766736  1.60636316  1.51897950  3.645682446
#> ind92  -0.05068369 -4.01622404  1.93167425 -1.18208648 -0.417538797
#> ind93   0.80241107  0.98501779 -1.11694930  1.79589046 -1.990226395
#> ind94   1.29397058  0.17236167 -2.04630270 -1.59885130 -0.193579004
#> ind95   2.49353681 -1.09009331  0.13008246  1.25824256  2.297954430
#> ind96  -1.05976665 -0.39539464 -0.75003592 -0.05082268 -0.620337969
#> ind97   0.33008355 -0.53266419  1.03687435 -1.87665222  0.210784827
#> ind98  -0.83493737 -0.15988114  0.76309113  3.92552094  0.273150296
#> ind99   1.68844227  0.88557367  0.54647508  0.20079094 -0.959181209
#> ind100 -1.40578768  1.08612229  1.28800080  0.87428549  1.935619720
#>        pc1_3_sd_outlier pc2_3_sd_outlier pc1_4_sd_outlier pc2_4_sd_outlier
#> ind1                 NA               NA               NA               NA
#> ind2                  1                0                1                0
#> ind3                  0                0                0                0
#> ind4                  0                0                0                0
#> ind5                  0                0                0                0
#> ind6                  0                0                0                0
#> ind7                  0                0                0                0
#> ind8                  0                0                0                0
#> ind9                 NA               NA               NA               NA
#> ind10                 0                0                0                0
#> ind11                 0                0                0                0
#> ind12                 0                0                0                0
#> ind13                 0                0                0                0
#> ind14                 0                0                0                0
#> ind15                 0                0                0                0
#> ind16                 0                0                0                0
#> ind17                 0                0                0                0
#> ind18                 0                0                0                0
#> ind19                 0                0                0                0
#> ind20                 0                0                0                0
#> ind21                 0                0                0                0
#> ind22                 0                0                0                0
#> ind23                 0                0                0                0
#> ind24                 0                0                0                0
#> ind25                 0                0                0                0
#> ind26                 0                0                0                0
#> ind27                 0                0                0                0
#> ind28                 0                0                0                0
#> ind29                 0                0                0                0
#> ind30                 0                0                0                0
#> ind31                 0                0                0                0
#> ind32                 0                0                0                0
#> ind33                 0                0                0                0
#> ind34                 0                0                0                0
#> ind35                 0                0                0                0
#> ind36                 0                0                0                0
#> ind37                 0                0                0                0
#> ind38                 0                0                0                0
#> ind39                 0                0                0                0
#> ind40                 0                0                0                0
#> ind41                 0                0                0                0
#> ind42                 0                0                0                0
#> ind43                 0                0                0                0
#> ind44                 0                0                0                0
#> ind45                 0                0                0                0
#> ind46                 0                0                0                0
#> ind47                 0                0                0                0
#> ind48                 0                0                0                0
#> ind49                 0                0                0                0
#> ind50                 0                0                0                0
#> ind51                 0                0                0                0
#> ind52                 0                0                0                0
#> ind53                 0                0                0                0
#> ind54                 0                0                0                0
#> ind55                 0                0                0                0
#> ind56                 0                0                0                0
#> ind57                 0                0                0                0
#> ind58                 0                0                0                0
#> ind59                NA               NA               NA               NA
#> ind60                 0                0                0                0
#> ind61                 0                0                0                0
#> ind62                 0                0                0                0
#> ind63                 0                0                0                0
#> ind64                 0                0                0                0
#> ind65                 0                0                0                0
#> ind66                 0                0                0                0
#> ind67                 0                0                0                0
#> ind68                 0                0                0                0
#> ind69                 0                0                0                0
#> ind70                 0                0                0                0
#> ind71                 0                0                0                0
#> ind72                 0                0                0                0
#> ind73                 0                0                0                0
#> ind74                 0                0                0                0
#> ind75                 0                0                0                0
#> ind76                 0                0                0                0
#> ind77                 0                0                0                0
#> ind78                 0                0                0                0
#> ind79                 0                0                0                0
#> ind80                 0                0                0                0
#> ind81                 0                0                0                0
#> ind82                 0                0                0                0
#> ind83                 0                0                0                0
#> ind84                 0                0                0                0
#> ind85                 0                0                0                0
#> ind86                 0                0                0                0
#> ind87                 0                0                0                0
#> ind88                 0                0                0                0
#> ind89                 0                0                0                0
#> ind90                 0                0                0                0
#> ind91                 0                0                0                0
#> ind92                 0                0                0                0
#> ind93                 0                0                0                0
#> ind94                 0                0                0                0
#> ind95                 0                0                0                0
#> ind96                 0                0                0                0
#> ind97                 0                0                0                0
#> ind98                 0                0                0                0
#> ind99                 0                0                0                0
#> ind100                0                0                0                0
#>        pc1_5_sd_outlier pc2_5_sd_outlier
#> ind1                 NA               NA
#> ind2                  0                0
#> ind3                  0                0
#> ind4                  0                0
#> ind5                  0                0
#> ind6                  0                0
#> ind7                  0                0
#> ind8                  0                0
#> ind9                 NA               NA
#> ind10                 0                0
#> ind11                 0                0
#> ind12                 0                0
#> ind13                 0                0
#> ind14                 0                0
#> ind15                 0                0
#> ind16                 0                0
#> ind17                 0                0
#> ind18                 0                0
#> ind19                 0                0
#> ind20                 0                0
#> ind21                 0                0
#> ind22                 0                0
#> ind23                 0                0
#> ind24                 0                0
#> ind25                 0                0
#> ind26                 0                0
#> ind27                 0                0
#> ind28                 0                0
#> ind29                 0                0
#> ind30                 0                0
#> ind31                 0                0
#> ind32                 0                0
#> ind33                 0                0
#> ind34                 0                0
#> ind35                 0                0
#> ind36                 0                0
#> ind37                 0                0
#> ind38                 0                0
#> ind39                 0                0
#> ind40                 0                0
#> ind41                 0                0
#> ind42                 0                0
#> ind43                 0                0
#> ind44                 0                0
#> ind45                 0                0
#> ind46                 0                0
#> ind47                 0                0
#> ind48                 0                0
#> ind49                 0                0
#> ind50                 0                0
#> ind51                 0                0
#> ind52                 0                0
#> ind53                 0                0
#> ind54                 0                0
#> ind55                 0                0
#> ind56                 0                0
#> ind57                 0                0
#> ind58                 0                0
#> ind59                NA               NA
#> ind60                 0                0
#> ind61                 0                0
#> ind62                 0                0
#> ind63                 0                0
#> ind64                 0                0
#> ind65                 0                0
#> ind66                 0                0
#> ind67                 0                0
#> ind68                 0                0
#> ind69                 0                0
#> ind70                 0                0
#> ind71                 0                0
#> ind72                 0                0
#> ind73                 0                0
#> ind74                 0                0
#> ind75                 0                0
#> ind76                 0                0
#> ind77                 0                0
#> ind78                 0                0
#> ind79                 0                0
#> ind80                 0                0
#> ind81                 0                0
#> ind82                 0                0
#> ind83                 0                0
#> ind84                 0                0
#> ind85                 0                0
#> ind86                 0                0
#> ind87                 0                0
#> ind88                 0                0
#> ind89                 0                0
#> ind90                 0                0
#> ind91                 0                0
#> ind92                 0                0
#> ind93                 0                0
#> ind94                 0                0
#> ind95                 0                0
#> ind96                 0                0
#> ind97                 0                0
#> ind98                 0                0
#> ind99                 0                0
#> ind100                0                0