Retrieve Probabilistic Principal Components

Extracts the probabilistic principal components (PCs) from a Metabolites object for a specified data layer, with optional exclusions applied.

Usage

get_prob_pcs(
  metabolites,
  layer,
  sample_ids = NULL,
  drop = FALSE,
  as_df = FALSE
)

Arguments

metabolites

A Metabolites object containing the probabilistic PC data.

layer

A character string specifying the data layer from which to extract the probabilistic PCs.

sample_ids

Optional. A vector of column names to extract. Defaults to NULL, meaning all columns are included.

drop

Logical. Whether to drop completely NA rows (i.e. excluded)

as_df

Logical. Whether to return the result as a data.frame. Defaults to FALSE, returning a matrix.

Value

A matrix or data.frame (if as_df = TRUE) containing the probabilistic principal components.

Examples

file <- system.file("extdata", "metabolon_v1_example.xlsx", package = "metaboprep2")
metabolites <- Metabolites(project_name = "MYPROJECT",
                           format       = "metabolon_v1",
                           filepath     = file)
metabolites <- import_data(metabolites)
#> Warning: 86 unmapped features:
#>  - 123, 124, 125, 126, 127, 128, 129, 130, 131, 132, 133, 134, 135, 137, 138, 139, 140, 141, 142, 143, 145, 146, 147, 148, 149, 150, 151, 152, 153, 154, 155, 156, 157, 159, 160, 161, 162, 163, 164, 165, 166, 167, 168, 169, 170, 172, 173, 174, 175, 176, 177, 178, 183, 184, 185, 186, 187, 188, 190, 191, 193, 195, 196, 198, 199, 200, 201, 202, 204, 205, 206, 207, 208, 209, 210, 211, 213, 214, 215, 216, 217, 218, 219, 220, 221, 222
metabolites <- metabolite_qc(metabolites, source="raw")
#> 
#> ── Starting Metabolite QC Process ──────────────────────────────────────────────
#> ℹ Validating input parameters...
#> ✔ Input validation complete.
#> ℹ Applying exclusion of xenobiotic features...
#> ✔ Xenobiotic features exclusions assessment complete. (3 features excluded)
#> ℹ Analysing exclusions...
#> ℹ Applying exclusion of derived features...
#> ✔ Derived features exclusions assessment complete. (0 features excluded)
#> ℹ Sample & Feature Summary Statistics for raw data...
#> ✔ Sample & Feature Summary Statistics for raw data...
#> ℹ Copying raw data to new 'qcing' data layer...
#> ✔ qcing data ready for QC.
#> ℹ Assessing for extreme sample missingness >=80%...
#> ✔ Extreme sample missingness exclusions assessment complete. (0 samples excluded)
#> ℹ Assessing for extreme feature missingness >=80%...
#> ✔ Extreme feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Assessing for sample missingness at specified level of >=20%...
#> ✔ Sample missingness exclusions assessment complete. (2 samples excluded)
#> ℹ Assessing for feature missingness at specified level of >=20%...
#> ✔ Feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Calculating total peak abundance outliers at +/- 5 Sdev...
#> ✔ Total peak abundance outlier exclusions assessment complete. (0 samples excluded)
#> ℹ Running data outlier analysis at +/- 5 Sdev...
#> ℹ Data outlier analysis complete.
#> ℹ Re-identify feature independence and PC outliers...
#> ℹ Re-running sample summary using current exclusions...
#> ℹ QC sample summary complete.
#> ℹ PCA outlier analysis 2 complete.
#> ℹ Creating final QC dataset...
#> ℹ Rerunning Sample & Feature Summary Statistics for QC data...
#> ✔ Sample & Feature Summary Statistics for QC data complete.
#> 
#> ── Metabolite QC Process Completed ─────────────────────────────────────────────
get_prob_pcs(metabolites, layer = "post_qc")
#>           prob_pc1    prob_pc2    prob_pc3    prob_pc4     prob_pc5    prob_pc6
#> ind1   12.61437120  2.47291203  1.96247382 -3.39933944  3.339498569  0.83363466
#> ind2   11.36047057 -0.69661854 -1.62821349 -0.77629082 -3.361826463 -1.01241704
#> ind3    0.25169154  4.93685035 -0.73989774 -2.69500866 -1.401680879  1.33486587
#> ind4    1.30931413  2.19236956 -0.15054611  0.11237877  1.943981542 -1.48301701
#> ind5    2.82722244  1.80642662  1.43939447 -3.07929543 -1.165086059 -0.53775635
#> ind6   -0.12070582  2.89094534 -0.79301625  2.73493607  0.181912129 -1.16043282
#> ind7   -0.26989163  0.70803203  0.63053330  1.31452733 -1.361012698 -1.19961461
#> ind8    0.09463535  0.75107969  3.36746283 -1.83807860  0.142673575 -0.50331317
#> ind9            NA          NA          NA          NA           NA          NA
#> ind10  -0.06271049 -0.97669134  1.04087346  1.72389437  1.455284578  0.70037284
#> ind11  -0.06313026 -1.54194306  0.03001117  0.01600709 -0.712814498 -1.12478502
#> ind12  -2.35831692  0.43698725 -0.72290344  0.36391652 -1.143932350  0.23710025
#> ind13  -2.05632405  3.95757573  1.97135964  1.19223979  0.224979135 -0.88142371
#> ind14   1.51355609  1.58463234 -0.97568757 -0.04754646  0.109986109  0.46872119
#> ind15  -0.18178129  3.87844853 -0.42291503  2.88231512 -0.339052780 -0.53938735
#> ind16  -3.68693755  2.96666932 -2.58108490  2.43131759  1.785189275  2.27539955
#> ind17  -1.81504201 -3.55936296  1.30782308  0.07799710 -2.692719407  2.76271441
#> ind18   1.50646330  0.59903745  0.49965640  3.09629583  1.097493729  0.29120412
#> ind19  -0.19825567  3.65559529  2.01471004  2.41262669 -0.544028793  0.57367970
#> ind20  -0.16800876  2.24283212  1.68653763  0.75125896  2.044546149 -0.48492005
#> ind21   2.78067737  0.60017692 -1.06045332  1.21253305  2.920455896 -2.53646612
#> ind22   0.93156011  1.88268335 -0.32706222  1.48171582  1.201905539 -1.00384345
#> ind23   1.81963568  1.90010102 -0.13393120  3.10423912 -0.200452951 -0.10129895
#> ind24  -1.01263748  2.12860023  2.38821878  0.83854237  1.914110784 -1.49173933
#> ind25   1.89915333 -0.75373363 -0.83139097  1.50559275  0.392841040 -2.36166010
#> ind26  -0.57850751 -0.51975915 -0.08158364  1.78764417 -0.003856669  2.00270664
#> ind27   1.33305435  3.62678959  3.60056495 -0.81228410  1.974687722  0.37890020
#> ind28   0.24177361  5.00357419 -1.48184638 -0.50558032 -1.014050035 -0.11323139
#> ind29  -0.93105753  0.88076888  1.59522952 -0.41900554 -2.196429099  1.26828693
#> ind30   2.62251089 -0.52654293 -1.28219745  0.84219057  0.220423986 -2.95893395
#> ind31   1.40631246  2.77224890 -0.78901236 -0.40825350 -1.379327928  0.81505819
#> ind32  -1.66048619  2.39053313 -2.67797194 -0.92941973 -2.055465043  0.86881730
#> ind33  -0.51607485  2.56400974  0.99482875 -0.66731405  0.488982352 -0.92212940
#> ind34   0.71019993 -0.43416824  0.13489035  1.64945350  1.630118331 -1.15345534
#> ind35  -1.09655758  0.25397991 -0.12268660  2.96339941 -4.428537690  2.05631620
#> ind36   2.13906239  0.27021301  0.15362745  2.47824047 -1.852865618  0.22892594
#> ind37  -0.64547463  5.84753908 -2.90307426 -1.32781963  0.816366025 -0.59943012
#> ind38  -2.69205857  3.26291820  2.19262811 -0.23146371 -1.276486975  1.65702297
#> ind39  -2.93566128  1.50397102 -2.98925596 -2.32235875  3.669176521 -0.19426320
#> ind40  -3.25779408  2.68145187  1.95552555 -3.00617170 -0.818966753  1.87011623
#> ind41  -2.69390765  2.59033349  2.04668318 -0.79852723 -3.792054969  1.00640174
#> ind42  -0.41210492  3.36064447  1.00786018  0.99506954  0.614433414 -0.18843801
#> ind43   0.46145878  3.64120919  1.45805293 -3.01891418  0.141802281 -0.20011909
#> ind44   1.92294336  3.45163454 -1.12592988 -0.42876228 -0.937527656 -0.38497418
#> ind45   0.95494129  1.62562235 -0.99019752  2.19350953 -0.906152353  1.79945893
#> ind46  -0.34896933  2.02315537  3.76778512 -0.79867628 -1.122928500 -0.55520743
#> ind47   2.12434568  3.40422156  2.02174171 -0.88974352 -1.041550679 -2.10937215
#> ind48  -2.69894200  3.69574942 -0.79845098  1.49271351  2.037873582  1.69897398
#> ind49   0.53309770  1.52690018  2.43814839 -1.23114222  1.206304134 -1.67778937
#> ind50   0.57014376 -1.86840067  0.75084221 -1.50611116  3.713923525  3.84446980
#> ind51   4.15344739 -3.21678826 -2.05397162  1.54078104  0.928016478  0.30542782
#> ind52   1.41132444 -0.38519347 -1.69625311 -0.64056229 -1.130866949 -0.42659954
#> ind53   2.11526741 -5.04569202  2.36204810  0.32343110  0.598493965  1.59555419
#> ind54   4.66763651 -2.99608418 -0.46833536 -1.46524967 -0.807087857  0.07292094
#> ind55   5.04796230  0.01433518 -1.76711211  1.59476065 -0.671169827  4.51812002
#> ind56   1.49066529 -0.67408852 -3.19005435  1.42028123  0.886884014  1.79714629
#> ind57   1.21091535 -1.53922845  0.80277360 -1.28161408 -3.257346281  0.97054295
#> ind58   1.66790157 -0.65009874  2.82921039 -0.17499277 -0.259905120  0.42064649
#> ind59           NA          NA          NA          NA           NA          NA
#> ind60   0.30397742 -3.68132961 -1.48080529 -0.31871299  3.161409428  2.26061690
#> ind61   2.98906253 -3.46555633  1.98844286  1.81343494 -0.729333458  1.13129384
#> ind62   2.16802137  0.08339744 -0.75336341 -3.86076046  1.816426702 -2.24527747
#> ind63   2.56707554 -2.37071777  1.43874922 -0.04036490 -1.949304700 -1.41325513
#> ind64   1.30466485 -2.43266666 -1.35095550 -3.21700365 -0.869459827  0.40222400
#> ind65   1.55803782 -0.87208925  0.72940706 -1.73128371 -0.248186717 -0.03204947
#> ind66   6.24562589 -1.08003423 -1.54925179 -1.38894028 -0.109647304  0.41863260
#> ind67   0.24147196 -4.74423375  1.67221622  2.21013804  1.545125530 -1.05991700
#> ind68   4.28462627 -2.22596400 -1.13194429  1.78895838 -0.199643824 -3.34654960
#> ind69   0.61266701 -3.32704823  4.28057230 -1.90853673  1.444115902 -0.46774462
#> ind70   3.95446112 -4.73225560 -1.74348430  0.13363263 -2.031101179 -1.29398417
#> ind71   2.07173730 -3.92974822  1.55601794 -2.83516270  0.823776172  1.16067409
#> ind72   5.12265287  1.17586318 -2.88295057  1.21083648 -1.440114578  0.07511080
#> ind73   1.43980271 -1.72609969  0.22428679  2.87774114  3.286963485  2.44720205
#> ind74   0.63696089 -2.98060983  1.57625175 -0.03985398  2.235023239  3.31052880
#> ind75   2.83858546 -1.82345626 -4.09529776 -1.00154639 -0.675762253  0.03801040
#> ind76  -4.57275606 -3.02027662 -1.30153880 -0.08109621 -2.092047825 -1.72375169
#> ind77  -1.27189166 -2.47409388 -3.06510964 -2.36903144 -1.332891927 -2.25790412
#> ind78  -6.05500503 -5.80930025  1.50778993  0.31912547  2.236360169 -3.05674174
#> ind79  -1.72174869 -2.15635440 -1.60264546  0.90084401 -2.857884567 -2.10510915
#> ind80  -3.36868541  0.73544524  3.76183339  0.46477264  0.516412638 -1.71616719
#> ind81  -1.39653432 -1.31979098 -1.04498196 -2.20514463  0.180086641  0.88307805
#> ind82  -3.82574934  1.33035224 -0.57455538 -1.05591736 -0.226632091 -1.80942985
#> ind83  -3.31022624  1.15694900  0.34246658  0.97666618  0.503110133 -1.22186904
#> ind84  -3.89724105 -3.95391522 -0.76987280  1.51875487 -1.687067694 -0.51187037
#> ind85  -2.20140986  1.89322611  2.29580558  2.35900978 -0.393892452  0.20350258
#> ind86  -1.76613825 -1.19948039 -2.00954483 -0.51720062  1.004894967  0.75722664
#> ind87  -5.61751665 -3.50005742 -0.63008022 -2.81906468 -1.958178167  1.30461407
#> ind88  -4.36838364 -1.55777940 -0.04974734  0.72500579  0.506970050 -2.45144115
#> ind89  -2.02926012 -2.59377545  1.44006836  2.55942521 -1.296777286  1.55695553
#> ind90  -3.80434508 -0.50486563 -0.11865862 -0.86617742  0.827561680  1.21344579
#> ind91  -2.16609140 -2.64244457  1.34122446 -0.30304112 -0.556891251  1.68428267
#> ind92  -5.89813196  2.14989750 -4.12770612 -4.80008843  1.454882165 -0.53840634
#> ind93  -6.35226296 -1.10086784 -1.75333089 -2.67202406 -0.484168473  2.35552122
#> ind94  -2.67110699  1.80489947 -3.72359018  2.24484578  0.869228261 -1.80598515
#> ind95  -2.37692871 -0.20359675 -3.08686096  0.63809871  2.260737930  1.75627825
#> ind96  -1.76768766 -3.17198416 -1.87119208 -0.33311683  0.937720202 -2.18834736
#> ind97  -2.48619916 -1.68898365  3.04627337 -0.69193523 -2.928976896 -2.11042264
#> ind98  -3.34657648 -4.59623977  0.39670358 -0.06342294  1.803592319 -0.15882528
#> ind99  -3.42397316  1.31052490 -1.31988511  0.28877508 -0.961923838  1.06165568
#> ind100 -1.90652873 -3.32689517  0.77477185 -0.07052946  1.751590096 -1.21555562
#>           prob_pc7    prob_pc8    prob_pc9    prob_pc10
#> ind1    1.64314260 -2.35890286 -1.62468952 -3.590045055
#> ind2    1.85064585  0.35596577 -1.00824638 -2.154831380
#> ind3    0.33249623 -1.77319779 -1.68069088  2.153346566
#> ind4   -2.25905901 -0.61923439 -0.23683798 -0.001227868
#> ind5   -0.19461029  1.20295143  0.94539472  2.641102655
#> ind6    0.29057566 -1.17122648  0.60693899 -1.071309982
#> ind7    1.38011603 -1.14064551 -2.12634198  0.135405354
#> ind8   -1.03388355 -1.18999475  1.04585100  1.073639762
#> ind9            NA          NA          NA           NA
#> ind10   0.30140882  0.94984906  1.89235379 -0.061938167
#> ind11   1.75543143 -0.00995555  1.46694890 -3.084524116
#> ind12  -0.88428089  2.45806156 -0.40359974  0.744741581
#> ind13   0.96773117 -0.16786303  0.58511050 -1.552447316
#> ind14  -0.60765844  0.01493405  1.10481670  0.813876990
#> ind15   0.56227484  1.97426346 -0.44323298  1.472251861
#> ind16  -0.57113628 -1.05798813 -0.94473114  0.250605564
#> ind17  -0.42860471  1.78064682 -2.41008263 -0.473739235
#> ind18   2.24663312  0.75835274 -0.36250666  0.739388004
#> ind19  -0.13387801 -0.31147388 -2.20777112 -0.169363045
#> ind20   1.38810676 -0.37740830  1.32106934 -0.422320968
#> ind21  -1.92505946 -1.05545609  0.02779357 -1.250645884
#> ind22  -0.59712205  2.05959715  2.74565006 -0.002228600
#> ind23  -0.20170258  0.01804098 -1.78290634 -1.737852245
#> ind24   2.13582017 -1.25548726  2.07343535  0.859912288
#> ind25   2.15639344 -0.98287621 -0.31079116  0.727037155
#> ind26  -0.88726311  1.20925292 -0.96794560 -0.339711434
#> ind27  -1.22429642  0.00794552  1.86192735  3.284418524
#> ind28   0.25450198 -0.58712288  2.15177771  1.255884187
#> ind29  -1.31371664  0.48733772 -1.15369825  0.561343105
#> ind30  -0.59974274  2.02484416 -1.47429956  3.416463056
#> ind31  -1.40492598  1.53641948  1.28769311 -2.375416960
#> ind32  -0.41742761  1.22180712 -0.66744585 -0.529299017
#> ind33   0.60156757 -1.31725440  0.74221414  1.734361110
#> ind34  -0.37135387 -0.88531277 -0.98278138  1.527584640
#> ind35   5.99556836 -0.56675263  2.10921627 -1.349983119
#> ind36  -1.13562442 -0.06150302  0.93493250 -1.748272578
#> ind37   3.74768263  2.59213807 -0.85046684 -0.976632646
#> ind38   2.61065627 -1.20637797  0.31208653 -1.483670174
#> ind39   0.70427817  1.75102893  1.74847256 -0.395339462
#> ind40  -2.54032321 -1.69101017  1.95228889  0.127065817
#> ind41   0.78190037 -1.44441328 -0.25984312  2.897782137
#> ind42  -0.48083724  0.84702653 -0.12749680  0.206250675
#> ind43  -2.83332398  1.31168001 -0.55802843 -1.746267231
#> ind44  -1.21306242  0.79510701  2.52695276 -0.344158819
#> ind45   0.17804833  0.87092947 -0.67885613  1.103768936
#> ind46  -1.48993131  0.41810568 -2.16714727 -1.249032524
#> ind47   0.31088090  1.56261311  0.05603408  0.045563867
#> ind48   1.31030163  1.24121517  1.95600121 -0.162756344
#> ind49  -0.11239433  1.12842295 -0.80415471 -0.145649842
#> ind50  -2.79419006 -0.36438897 -1.31275917 -0.175913857
#> ind51  -2.47614035 -3.61234307  1.81696761  1.073829571
#> ind52  -0.51433116  0.19081783 -0.02173861 -1.027854423
#> ind53   2.11588607  0.58559382 -1.92012660  0.288606811
#> ind54   1.46478559 -1.71117742  0.45983287  2.702229806
#> ind55  -1.62828077  0.94001237  1.75143907  1.898088399
#> ind56   1.69304119 -1.83809393  0.40158499 -0.319485093
#> ind57   1.59020616  0.87987548  2.54539205 -0.293434906
#> ind58  -1.75682996 -1.00073042  0.52629291 -1.125701960
#> ind59           NA          NA          NA           NA
#> ind60   1.33741387 -0.03613437  2.36219628  0.210703886
#> ind61   1.89469775 -0.34862616 -1.82161416 -0.452738636
#> ind62   0.36965271 -2.28581695  0.61920256 -1.432307042
#> ind63  -0.16719237 -2.24461631 -0.05129502  0.472523749
#> ind64   2.21509895  3.35049010  1.23782656 -0.317380813
#> ind65  -0.41011419 -0.26324888 -0.61631238  3.793853067
#> ind66  -0.86498279 -0.76820981  0.34762514 -0.645719310
#> ind67   1.01801201  1.70195306 -1.26535896  1.737132959
#> ind68  -0.33011206  2.34251823  0.16929750  3.455845323
#> ind69   1.23998787 -0.68557134  2.52621596 -0.816597552
#> ind70  -1.96230660  0.49172429  1.60216040  0.101746078
#> ind71   1.96315768  2.24480825 -1.75826587  1.215331194
#> ind72  -3.12044038 -0.20186610 -2.02105316 -2.318232288
#> ind73  -0.77028322  3.25087926  0.49332608  0.509433420
#> ind74   0.90565686  0.64525891 -2.40823041  1.010963473
#> ind75   0.13235084 -0.49628007 -1.32888189 -0.210657205
#> ind76  -1.15615894  4.15743994  1.41620758 -0.489636044
#> ind77  -0.30163770 -1.35076020  1.76427445  1.143259642
#> ind78   1.73435331 -0.29932160 -0.50555788 -0.344889560
#> ind79  -2.17869243 -2.51615920 -0.18676279  0.632483255
#> ind80   0.59544287  1.07621996  0.24502600 -0.779288672
#> ind81  -1.97095092  2.75433370 -3.47136606  0.115771661
#> ind82   1.58840678 -1.19052110 -1.88430918  1.274906207
#> ind83   0.60143186 -0.99227702 -1.45576456 -0.976781000
#> ind84  -1.07059113 -0.59058616  1.66919582 -0.492074215
#> ind85  -4.36120708 -0.55106450  0.75801714  0.145789630
#> ind86   0.80390143 -1.99297225  0.06532676  0.415749983
#> ind87   0.63914185 -1.16576873  1.61195919 -0.866216665
#> ind88  -0.41208621 -0.62368928 -0.89952896 -2.196464273
#> ind89   0.08414955 -1.17823229  0.67742927 -0.518508392
#> ind90   0.74670641 -3.15798629 -2.45215428  1.691126207
#> ind91  -1.94389053 -0.09158413  0.73762380 -3.390029940
#> ind92   1.17315383  2.43798193 -1.98949369 -1.107695905
#> ind93  -0.48632373 -2.17698038  0.38256330  1.659735139
#> ind94   0.17108303 -2.34061528 -2.08508220 -0.057356675
#> ind95   0.36377540 -0.32331018  1.07032544 -0.025782793
#> ind96  -0.15757801 -0.81571713 -0.61095286  0.048339553
#> ind97  -0.85023050  0.91864693 -1.38008950 -0.486030003
#> ind98  -1.02575680 -0.90016202 -0.60619020 -2.378577263
#> ind99  -0.72804407 -1.10635474 -2.07179181 -0.616017346
#> ind100 -1.69582400  1.98648189  1.77518754 -1.088240879