Retrieve Sample Summary Data

Extracts sample-level summary statistics from a Metabolites object for a specified data layer, with optional exclusions and subsetting.

Usage

get_sample_summary(
  metabolites,
  layer,
  sample_ids = NULL,
  drop = FALSE,
  as_df = FALSE
)

Arguments

metabolites

A Metabolites object containing sample summary data.

layer

A character string specifying the data layer from which to extract sample summary data.

sample_ids

Optional. A vector of column names (samples) to extract. Defaults to NULL, meaning all samples are included.

drop

Logical. Whether to drop all NA (excluded samples)

as_df

Logical. Whether to return the result as a data.frame. Defaults to FALSE, meaning the result is returned as a matrix.

Value

A matrix or data.frame (if as_df = TRUE) containing the sample summary data for the specified layer.

Examples

file <- system.file("extdata", "metabolon_v1_example.xlsx", package = "metaboprep2")
metabolites <- Metabolites(project_name = "MYPROJECT",
                           format       = "metabolon_v1",
                           filepath     = file)
metabolites <- import_data(metabolites)
#> Warning: 86 unmapped features:
#>  - 123, 124, 125, 126, 127, 128, 129, 130, 131, 132, 133, 134, 135, 137, 138, 139, 140, 141, 142, 143, 145, 146, 147, 148, 149, 150, 151, 152, 153, 154, 155, 156, 157, 159, 160, 161, 162, 163, 164, 165, 166, 167, 168, 169, 170, 172, 173, 174, 175, 176, 177, 178, 183, 184, 185, 186, 187, 188, 190, 191, 193, 195, 196, 198, 199, 200, 201, 202, 204, 205, 206, 207, 208, 209, 210, 211, 213, 214, 215, 216, 217, 218, 219, 220, 221, 222
metabolites <- metabolite_qc(metabolites, source="raw")
#> 
#> ── Starting Metabolite QC Process ──────────────────────────────────────────────
#> ℹ Validating input parameters...
#> ✔ Input validation complete.
#> ℹ Applying exclusion of xenobiotic features...
#> ✔ Xenobiotic features exclusions assessment complete. (3 features excluded)
#> ℹ Analysing exclusions...
#> ℹ Applying exclusion of derived features...
#> ✔ Derived features exclusions assessment complete. (0 features excluded)
#> ℹ Sample & Feature Summary Statistics for raw data...
#> ✔ Sample & Feature Summary Statistics for raw data...
#> ℹ Copying raw data to new 'qcing' data layer...
#> ✔ qcing data ready for QC.
#> ℹ Assessing for extreme sample missingness >=80%...
#> ✔ Extreme sample missingness exclusions assessment complete. (0 samples excluded)
#> ℹ Assessing for extreme feature missingness >=80%...
#> ✔ Extreme feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Assessing for sample missingness at specified level of >=20%...
#> ✔ Sample missingness exclusions assessment complete. (2 samples excluded)
#> ℹ Assessing for feature missingness at specified level of >=20%...
#> ✔ Feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Calculating total peak abundance outliers at +/- 5 Sdev...
#> ✔ Total peak abundance outlier exclusions assessment complete. (0 samples excluded)
#> ℹ Running data outlier analysis at +/- 5 Sdev...
#> ℹ Data outlier analysis complete.
#> ℹ Re-identify feature independence and PC outliers...
#> ℹ Re-running sample summary using current exclusions...
#> ℹ QC sample summary complete.
#> ℹ PCA outlier analysis 2 complete.
#> ℹ Creating final QC dataset...
#> ℹ Rerunning Sample & Feature Summary Statistics for QC data...
#> ✔ Sample & Feature Summary Statistics for QC data complete.
#> 
#> ── Metabolite QC Process Completed ─────────────────────────────────────────────
get_sample_summary(metabolites, layer = "post_qc")
#>        missingness missingness_w_exclusions tpa_total tpa_complete_features
#> ind1          0.00                        0  429.3981              418.1846
#> ind2          0.00                        0  445.8317              432.1929
#> ind3          0.02                        0  305.6836              300.3210
#> ind4          0.00                        0  330.2005              320.3710
#> ind5          0.00                        0  350.4338              340.8450
#> ind6          0.00                        0  323.3249              312.7040
#> ind7          0.00                        0  331.3949              321.6856
#> ind8          0.00                        0  348.0044              337.8921
#> ind9            NA                       NA        NA                    NA
#> ind10         0.00                        0  357.0769              348.1124
#> ind11         0.00                        0  347.4778              336.1425
#> ind12         0.01                        0  316.7135              311.2105
#> ind13         0.00                        0  299.7482              290.3293
#> ind14         0.00                        0  336.7379              327.6420
#> ind15         0.01                        0  300.6866              294.7594
#> ind16         0.00                        0  293.8122              284.0910
#> ind17         0.00                        0  361.8988              352.9619
#> ind18         0.01                        0  347.2309              341.4984
#> ind19         0.01                        0  312.1795              306.2072
#> ind20         0.00                        0  334.9327              325.2223
#> ind21         0.01                        0  347.6117              340.8737
#> ind22         0.01                        0  328.7270              322.1908
#> ind23         0.01                        0  337.3916              333.0592
#> ind24         0.01                        0  321.7626              316.2721
#> ind25         0.01                        0  347.4914              342.7891
#> ind26         0.02                        0  338.3327              335.7303
#> ind27         0.01                        0  317.4020              315.6527
#> ind28         0.01                        0  307.3521              301.3430
#> ind29         0.00                        0  329.5923              319.2297
#> ind30         0.00                        0  359.1920              347.7323
#> ind31         0.01                        0  326.3588              320.8185
#> ind32         0.01                        0  314.4292              306.2326
#> ind33         0.00                        0  320.5111              311.7081
#> ind34         0.01                        0  348.1397              340.9728
#> ind35         0.00                        0  339.5943              329.9637
#> ind36         0.00                        0  356.3353              345.5204
#> ind37         0.00                        0  301.1192              291.2797
#> ind38         0.01                        0  310.7453              303.6725
#> ind39         0.01                        0  318.9477              314.0924
#> ind40         0.01                        0  307.0884              301.5534
#> ind41         0.00                        0  311.0102              301.3099
#> ind42         0.00                        0  318.0804              307.9378
#> ind43         0.00                        0  323.0416              310.1692
#> ind44         0.01                        0  309.5491              305.7485
#> ind45         0.00                        0  341.2931              334.1877
#> ind46         0.01                        0  327.6826              323.6172
#> ind47         0.00                        0  323.4631              314.7083
#> ind48         0.00                        0  300.3607              291.4474
#> ind49         0.00                        0  338.5495              326.1048
#> ind50         0.00                        0  373.3480              362.8027
#> ind51         0.00                        0  403.8585              395.0482
#> ind52         0.01                        0  351.5407              343.2161
#> ind53         0.00                        0  406.9835              395.8279
#> ind54         0.01                        0  406.1964              401.0412
#> ind55         0.00                        0  383.5741              377.1375
#> ind56         0.00                        0  368.1418              359.0292
#> ind57         0.00                        0  380.3077              367.5933
#> ind58         0.00                        0  372.2861              362.5520
#> ind59           NA                       NA        NA                    NA
#> ind60         0.00                        0  379.0648              368.3538
#> ind61         0.00                        0  404.6702              394.0665
#> ind62         0.01                        0  356.7894              348.3922
#> ind63         0.00                        0  382.0445              372.4209
#> ind64         0.00                        0  371.5869              359.1085
#> ind65         0.00                        0  363.0435              353.2881
#> ind66         0.01                        0  393.3985              385.2493
#> ind67         0.00                        0  385.9002              376.5585
#> ind68         0.00                        0  387.5416              376.1354
#> ind69         0.00                        0  380.6284              368.9180
#> ind70         0.00                        0  395.8335              384.5292
#> ind71         0.01                        0  399.1562              388.2660
#> ind72         0.00                        0  367.8114              355.9107
#> ind73         0.01                        0  363.2988              356.2844
#> ind74         0.01                        0  389.5995              381.5605
#> ind75         0.00                        0  377.2995              364.2316
#> ind76         0.01                        0  333.6146              324.1795
#> ind77         0.00                        0  358.5985              346.0571
#> ind78         0.00                        0  356.8877              345.2503
#> ind79         0.00                        0  355.5831              345.2696
#> ind80         0.00                        0  319.9851              307.8824
#> ind81         0.02                        0  338.6808              333.9001
#> ind82         0.00                        0  307.9087              296.3680
#> ind83         0.01                        0  308.5287              301.5340
#> ind84         0.00                        0  351.9746              343.1257
#> ind85         0.00                        0  325.4163              313.7341
#> ind86         0.02                        0  338.6869              334.6070
#> ind87         0.00                        0  338.2676              327.8050
#> ind88         0.01                        0  321.7062              314.1764
#> ind89         0.01                        0  356.5756              349.5830
#> ind90         0.01                        0  329.1736              324.0362
#> ind91         0.00                        0  363.1633              349.1504
#> ind92         0.01                        0  288.2985              277.5573
#> ind93         0.00                        0  325.6924              314.3378
#> ind94         0.00                        0  309.1097              299.8323
#> ind95         0.01                        0  334.2227              327.1458
#> ind96         0.01                        0  356.1463              348.7503
#> ind97         0.00                        0  349.0920              336.2951
#> ind98         0.01                        0  360.7229              351.4024
#> ind99         0.01                        0  311.6307              303.4713
#> ind100        0.00                        0  356.6426              344.5812
#>        outlier_count
#> ind1               1
#> ind2               1
#> ind3               0
#> ind4               0
#> ind5               0
#> ind6               0
#> ind7               0
#> ind8               0
#> ind9              NA
#> ind10              0
#> ind11              0
#> ind12              0
#> ind13              0
#> ind14              0
#> ind15              0
#> ind16              0
#> ind17              0
#> ind18              0
#> ind19              0
#> ind20              0
#> ind21              0
#> ind22              0
#> ind23              0
#> ind24              0
#> ind25              0
#> ind26              0
#> ind27              0
#> ind28              0
#> ind29              0
#> ind30              0
#> ind31              0
#> ind32              0
#> ind33              0
#> ind34              0
#> ind35              0
#> ind36              0
#> ind37              0
#> ind38              0
#> ind39              0
#> ind40              0
#> ind41              0
#> ind42              0
#> ind43              0
#> ind44              0
#> ind45              0
#> ind46              0
#> ind47              0
#> ind48              0
#> ind49              0
#> ind50              0
#> ind51              0
#> ind52              0
#> ind53              0
#> ind54              0
#> ind55              0
#> ind56              0
#> ind57              0
#> ind58              0
#> ind59             NA
#> ind60              0
#> ind61              0
#> ind62              0
#> ind63              0
#> ind64              0
#> ind65              0
#> ind66              0
#> ind67              0
#> ind68              0
#> ind69              0
#> ind70              0
#> ind71              0
#> ind72              0
#> ind73              0
#> ind74              0
#> ind75              0
#> ind76              0
#> ind77              0
#> ind78              0
#> ind79              0
#> ind80              0
#> ind81              0
#> ind82              0
#> ind83              0
#> ind84              0
#> ind85              0
#> ind86              0
#> ind87              0
#> ind88              0
#> ind89              0
#> ind90              0
#> ind91              0
#> ind92              0
#> ind93              0
#> ind94              0
#> ind95              0
#> ind96              0
#> ind97              0
#> ind98              0
#> ind99              0
#> ind100             0