Retrieve Variance Explained by PCs

Extracts the variance explained by each principal components (PCs) from a Metabolites object for a specified data layer, with optional exclusions and subsetting.

Usage

get_var_exp(metabolites, layer, as_df = FALSE)

Arguments

metabolites

A Metabolites object containing the principal component data.

layer

A character string specifying the data layer from which to extract the PCs.

as_df

Logical. Whether to return the result as a data.frame. Defaults to FALSE, returning a matrix.

Value

A matrix or data.frame (if as_df = TRUE) containing the principal components.

Examples

file <- system.file("extdata", "metabolon_v1_example.xlsx", package = "metaboprep2")
metabolites <- Metabolites(project_name = "MYPROJECT",
                           format       = "metabolon_v1",
                           filepath     = file)
metabolites <- import_data(metabolites)
#> Using cached MetaboAnalyst DB.
#> Attempting join [features$hmdb] ~ [annotation$hmdb_id] 
#> Attempting join [features$kegg] ~ [annotation$kegg_id] 
#> Attempting join [features$metabolite_id] ~ [annotation$name] 
#> Attempting join [features$comp_id] ~ [annotation$comp_id] 
#> Warning: 100 unannotated metabolites remain after merging with metabolomics database
metabolites <- metabolite_qc(metabolites, source="raw")
#> 
#> ── Starting Metabolite QC Process ──────────────────────────────────────────────
#> ℹ Validating input parameters...
#> ✔ Input validation complete.
#> ℹ Batch normalising raw data...
#> ✔ Normalised raw data.
#> ℹ Applying exclusion of xenobiotic features...
#> ✔ Xenobiotic features exclusions assessment complete. (3 features excluded)
#> ℹ Analysing exclusions...
#> ℹ Applying exclusion of derived features...
#> ✔ Derived features exclusions assessment complete. (0 features excluded)
#> ℹ Sample & Feature Summary Statistics for raw data...
#> ✔ Sample & Feature Summary Statistics for raw data...
#> ℹ Copying batch_normalised data to new 'qcing' data layer...
#> ✔ qcing data ready for QC.
#> ℹ Assessing for extreme sample missingness >=80%...
#> ✔ Extreme sample missingness exclusions assessment complete. (0 samples excluded)
#> ℹ Assessing for extreme feature missingness >=80%...
#> ✔ Extreme feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Assessing for sample missingness at specified level of >=20%...
#> ✔ Sample missingness exclusions assessment complete. (2 samples excluded)
#> ℹ Assessing for feature missingness at specified level of >=20%...
#> ✔ Feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Calculating total peak abundance outliers at +/- 5 Sdev...
#> ✔ Total peak abundance outlier exclusions assessment complete. (0 samples excluded)
#> ℹ Running data outlier analysis at +/- 5 Sdev...
#> ℹ Data outlier analysis complete.
#> ℹ Re-identify feature independence and PC outliers...
#> ℹ Re-running sample summary using current exclusions...
#> ℹ QC sample summary complete.
#> ℹ PCA outlier analysis 2 complete.
#> ℹ Creating final QC dataset...
#> ℹ Rerunning Sample & Feature Summary Statistics for QC data...
#> ✔ Sample & Feature Summary Statistics for QC data complete.
#> 
#> ── Metabolite QC Process Completed ─────────────────────────────────────────────
get_var_exp(metabolites, layer = "post_qc")
#>     PC1     PC2     PC3     PC4     PC5     PC6     PC7     PC8     PC9    PC10 
#> 0.07369 0.06196 0.04935 0.04255 0.04059 0.03683 0.03473 0.03152 0.03019 0.02847 
#>    PC11    PC12    PC13    PC14    PC15    PC16    PC17    PC18    PC19    PC20 
#> 0.02707 0.02529 0.02489 0.02379 0.02319 0.02078 0.02074 0.01959 0.01867 0.01839 
#>    PC21    PC22    PC23    PC24    PC25    PC26    PC27    PC28    PC29    PC30 
#> 0.01763 0.01663 0.01587 0.01517 0.01454 0.01387 0.01370 0.01318 0.01240 0.01222 
#>    PC31    PC32    PC33    PC34    PC35    PC36    PC37    PC38    PC39    PC40 
#> 0.01157 0.01054 0.01014 0.00951 0.00924 0.00916 0.00799 0.00788 0.00761 0.00733 
#>    PC41    PC42    PC43    PC44    PC45    PC46    PC47    PC48    PC49    PC50 
#> 0.00702 0.00651 0.00634 0.00580 0.00567 0.00532 0.00502 0.00482 0.00455 0.00447 
#>    PC51    PC52    PC53    PC54    PC55    PC56    PC57    PC58    PC59    PC60 
#> 0.00423 0.00402 0.00352 0.00340 0.00331 0.00305 0.00284 0.00279 0.00259 0.00242 
#>    PC61    PC62    PC63    PC64    PC65    PC66    PC67    PC68    PC69    PC70 
#> 0.00230 0.00207 0.00202 0.00184 0.00181 0.00166 0.00152 0.00131 0.00126 0.00112 
#>    PC71    PC72    PC73    PC74    PC75    PC76    PC77    PC78    PC79    PC80 
#> 0.00102 0.00092 0.00081 0.00069 0.00063 0.00056 0.00049 0.00041 0.00033 0.00030 
#>    PC81    PC82    PC83    PC84    PC85    PC86    PC87 
#> 0.00025 0.00017 0.00013 0.00011 0.00007 0.00003 0.00002