Estimates total peak abundance

This function estimates total peak abundance|area for numeric data in a matrix, for (1) all features and (2) all features with complete data.

Usage

total_peak_area(data, features_exclude = NA_character_, ztransform = TRUE)

Arguments

data

matrix, the metabolite data matrix. Samples in rows, metabolites in columns

features_exclude

character, a vector of feature|metabolite names to exclude from the tree building, independent feature identification process.

ztransform

logical, should the feature data be z-transformed and absolute value minimum, mean shifted prior to summing the feature values. TRUE or FALSE.

Value

a data frame of estimates for (1) total peak abundance and (2) total peak abundance at complete features for each samples

Examples

set.seed(1110)
ex_data = sapply(1:5, function(x){ rnorm(10, 40, 5) })
rownames(ex_data) = paste0("ind", 1:nrow(ex_data))
ex_data[ sample(1:50, 4) ] = NA
tpa_est = total_peak_area(ex_data)