Metabolites Object

A Metabolites object is a container for matrices of metabolite data, along with associated metadata. It allows for efficient storage and manipulation of data, supporting quality control, transformations, and various analyses. This object facilitates easy access to data layers, sample and feature summaries, outlier treatment, and more.

Usage

Metabolites(
  project_name = "",
  created = Sys.time(),
  filepath = character(0),
  format = "metabolon_v1",
  source = character(0),
  feature_missingness = 0.2,
  sample_missingness = 0.2,
  total_peak_area_sd = 5,
  outlier_udist = 5,
  outlier_treatment = "leave_be",
  winsorize_quantile = 1,
  tree_cut_height = 0.5,
  pc_outlier_sd = 5,
  derived_var_exclusion = TRUE,
  xenobiotics_var_exclusion = TRUE,
  samples = data.table::data.table(sample_id = character()),
  features = data.table::data.table(feature_id = character(), platform = character(),
    pathway = character(), derived_feature = logical()),
  data = integer(0),
  exclusions = list(raw = list(samples = list(extreme_sample_missingness = character(0),
    user_defined_sample_missingness = character(0), user_defined_sample_totalpeakarea =
    character(0), user_defined_sample_pca_outlier = character(0)), features =
    list(extreme_feature_missingness = character(0), user_defined_feature_missingness =
    character(0)))),
  feature_summary = integer(0),
  feature_tree = list(),
  sample_summary = integer(0),
  pcs = integer(0),
  prob_pcs = integer(0),
  var_exp = integer(0),
  acceleration_factor = integer(0),
  n_parallel = integer(0)
)

Arguments

project_name

character, the name of the project associated with the data.

created

POSIXct, the timestamp when the Metabolites object was created.

filepath

character, the file path to the original raw data file.

format

character, the format of the metabolomic data (e.g., "metabolon_v1").

source

named character, the source for this data layer (e.g., 'raw' for 'post_qc').

feature_missingness

numeric, the proportion (0-1) of missingness in features, used as a QC threshold. Default is 0.2.

sample_missingness

numeric, the proportion (0-1) of missingness in samples, used as a QC threshold. Default is 0.2.

total_peak_area_sd

numeric, the number of standard deviations from the mean to use for sample QC based on total abundance. Default is 5. If set to NA, total sum abundance will not be used for QC.

outlier_udist

numeric, the number of interquartile range (IQR) unit distances from the median used to identify outliers for features. Default is 5.

outlier_treatment

character, specifies how to treat outliers for PCA analysis. Options: "leave_be" (no action on outliers), "winsorize" (winsorize to the 100th quantile), "turn_NA" (convert outliers to NA, which will be imputed to the median).

winsorize_quantile

numeric, the quantile of remaining values to use for winsorizing outliers. Default is 1.

tree_cut_height

numeric, the threshold for feature independence in hierarchical clustering. Default is 0.5.

pc_outlier_sd

numeric, the number of standard deviations from the mean used for PCA-based sample QC. Default is 5. Set to NA to exclude PCA-based QC.

derived_var_exclusion

logical, whether to exclude derived variables (from multiple existing variables) in Nightingale data. Default is TRUE.

xenobiotics_var_exclusion

logical, whether to exclude xenobiotics from missingness estimates. Default is TRUE.

samples

data.table, a data table containing sample-related information (not to be set directly).

features

data.table, a data table containing feature-related information (not to be set directly).

data

numeric matrix, the data matrix containing metabolite values (not to be set directly).

exclusions

list, holds exclusion codes for data masking (not to be set directly).

feature_summary

numeric matrix, summary statistics for features (not to be set directly).

feature_tree

named list of tree objects, hierarchical clustering results for features (not to be set directly).

sample_summary

numeric matrix, summary statistics for samples (not to be set directly).

pcs

numeric matrix, principal component scores for the data (not to be set directly).

prob_pcs

numeric matrix, probabilities associated with principal components (not to be set directly).

var_exp

named numeric, variance explained by each principal component (not to be set directly).

acceleration_factor

named numeric, the acceleration factor for parallel analysis (not to be set directly).

n_parallel

named numeric, the number of parallel components in the analysis (not to be set directly).

Value

An object of class Metabolites (S7 class).

Slots

project_name

character, name for this project.

created

POSIXct, the timestamp when the Metabolites object was created.

filepath

character, path to the original raw data file.

format

character, a valid metabolomic data source.

samples

data.table, the samples data table.

features

data.table, the features data table.

data

numeric matrix, the metabolite data.

source

named character, the source for this data layer (e.g., 'raw' for 'post_qc').

exclusions

list, exclusion codes (mask for data).

feature_summary

numeric matrix, feature summary statistics.

sample_summary

numeric matrix, sample summary statistics.

feature_missingness

numeric, proportion of missingness in features (QC threshold).

sample_missingness

numeric, proportion of missingness in samples (QC threshold).

total_peak_area_sd

numeric, standard deviation for sample QC based on total abundance.

outlier_udist

numeric, outlier threshold in terms of IQR for features.

outlier_treatment

character, outlier treatment for PCA analysis.

winsorize_quantile

numeric, quantile used for winsorizing outliers.

tree_cut_height

numeric, threshold for feature independence in clustering.

pc_outlier_sd

numeric, standard deviation for PCA-based sample QC.

derived_var_exclusion

logical, whether to exclude derived variables from Nightingale data.

xenobiotics_var_exclusion

logical, whether to exclude xenobiotics from missingness estimate.

feature_tree

object, hierarchical clustering results for features.

pcs

matrix, principal component scores.

prob_pcs

matrix, probabilities for principal components.

var_exp

matrix, variance explained by each principal component.

acceleration_factor

numeric, acceleration factor for parallel analysis.

n_parallel

numeric, number of parallel components.

Examples

# Example of importing data into a Metabolites object and running QC
file <- system.file("extdata", "metabolon_v1_example.xlsx", package = "metaboprep2")
metabolites <- Metabolites(project_name = "MYPROJECT",
                           format       = "metabolon_v1",
                           filepath     = file)
metabolites <- import_data(metabolites)
#> Warning: 86 unmapped features:
#>  - 123, 124, 125, 126, 127, 128, 129, 130, 131, 132, 133, 134, 135, 137, 138, 139, 140, 141, 142, 143, 145, 146, 147, 148, 149, 150, 151, 152, 153, 154, 155, 156, 157, 159, 160, 161, 162, 163, 164, 165, 166, 167, 168, 169, 170, 172, 173, 174, 175, 176, 177, 178, 183, 184, 185, 186, 187, 188, 190, 191, 193, 195, 196, 198, 199, 200, 201, 202, 204, 205, 206, 207, 208, 209, 210, 211, 213, 214, 215, 216, 217, 218, 219, 220, 221, 222
metabolites <- metabolite_qc(metabolites, source="raw")
#> 
#> ── Starting Metabolite QC Process ──────────────────────────────────────────────
#> ℹ Validating input parameters...
#> ✔ Input validation complete.
#> ℹ Sample & Feature Summary Statistics for raw data...
#> ✔ Sample & Feature Summary Statistics for raw data...
#> ℹ Copying raw data to new 'qcing' data layer...
#> ✔ qcing data ready for QC.
#> ℹ Analysing exclusions...
#> ℹ Applying exclusion of derived features...
#> ✔ Derived features exclusions assessment complete. (0 features excluded)
#> ℹ Applying exclusion of xenobiotic features...
#> ✔ Xenobiotic features exclusions assessment complete. (3 features excluded)
#> ℹ Assessing for extreme sample missingness >=80%...
#> ✔ Extreme sample missingness exclusions assessment complete. (0 samples excluded)
#> ℹ Assessing for extreme feature missingness >=80%...
#> ✔ Extreme feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Assessing for sample missingness at specified level of >=20%...
#> ✔ Sample missingness exclusions assessment complete. (2 samples excluded)
#> ℹ Assessing for feature missingness at specified level of >=20%...
#> ✔ Feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Calculating total peak abundance outliers at +/- 5 Sdev...
#> ✔ Total peak abundance outlier exclusions assessment complete. (0 samples excluded)
#> ℹ Running principal component outlier analysis at +/- 5 Sdev...
#> ℹ PCA outlier analysis complete.
#> ℹ Re-running sample summary using current exclusions...
#> ℹ QC sample summary complete.
#> ℹ Re-identify feature independence and PC outliers...
#> ℹ PCA outlier analysis 2 complete.
#> ℹ Creating final QC dataset...
#> ℹ Rerunning Sample & Feature Summary Statistics for QC data...
#> ✔ Sample & Feature Summary Statistics for QC data complete.
#> 
#> ── Metabolite QC Process Completed ─────────────────────────────────────────────
metabolites
#> <metaboprep2::Metabolites>
#>  @ project_name             : chr "MYPROJECT"
#>  @ created                  : POSIXct[1:1], format: "2025-04-05 22:24:33"
#>  @ filepath                 : chr "/home/runner/work/_temp/Library/metaboprep2/extdata/metabolon_v1_example.xlsx"
#>  @ format                   : chr "metabolon_v1"
#>  @ source                   : Named chr [1:2] "raw" "qcing"
#>  .. - attr(*, "names")= chr [1:2] "qcing" "post_qc"
#>  @ feature_missingness      : num 0.2
#>  @ sample_missingness       : num 0.2
#>  @ total_peak_area_sd       : num 5
#>  @ outlier_udist            : num 5
#>  @ outlier_treatment        : chr "leave_be"
#>  @ winsorize_quantile       : num 1
#>  @ tree_cut_height          : num 0.5
#>  @ pc_outlier_sd            : num 5
#>  @ derived_var_exclusion    : logi TRUE
#>  @ xenobiotics_var_exclusion: logi TRUE
#>  @ samples                  :Classes ‘data.table’ and 'data.frame':	100 obs. of  7 variables:
#>  .. $ sample_id  : chr  "ind1" "ind2" "ind3" "ind4" ...
#>  .. $ sample_name: chr  "ind1" "ind2" "ind3" "ind4" ...
#>  .. $ neg        : chr  "batch1" "batch1" "batch1" "batch1" ...
#>  .. $ pos        : chr  "batch1" "batch1" "batch1" "batch1" ...
#>  .. $ run_day    : chr  "day1" "day1" "day1" "day1" ...
#>  .. $ box_id     : chr  "box1" "box1" "box1" "box1" ...
#>  .. $ lot        : chr  "lot250" "lot250" "lot250" "lot250" ...
#>  .. - attr(*, ".internal.selfref")=<externalptr> 
#>  @ features                 :Classes ‘data.table’ and 'data.frame':	100 obs. of  10 variables:
#>  .. $ feature_id     : chr  "unmapped_123" "unmapped_124" "unmapped_125" "unmapped_126" ...
#>  .. $ comp_id        : chr  "123" "124" "125" "126" ...
#>  .. $ metabolite_id  : chr  "met1" "met2" "met3" "met4" ...
#>  .. $ platform       : chr  NA NA NA NA ...
#>  .. $ pathway        : chr  "xenobiotic" "xenobiotic" "xenobiotic" "lipid" ...
#>  .. $ kegg           : chr  "K100" "K100" "K100" "K200" ...
#>  .. $ hmdb           : chr  "HMDB1" "HMDB2" "HMDB3" "HMDB4" ...
#>  .. $ derived_feature: logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
#>  .. $ id_source      : chr  NA NA NA NA ...
#>  .. $ chemical_name  : chr  NA NA NA NA ...
#>  .. - attr(*, ".internal.selfref")=<externalptr> 
#>  .. - attr(*, "index")= int(0) 
#>  ..  ..- attr(*, "__derived_feature")= int(0) 
#>  @ data                     : num [1:100, 1:100, 1:4] 11.03 12.5 NA 6.98 8.56 ...
#>  .. - attr(*, "dimnames")=List of 3
#>  ..  ..$ : chr [1:100] "ind1" "ind2" "ind3" "ind4" ...
#>  ..  ..$ : chr [1:100] "unmapped_123" "unmapped_124" "unmapped_125" "unmapped_126" ...
#>  ..  ..$ : chr [1:4] "raw" "scaled" "qcing" "post_qc"
#>  @ exclusions               :List of 3
#>  .. $ raw    :List of 2
#>  ..  ..$ samples :List of 4
#>  ..  .. ..$ extreme_sample_missingness       : chr(0) 
#>  ..  .. ..$ user_defined_sample_missingness  : chr(0) 
#>  ..  .. ..$ user_defined_sample_totalpeakarea: chr(0) 
#>  ..  .. ..$ user_defined_sample_pca_outlier  : chr(0) 
#>  ..  ..$ features:List of 2
#>  ..  .. ..$ extreme_feature_missingness     : chr(0) 
#>  ..  .. ..$ user_defined_feature_missingness: chr(0) 
#>  .. $ qcing  :List of 2
#>  ..  ..$ samples :List of 4
#>  ..  .. ..$ extreme_sample_missingness       : chr(0) 
#>  ..  .. ..$ user_defined_sample_missingness  : chr [1:2] "ind9" "ind59"
#>  ..  .. ..$ user_defined_sample_totalpeakarea: chr(0) 
#>  ..  .. ..$ user_defined_sample_pca_outlier  : chr(0) 
#>  ..  ..$ features:List of 2
#>  ..  .. ..$ extreme_feature_missingness     : chr(0) 
#>  ..  .. ..$ user_defined_feature_missingness: chr(0) 
#>  .. $ post_qc:List of 2
#>  ..  ..$ samples :List of 4
#>  ..  .. ..$ extreme_sample_missingness       : chr(0) 
#>  ..  .. ..$ user_defined_sample_missingness  : chr [1:2] "ind9" "ind59"
#>  ..  .. ..$ user_defined_sample_totalpeakarea: chr(0) 
#>  ..  .. ..$ user_defined_sample_pca_outlier  : chr(0) 
#>  ..  ..$ features:List of 2
#>  ..  .. ..$ extreme_feature_missingness     : chr(0) 
#>  ..  .. ..$ user_defined_feature_missingness: chr(0) 
#>  @ feature_summary          : num [1:21, 1:100, 1:3] 0.2 NA 0 80 10.5 ...
#>  .. - attr(*, "dimnames")=List of 3
#>  ..  ..$ : chr [1:21] "missingness" "missingness_w_exclusions" "outlier_count" "n" ...
#>  ..  ..$ : chr [1:100] "unmapped_123" "unmapped_124" "unmapped_125" "unmapped_126" ...
#>  ..  ..$ : chr [1:3] "raw" "qcing" "post_qc"
#>  @ feature_tree             :List of 3
#>  .. $ raw    :List of 7
#>  ..  ..$ merge      : int [1:99, 1:2] -11 -3 -75 -85 -54 -57 -55 -4 -65 -53 ...
#>  ..  ..$ height     : num [1:99] 0.149 0.155 0.21 0.218 0.26 ...
#>  ..  ..$ order      : int [1:100] 58 81 67 99 72 91 94 74 65 89 ...
#>  ..  ..$ labels     : chr [1:100] "unmapped_123" "unmapped_124" "unmapped_125" "unmapped_126" ...
#>  ..  ..$ method     : chr "complete"
#>  ..  ..$ call       : language stats::hclust(d = dist_matrix, method = "complete")
#>  ..  ..$ dist.method: NULL
#>  ..  ..- attr(*, "class")= chr "hclust"
#>  .. $ qcing  :List of 7
#>  ..  ..$ merge      : int [1:96, 1:2] -8 -72 -82 -51 -52 -1 -50 -54 -62 -88 ...
#>  ..  ..$ height     : num [1:96] 0.149 0.218 0.218 0.26 0.359 ...
#>  ..  ..$ order      : int [1:97] 55 78 64 96 69 88 91 71 62 86 ...
#>  ..  ..$ labels     : chr [1:97] "unmapped_126" "unmapped_127" "unmapped_128" "unmapped_129" ...
#>  ..  ..$ method     : chr "complete"
#>  ..  ..$ call       : language stats::hclust(d = dist_matrix, method = "complete")
#>  ..  ..$ dist.method: NULL
#>  ..  ..- attr(*, "class")= chr "hclust"
#>  .. $ post_qc:List of 7
#>  ..  ..$ merge      : int [1:99, 1:2] -8 -3 -72 -82 -51 -52 -1 -50 -54 -62 ...
#>  ..  ..$ height     : num [1:99] 0.149 0.155 0.218 0.218 0.26 ...
#>  ..  ..$ order      : int [1:100] 23 45 20 21 1 38 10 74 15 41 ...
#>  ..  ..$ labels     : chr [1:100] "unmapped_126" "unmapped_127" "unmapped_128" "unmapped_129" ...
#>  ..  ..$ method     : chr "complete"
#>  ..  ..$ call       : language stats::hclust(d = dist_matrix, method = "complete")
#>  ..  ..$ dist.method: NULL
#>  ..  ..- attr(*, "class")= chr "hclust"
#>  @ sample_summary           : num [1:100, 1:5, 1:3] 0 0 0.02 0 0 0 0 0 0.29 0 ...
#>  .. - attr(*, "dimnames")=List of 3
#>  ..  ..$ : chr [1:100] "ind1" "ind2" "ind3" "ind4" ...
#>  ..  ..$ : chr [1:5] "missingness" "missingness_w_exclusions" "tpa_total" "tpa_complete_features" ...
#>  ..  ..$ : chr [1:3] "raw" "qcing" "post_qc"
#>  @ pcs                      : num [1:100, 1:16, 1:3] -11.67 -12.441 0.319 -0.826 -3.309 ...
#>  .. - attr(*, "dimnames")=List of 3
#>  ..  ..$ : chr [1:100] "ind1" "ind2" "ind3" "ind4" ...
#>  ..  ..$ : chr [1:16] "pc1" "pc2" "pc3" "pc4" ...
#>  ..  ..$ : chr [1:3] "raw" "qcing" "post_qc"
#>  @ prob_pcs                 : num [1:100, 1:10, 1:3] -11.523 -12.42 0.44 -0.777 -3.216 ...
#>  .. - attr(*, "dimnames")=List of 3
#>  ..  ..$ : chr [1:100] "ind1" "ind2" "ind3" "ind4" ...
#>  ..  ..$ : chr [1:10] "prob_pc1" "prob_pc2" "prob_pc3" "prob_pc4" ...
#>  ..  ..$ : chr [1:3] "raw" "qcing" "post_qc"
#>  @ var_exp                  : num [1:77, 1, 1:3] 0.1316 0.093 0.0449 0.0397 0.0353 ...
#>  .. - attr(*, "dimnames")=List of 3
#>  ..  ..$ : chr [1:77] "PC1" "PC2" "PC3" "PC4" ...
#>  ..  ..$ : chr "value"
#>  ..  ..$ : chr [1:3] "raw" "qcing" "post_qc"
#>  @ acceleration_factor      : Named num [1:3] 2 2 2
#>  .. - attr(*, "names")= chr [1:3] "raw" "qcing" "post_qc"
#>  @ n_parallel               : Named int [1:3] 7 8 5
#>  .. - attr(*, "names")= chr [1:3] "raw" "qcing" "post_qc"