Metabolites Object

A Metabolites object is a container for matrices of metabolite data, along with associated metadata. It allows for efficient storage and manipulation of data, supporting quality control, transformations, and various analyses. This object facilitates easy access to data layers, sample and feature summaries, outlier treatment, and more.

Usage

Metabolites(
  project_name = "",
  created = Sys.time(),
  filepath = character(0),
  format = "metabolon_v1",
  source = character(0),
  feature_missingness = 0.2,
  sample_missingness = 0.2,
  total_peak_area_sd = 5,
  outlier_udist = 5,
  outlier_treatment = "leave_be",
  winsorize_quantile = 1,
  tree_cut_height = 0.5,
  pc_outlier_sd = 5,
  derived_var_exclusion = TRUE,
  xenobiotics_var_exclusion = TRUE,
  samples = data.table::data.table(),
  features = data.table::data.table(),
  data = integer(0),
  exclusions = list(raw = list(samples = list(extreme_sample_missingness = character(0),
    user_defined_sample_missingness = character(0), user_defined_sample_totalpeakarea =
    character(0), user_defined_sample_pca_outlier = character(0)), features =
    list(extreme_feature_missingness = character(0), user_defined_feature_missingness =
    character(0)))),
  feature_summary = integer(0),
  feature_tree = list(),
  sample_summary = integer(0),
  pcs = integer(0),
  prob_pcs = integer(0),
  var_exp = integer(0),
  acceleration_factor = integer(0),
  n_parallel = integer(0)
)

Arguments

project_name

character, the name of the project associated with the data.

created

POSIXct, the timestamp when the Metabolites object was created.

filepath

character, the file path to the original raw data file.

format

character, the format of the metabolomic data (e.g., "metabolon_v1").

source

named character, the source for this data layer (e.g., 'raw' for 'post_qc').

feature_missingness

numeric, the proportion (0-1) of missingness in features, used as a QC threshold. Default is 0.2.

sample_missingness

numeric, the proportion (0-1) of missingness in samples, used as a QC threshold. Default is 0.2.

total_peak_area_sd

numeric, the number of standard deviations from the mean to use for sample QC based on total abundance. Default is 5. If set to NA, total sum abundance will not be used for QC.

outlier_udist

numeric, the number of interquartile range (IQR) unit distances from the median used to identify outliers for features. Default is 5.

outlier_treatment

character, specifies how to treat outliers for PCA analysis. Options: "leave_be" (no action on outliers), "winsorize" (winsorize to the 100th quantile), "turn_NA" (convert outliers to NA, which will be imputed to the median).

winsorize_quantile

numeric, the quantile of remaining values to use for winsorizing outliers. Default is 1.

tree_cut_height

numeric, the threshold for feature independence in hierarchical clustering. Default is 0.5.

pc_outlier_sd

numeric, the number of standard deviations from the mean used for PCA-based sample QC. Default is 5. Set to NA to exclude PCA-based QC.

derived_var_exclusion

logical, whether to exclude derived variables (from multiple existing variables) in Nightingale data. Default is TRUE.

xenobiotics_var_exclusion

logical, whether to exclude xenobiotics from missingness estimates. Default is TRUE.

samples

data.table, a data table containing sample-related information (not to be set directly).

features

data.table, a data table containing feature-related information (not to be set directly).

data

numeric matrix, the data matrix containing metabolite values (not to be set directly).

exclusions

list, holds exclusion codes for data masking (not to be set directly).

feature_summary

numeric matrix, summary statistics for features (not to be set directly).

feature_tree

named list of tree objects, hierarchical clustering results for features (not to be set directly).

sample_summary

numeric matrix, summary statistics for samples (not to be set directly).

pcs

numeric matrix, principal component scores for the data (not to be set directly).

prob_pcs

numeric matrix, probabilities associated with principal components (not to be set directly).

var_exp

named numeric, variance explained by each principal component (not to be set directly).

acceleration_factor

named numeric, the acceleration factor for parallel analysis (not to be set directly).

n_parallel

named numeric, the number of parallel components in the analysis (not to be set directly).

Value

An object of class Metabolites (S7 class).

Fields

project_name:

Object of class "character", name for this project

created:

Object of class "POSIXct", the timestamp when the Metabolites object was created

Examples

# Example of importing data into a Metabolites object and running QC
file <- system.file("extdata", "metabolon_v1_example.xlsx", package = "metaboprep2")
metabolites <- Metabolites(project_name = "MYPROJECT",
                           format       = "metabolon_v1",
                           filepath     = file)
metabolites <- import_data(metabolites)
metabolites <- metabolite_qc(metabolites, source="raw")
#> 
#> ── Starting Metabolite QC Process ──────────────────────────────────────────────
#> ℹ Validating input parameters...
#> ✔ Input validation complete.
#> ℹ Normalising raw data...
#> ✔ Normalised raw data.
#> ℹ Sample & Feature Summary Statistics for raw data...
#> ✔ Sample & Feature Summary Statistics for raw data...
#> ℹ Copying batch_normalised data to new 'qcing' data layer...
#> ✔ qcing data ready for QC.
#> ℹ Analysing exclusions...
#> ℹ Applying exclusion of derived features...
#> ✔ Derived features exclusions assessment complete. (0 features excluded)
#> ℹ Applying exclusion of xenobiotic features...
#> ✔ Xenobiotic features exclusions assessment complete. (3 features excluded)
#> ℹ Assessing for extreme sample missingness >=80%...
#> ✔ Extreme sample missingness exclusions assessment complete. (0 samples excluded)
#> ℹ Assessing for extreme feature missingness >=80%...
#> ✔ Extreme feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Assessing for sample missingness at specified level of >=20%...
#> ✔ Sample missingness exclusions assessment complete. (2 samples excluded)
#> ℹ Assessing for feature missingness at specified level of >=20%...
#> ✔ Feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Calculating total peak abundance outliers at +/- 5 Sdev...
#> ✔ Total peak abundance outlier exclusions assessment complete. (0 samples excluded)
#> ℹ Running principal component outlier analysis at +/- 5 Sdev...
#> ℹ PCA outlier analysis complete.
#> ℹ Re-running sample summary using current exclusions...
#> ℹ QC sample summary complete.
#> ℹ Re-identify feature independence and PC outliers...
#> ℹ PCA outlier analysis 2 complete.
#> ℹ Creating final QC dataset...
#> ℹ Rerunning Sample & Feature Summary Statistics for QC data...
#> ✔ Sample & Feature Summary Statistics for QC data complete.
#> 
#> ── Metabolite QC Process Completed ─────────────────────────────────────────────
metabolites
#> <metaboprep2::Metabolites>
#>  @ project_name             : chr "MYPROJECT"
#>  @ created                  : POSIXct[1:1], format: "2025-04-01 18:00:25"
#>  @ filepath                 : chr "/home/runner/work/_temp/Library/metaboprep2/extdata/metabolon_v1_example.xlsx"
#>  @ format                   : chr "metabolon_v1"
#>  @ source                   : Named chr [1:2] "batch_normalised" "qcing"
#>  .. - attr(*, "names")= chr [1:2] "qcing" "post_qc"
#>  @ feature_missingness      : num 0.2
#>  @ sample_missingness       : num 0.2
#>  @ total_peak_area_sd       : num 5
#>  @ outlier_udist            : num 5
#>  @ outlier_treatment        : chr "leave_be"
#>  @ winsorize_quantile       : num 1
#>  @ tree_cut_height          : num 0.5
#>  @ pc_outlier_sd            : num 5
#>  @ derived_var_exclusion    : logi TRUE
#>  @ xenobiotics_var_exclusion: logi TRUE
#>  @ samples                  :Classes ‘data.table’ and 'data.frame':	100 obs. of  7 variables:
#>  .. $ sample_id  : chr  "ind1" "ind2" "ind3" "ind4" ...
#>  .. $ sample_name: chr  "ind1" "ind2" "ind3" "ind4" ...
#>  .. $ neg        : chr  "batch1" "batch1" "batch1" "batch1" ...
#>  .. $ pos        : chr  "batch1" "batch1" "batch1" "batch1" ...
#>  .. $ run_day    : chr  "day1" "day1" "day1" "day1" ...
#>  .. $ box_id     : chr  "box1" "box1" "box1" "box1" ...
#>  .. $ lot        : chr  "lot250" "lot250" "lot250" "lot250" ...
#>  .. - attr(*, ".internal.selfref")=<externalptr> 
#>  @ features                 :Classes ‘data.table’ and 'data.frame':	100 obs. of  8 variables:
#>  .. $ feature_id     : chr  "compid_123" "compid_124" "compid_125" "compid_126" ...
#>  .. $ platform       : chr  "neg" "neg" "neg" "neg" ...
#>  .. $ pathway        : chr  "xenobiotic" "xenobiotic" "xenobiotic" "lipid" ...
#>  .. $ derived_feature: logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
#>  .. $ metabolite_id  : chr  "met1" "met2" "met3" "met4" ...
#>  .. $ comp_id        : chr  "123" "124" "125" "126" ...
#>  .. $ kegg           : chr  "K100" "K100" "K100" "K200" ...
#>  .. $ hmdb           : chr  "HMDB1" "HMDB2" "HMDB3" "HMDB4" ...
#>  .. - attr(*, ".internal.selfref")=<externalptr> 
#>  .. - attr(*, "index")= int(0) 
#>  ..  ..- attr(*, "__derived_feature")= int(0) 
#>  @ data                     : num [1:100, 1:100, 1:5] 11.03 12.5 NA 6.98 8.56 ...
#>  .. - attr(*, "dimnames")=List of 3
#>  ..  ..$ : chr [1:100] "ind1" "ind2" "ind3" "ind4" ...
#>  ..  ..$ : chr [1:100] "compid_123" "compid_124" "compid_125" "compid_126" ...
#>  ..  ..$ : chr [1:5] "raw" "scaled" "batch_normalised" "qcing" ...
#>  @ exclusions               :List of 3
#>  .. $ raw    :List of 2
#>  ..  ..$ samples :List of 4
#>  ..  .. ..$ extreme_sample_missingness       : chr(0) 
#>  ..  .. ..$ user_defined_sample_missingness  : chr(0) 
#>  ..  .. ..$ user_defined_sample_totalpeakarea: chr(0) 
#>  ..  .. ..$ user_defined_sample_pca_outlier  : chr(0) 
#>  ..  ..$ features:List of 2
#>  ..  .. ..$ extreme_feature_missingness     : chr(0) 
#>  ..  .. ..$ user_defined_feature_missingness: chr(0) 
#>  .. $ qcing  :List of 2
#>  ..  ..$ samples :List of 4
#>  ..  .. ..$ extreme_sample_missingness       : chr(0) 
#>  ..  .. ..$ user_defined_sample_missingness  : chr [1:2] "ind9" "ind59"
#>  ..  .. ..$ user_defined_sample_totalpeakarea: chr(0) 
#>  ..  .. ..$ user_defined_sample_pca_outlier  : chr "ind1"
#>  ..  ..$ features:List of 2
#>  ..  .. ..$ extreme_feature_missingness     : chr(0) 
#>  ..  .. ..$ user_defined_feature_missingness: chr(0) 
#>  .. $ post_qc:List of 2
#>  ..  ..$ samples :List of 4
#>  ..  .. ..$ extreme_sample_missingness       : chr(0) 
#>  ..  .. ..$ user_defined_sample_missingness  : chr [1:2] "ind9" "ind59"
#>  ..  .. ..$ user_defined_sample_totalpeakarea: chr(0) 
#>  ..  .. ..$ user_defined_sample_pca_outlier  : chr "ind1"
#>  ..  ..$ features:List of 2
#>  ..  .. ..$ extreme_feature_missingness     : chr(0) 
#>  ..  .. ..$ user_defined_feature_missingness: chr(0) 
#>  @ feature_summary          : num [1:21, 1:100, 1:3] 0.2 NA 0 80 0.984 ...
#>  .. - attr(*, "dimnames")=List of 3
#>  ..  ..$ : chr [1:21] "missingness" "missingness_w_exclusions" "outlier_count" "n" ...
#>  ..  ..$ : chr [1:100] "compid_123" "compid_124" "compid_125" "compid_126" ...
#>  ..  ..$ : chr [1:3] "batch_normalised" "qcing" "post_qc"
#>  @ feature_tree             :List of 3
#>  .. $ batch_normalised:List of 7
#>  ..  ..$ merge      : int [1:99, 1:2] -11 -3 -75 -85 -4 -65 -55 -54 -60 -53 ...
#>  ..  ..$ height     : num [1:99] 0.169 0.187 0.232 0.275 0.374 ...
#>  ..  ..$ order      : int [1:100] 75 92 2 34 30 42 5 40 58 81 ...
#>  ..  ..$ labels     : chr [1:100] "compid_123" "compid_124" "compid_125" "compid_126" ...
#>  ..  ..$ method     : chr "complete"
#>  ..  ..$ call       : language stats::hclust(d = dist_matrix, method = "complete")
#>  ..  ..$ dist.method: NULL
#>  ..  ..- attr(*, "class")= chr "hclust"
#>  .. $ qcing           :List of 7
#>  ..  ..$ merge      : int [1:96, 1:2] -8 -72 -82 -1 -52 -62 -51 -50 -57 -88 ...
#>  ..  ..$ height     : num [1:96] 0.169 0.24 0.273 0.368 0.376 ...
#>  ..  ..$ order      : int [1:97] 83 90 16 29 14 33 95 3 17 43 ...
#>  ..  ..$ labels     : chr [1:97] "compid_126" "compid_127" "compid_128" "compid_129" ...
#>  ..  ..$ method     : chr "complete"
#>  ..  ..$ call       : language stats::hclust(d = dist_matrix, method = "complete")
#>  ..  ..$ dist.method: NULL
#>  ..  ..- attr(*, "class")= chr "hclust"
#>  .. $ post_qc         :List of 7
#>  ..  ..$ merge      : int [1:99, 1:2] -8 -3 -72 -82 -1 -52 -62 -51 -29 -50 ...
#>  ..  ..$ height     : num [1:99] 0.166 0.189 0.243 0.282 0.366 ...
#>  ..  ..$ order      : int [1:100] 92 52 68 11 31 99 48 82 97 56 ...
#>  ..  ..$ labels     : chr [1:100] "compid_126" "compid_127" "compid_128" "compid_129" ...
#>  ..  ..$ method     : chr "complete"
#>  ..  ..$ call       : language stats::hclust(d = dist_matrix, method = "complete")
#>  ..  ..$ dist.method: NULL
#>  ..  ..- attr(*, "class")= chr "hclust"
#>  @ sample_summary           : num [1:100, 1:5, 1:3] 0 0 0.02 0 0 0 0 0 0.29 0 ...
#>  .. - attr(*, "dimnames")=List of 3
#>  ..  ..$ : chr [1:100] "ind1" "ind2" "ind3" "ind4" ...
#>  ..  ..$ : chr [1:5] "missingness" "missingness_w_exclusions" "tpa_total" "tpa_complete_features" ...
#>  ..  ..$ : chr [1:3] "batch_normalised" "qcing" "post_qc"
#>  @ pcs                      : num [1:100, 1:16, 1:3] -15.45 -14.14 1.42 -1.09 -2.86 ...
#>  .. - attr(*, "dimnames")=List of 3
#>  ..  ..$ : chr [1:100] "ind1" "ind2" "ind3" "ind4" ...
#>  ..  ..$ : chr [1:16] "pc1" "pc2" "pc3" "pc4" ...
#>  ..  ..$ : chr [1:3] "batch_normalised" "qcing" "post_qc"
#>  @ prob_pcs                 : num [1:100, 1:10, 1:3] -15.31 -14.08 1.73 -1.03 -2.78 ...
#>  .. - attr(*, "dimnames")=List of 3
#>  ..  ..$ : chr [1:100] "ind1" "ind2" "ind3" "ind4" ...
#>  ..  ..$ : chr [1:10] "prob_pc1" "prob_pc2" "prob_pc3" "prob_pc4" ...
#>  ..  ..$ : chr [1:3] "batch_normalised" "qcing" "post_qc"
#>  @ var_exp                  : num [1:87, 1, 1:3] 0.0923 0.0617 0.0486 0.0413 0.0396 ...
#>  .. - attr(*, "dimnames")=List of 3
#>  ..  ..$ : chr [1:87] "PC1" "PC2" "PC3" "PC4" ...
#>  ..  ..$ : chr "value"
#>  ..  ..$ : chr [1:3] "batch_normalised" "qcing" "post_qc"
#>  @ acceleration_factor      : Named num [1:3] 2 2 2
#>  .. - attr(*, "names")= chr [1:3] "batch_normalised" "qcing" "post_qc"
#>  @ n_parallel               : Named int [1:3] 13 12 15
#>  .. - attr(*, "names")= chr [1:3] "batch_normalised" "qcing" "post_qc"