Metabolites object

A Metabolites object is a container for matrices of metabolites data and associated metadata.

Usage

Metabolites(
  project_name = "",
  created = Sys.time(),
  filepath = character(0),
  format = "metabolon_v1",
  samples = (function (.data = list(), row.names = NULL) 
 {
     if (is.null(row.names))
    {
         list2DF(.data)
     }
     else {
         out <- list2DF(.data,
    length(row.names))
attr(out, "row.names") <- row.names
         out
     }

    })(),
  features = (function (.data = list(), row.names = NULL) 
 {
     if
    (is.null(row.names)) {
         list2DF(.data)
     }
     else {
         out <-
    list2DF(.data, length(row.names))
attr(out, "row.names") <- row.names
      
      out
     }
 })(),
  data = integer(0),
  feature_missingness = 0.2,
  sample_missingness = 0.2,
  total_peak_area_sd = 5,
  outlier_udist = 5,
  outlier_treatment = "leave_be",
  tree_cut_height = 0.5,
  pc_outlier_sd = 5,
  derived_var_exclusion = TRUE,
  xenobiotics_var_exclusion = TRUE
)

Arguments

project_name

character, name for this project

filepath

character, path to the original raw data file

format

character, a valid metabolomic data format

...

optional other slots to fill, see below.

Value

an S7 class metaboprep2::Metabolites object

Slots

project_name

character, name for this project

filepath

character, path to the original raw data file

source

character, a valid metabolomic data source

samples

data.table,

features

data.table,

data

multi-dimentional matrix,

feature_missingness

numeric 0-1 default 0.2, proportion of missinginess on features used as a QC threshold

sample_missingness

numeric 0-1 default 0.2, proportion of missinginess on samples used as a QC threshold

total_peak_area_sd

numeric default 5, number of standard deviation (SD) units from the mean to perform a sample QC based on total abundance. If you would like to exclude total sum abundance as a QC parameter set the value to NA

outlier_udist

numeric default 5, outlier values for features. The number of interquartile range (IQR) unit distances from the median to call a value an outlier. "Outliers" here are extreme values that are the product of error.

outlier_treatment

character default 'leave_be', what to do with outliers for the purposes of the PCA, and for the PCA only. Options: "leave_be" if you would like no action on outliers; "winsorize" if you would like outliers to be winsorized to the 100th quantile of all remaining (non-outlying) values, at a feature; "turn_NA" if you would like outliers converted to NA. This means they will be imputed to the median for the purposes of the PCA.

tree_cut_height

numeric default 0.5, feature|metabolite independence. To identify "independent" features in your data set, how similar should clustered|grouped features be? tree_cut_height = 1-absolute(spearman's rho), such that a tree_cut_height of 0.2 would indicate a intra-cluster similarity of >0.8, and a tree_cut_height of 0.8 would indicate a intra-cluster similarity of >0.2. Larger tree_cut_height values yield more clustering and thus fewer "representative" or "independent" features.

pc_outlier_sd

numeric default 5, the number of standard deviation (SD) units from the mean to perform a sample QC based on principal components. If you would like to exclude PC exclusions as a QC parameter set the value to NA. Setting to NA may be advisable if you expect significant structure among individuals for example your data is derived from different tissues, different geographies/ecologies/environments.

derived_var_exclusion

logical default TRUE, Nightingale derived variable exclusion. Derived variables in Nightingale are all variables derived from two or more variables already present in the data set. In this instance it represents all ratios that Nightingale supply in their data releases.

xenobiotics_var_exclusion

logical default TRUE, whether to exclude Xenobiotics from the missingness estimate

feature_tree

object

summary_data

list