Annotate Features with Common IDs

This function annotates a data.table by mapping feature identifiers (id_col) to standardized annotations. Optionally, pathway and platform information can be included.

Usage

annotate_features(
  features,
  fixed_match_cols = list(pubchem_id = NULL, hmdb_id = NULL, kegg_id = NULL, chebi_id =
    NULL, metlin_id = NULL, name = NULL, smiles = NULL, inchi_key = NULL, chem_id = NULL,
    comp_id = NULL),
  fuzzy_match_cols = list(name = NULL),
  fuzzy_method = "jw",
  fuzzy_dist = 0.1
)

Arguments

features

data.table The dataset containing features to be annotated.

fixed_match_cols

list The name of the column containing feature IDs - can be a vector of column names if multiple columns are to be tried. This can be a column with identifiers such as comp_ids, chem_ids, or identifiers from Nightingale Excel or flat text files. If multiple columns are given then the resulting match precedence will be in order of the columns given.

fuzzy_match_cols

list, The name of the column containing pathway information. If provided, source pathway data from features will be included in the annotation with standardised column name pathway.

fuzzy_method

method for fuzzy matching see fuzzyjoin::stringdist_inner_join

fuzzy_dist

distance for fuzzy matching see fuzzyjoin::stringdist_inner_join

Value

data.table A modified version of features with annotations added based on the provided IDs.

Examples

# Example dataset
features <- data.table::data.table(comp_ids = c("63042", "62919", "62921"))

# Annotate features using 'comp_ids'
annotated_features <- annotate_features(features, fixed_match_cols = list(comp_id = 'comp_ids'))
#> Using cached MetaboAnalyst DB.
#> Attempting join [features$comp_ids] ~ [annotation$comp_id] 
annotated_features
#>    comp_ids           join_key anno_hmdb_id                   anno_name
#>      <char>             <char>       <char>                      <char>
#> 1:    63042 comp_ids ~ comp_id         <NA>   2,6-dihydroxybenzoic acid
#> 2:    62919 comp_ids ~ comp_id         <NA> tridecenedioate (C13:1-DC)*
#> 3:    62921 comp_ids ~ comp_id         <NA>    pregnenetriol disulfate*
#>    anno_pubchem_id anno_chebi_id anno_kegg_id anno_metlin_id anno_lipid
#>             <char>        <char>       <char>         <char>     <char>
#> 1:            <NA>          <NA>                        <NA>       <NA>
#> 2:            <NA>          <NA>                        <NA>       <NA>
#> 3:            <NA>          <NA>                        <NA>       <NA>
#>    anno_smiles anno_exactmass anno_formula anno_inchi_key anno_super_class
#>         <char>         <char>       <char>         <char>           <char>
#> 1:                       <NA>         <NA>                            <NA>
#> 2:                       <NA>         <NA>                            <NA>
#> 3:                       <NA>         <NA>                            <NA>
#>    anno_main_class anno_sub_class                anno_synonym anno_chem_id
#>             <char>         <char>                      <char>       <char>
#> 1:                                  2,6-dihydroxybenzoic acid    100020837
#> 2:                                tridecenedioate (C13:1-DC)*    100002199
#> 3:                                   pregnenetriol disulfate*    100021104
#>    anno_comp_id anno_derived_feature          anno_split_synonym
#>          <char>               <lgcl>                      <char>
#> 1:        63042                FALSE   2,6-dihydroxybenzoic acid
#> 2:        62919                FALSE tridecenedioate (C13:1-DC)*
#> 3:        62921                FALSE    pregnenetriol disulfate*
#>    anno_fuzzy_dist
#>              <num>
#> 1:              NA
#> 2:              NA
#> 3:              NA