Export Data from a Metabolites Object

Exports the data from a Metabolites object to a specified directory. The function allows the user to export the data in a format that is compatible with data.table::fwrite for efficient writing of large datasets.

Usage

export_data(metabolites, directory)

Arguments

metabolites

A Metabolites object containing the data to be exported.

directory

A character string specifying the path to the directory where the data should be written.

Value

This function does not return a value. It writes the data to the specified directory as a file (e.g., CSV).

Examples

file <- system.file("extdata", "metabolon_v1_example.xlsx", package = "metaboprep2")
metabolites <- Metabolites(project_name = "MYPROJECT",
                           format       = "metabolon_v1",
                           filepath     = file)
metabolites <- import_data(metabolites)
if (FALSE) { # \dontrun{
# Specify the directory to export the data
export_data(metabolites, directory = "path/to/directory")
} # }