Summary Statistics for Features

This function allows you to 'describe' metabolite features using the describe() function from the psych package, as well as estimate variance, a dispersion index, the coeficent of variation, and shapiro's W-statistic.

Usage

feature_describe(data)

Arguments

data

matrix, the metabolite data matrix. Samples in row, metabolites in columns

Value

a data frame of summary statistics for features (columns) of a matrix

Examples

ex_data = sapply(1:20, function(x){ rnorm(250, 40, 5) })
feature_describe(ex_data)
#>       n     mean       sd   median      min      max    range         skew
#> X1  250 40.27526 5.172278 40.36540 26.93833 53.77709 26.83876 -0.056458353
#> X2  250 40.06002 5.154675 39.54348 25.30511 50.79378 25.48867  0.028685791
#> X3  250 39.51806 4.852301 39.37071 25.95995 53.27449 27.31454  0.038156775
#> X4  250 40.23345 4.802881 40.45265 28.18896 53.46186 25.27291 -0.033300476
#> X5  250 39.36757 5.092792 39.08372 26.19066 55.54591 29.35525  0.243294820
#> X6  250 40.03427 4.673586 39.93544 25.25458 52.92894 27.67436 -0.112163593
#> X7  250 40.39760 4.759053 40.17601 28.14909 54.19633 26.04724  0.082062001
#> X8  250 39.87530 4.784242 39.72621 27.88544 52.62554 24.74010  0.154642782
#> X9  250 39.90066 4.980441 39.76576 25.89100 55.04046 29.14946  0.002828426
#> X10 250 40.22230 4.982461 40.52634 27.63224 55.48615 27.85391  0.039310864
#> X11 250 40.24081 4.908312 40.19305 27.36124 53.08089 25.71966  0.023109163
#> X12 250 40.40534 5.138749 40.34793 28.49693 55.35504 26.85811  0.026096642
#> X13 250 39.94624 4.897504 39.73896 27.90090 52.78462 24.88372  0.014534053
#> X14 250 39.56150 4.765366 39.48203 26.86979 57.93473 31.06494  0.281380210
#> X15 250 39.89009 4.666954 39.80346 25.47890 52.23391 26.75501 -0.164008017
#> X16 250 39.77211 5.349866 39.89135 23.47909 53.93702 30.45793 -0.179582491
#> X17 250 40.70237 5.299425 40.86785 24.97725 57.40270 32.42545 -0.145757676
#> X18 250 39.69223 4.699183 40.13247 27.35291 52.51339 25.16049  0.100622736
#> X19 250 40.61697 5.125332 40.83701 26.94429 53.09422 26.14993 -0.153575831
#> X20 250 40.14265 4.880132 40.13518 26.14487 53.02527 26.88040 -0.111417552
#>        kurtosis        se missing      var disp_index coef_variance         W
#> X1  -0.15215254 0.3271236       0 26.75246  0.6642406     0.1284232 0.9943430
#> X2  -0.50914152 0.3260103       0 26.57067  0.6632716     0.1286738 0.9907720
#> X3   0.15620612 0.3068864       0 23.54482  0.5957990     0.1227869 0.9949403
#> X4  -0.33819027 0.3037609       0 23.06767  0.5733454     0.1193753 0.9956432
#> X5   0.03306867 0.3220964       0 25.93653  0.6588298     0.1293652 0.9937078
#> X6  -0.27665780 0.2955835       0 21.84240  0.5455927     0.1167396 0.9947771
#> X7  -0.24083805 0.3009889       0 22.64859  0.5606419     0.1178053 0.9955616
#> X8  -0.30428688 0.3025820       0 22.88897  0.5740137     0.1199801 0.9948012
#> X9  -0.01714174 0.3149907       0 24.80479  0.6216636     0.1248210 0.9977853
#> X10 -0.25398512 0.3151185       0 24.82492  0.6171929     0.1238731 0.9970326
#> X11 -0.37335697 0.3104289       0 24.09152  0.5986839     0.1219735 0.9965254
#> X12 -0.18807801 0.3250030       0 26.40674  0.6535457     0.1271799 0.9937110
#> X13 -0.28242050 0.3097453       0 23.98554  0.6004455     0.1226024 0.9956216
#> X14  0.83320414 0.3013882       0 22.70871  0.5740105     0.1204547 0.9896402
#> X15  0.46694407 0.2951641       0 21.78046  0.5460117     0.1169953 0.9914818
#> X16 -0.13303653 0.3383553       0 28.62107  0.7196267     0.1345130 0.9939993
#> X17  0.30583435 0.3351651       0 28.08390  0.6899820     0.1301994 0.9933469
#> X18 -0.27985525 0.2972024       0 22.08232  0.5563387     0.1183905 0.9948965
#> X19 -0.34989357 0.3241545       0 26.26903  0.6467501     0.1261870 0.9939497
#> X20 -0.19222055 0.3086467       0 23.81569  0.5932766     0.1215698 0.9953033
#>       log10_W
#> X1  0.9842886
#> X2  0.9874852
#> X3  0.9870036
#> X4  0.9886926
#> X5  0.9960951
#> X6  0.9847454
#> X7  0.9927975
#> X8  0.9952146
#> X9  0.9896302
#> X10 0.9920821
#> X11 0.9917882
#> X12 0.9869008
#> X13 0.9895151
#> X14 0.9917280
#> X15 0.9731672
#> X16 0.9769094
#> X17 0.9754657
#> X18 0.9931757
#> X19 0.9813313
#> X20 0.9836158