Summary Statistics for Features

This function allows you to 'describe' metabolite features using the describe() function from the psych package, as well as estimate variance, a dispersion index, the coeficent of variation, and shapiro's W-statistic.

Usage

feature_describe(data)

Arguments

data

matrix, the metabolite data matrix. Samples in row, metabolites in columns

Value

a data frame of summary statistics for features (columns) of a matrix

Examples

ex_data = sapply(1:20, function(x){ rnorm(250, 40, 5) })
feature_describe(ex_data)
#>       n     mean       sd   median      min      max    range         skew
#> X1  250 40.02612 5.485593 40.29187 23.78556 53.09653 29.31097 -0.183357484
#> X2  250 39.75038 5.110470 39.13285 26.25877 56.98186 30.72309  0.185380112
#> X3  250 39.70952 4.922882 39.89575 26.09766 52.71135 26.61368  0.099228500
#> X4  250 40.24168 4.580230 40.08521 25.94700 52.45746 26.51045 -0.153332786
#> X5  250 40.48538 5.128406 40.36834 22.56377 52.99907 30.43529 -0.076047986
#> X6  250 40.39414 4.796911 40.49903 25.80435 60.31731 34.51296  0.117461819
#> X7  250 39.79121 5.048694 39.48523 24.24222 55.02006 30.77785  0.014078652
#> X8  250 39.82109 5.033146 39.83822 22.55833 51.41472 28.85639 -0.085018252
#> X9  250 40.09075 4.672263 40.28910 26.14353 52.36189 26.21837 -0.214095221
#> X10 250 40.04637 5.174986 40.34156 25.64219 54.36973 28.72753  0.063619390
#> X11 250 39.89832 5.278189 39.73322 24.36818 53.62759 29.25941  0.039110619
#> X12 250 40.04502 4.855123 39.89259 25.54773 54.04576 28.49804  0.025110293
#> X13 250 39.98317 5.098151 39.47268 27.81008 53.47396 25.66388  0.195494603
#> X14 250 39.64356 5.244423 39.61339 21.85514 54.30136 32.44622  0.009347551
#> X15 250 39.74456 5.005813 40.22353 27.14815 52.81507 25.66691 -0.085402056
#> X16 250 40.04645 5.198383 39.97718 26.82679 58.27249 31.44570  0.114756448
#> X17 250 40.05496 5.149754 39.86412 26.21384 54.16155 27.94772  0.024032228
#> X18 250 39.97480 4.898688 39.73195 23.97944 52.31933 28.33989 -0.021513239
#> X19 250 40.46000 4.951577 40.37423 30.06999 53.69132 23.62133  0.093241331
#> X20 250 40.21402 4.768120 40.22599 28.28947 57.66385 29.37438  0.233148114
#>         kurtosis        se missing      var disp_index coef_variance         W
#> X1  -0.415204628 0.3469394       0 30.09173  0.7518023     0.1370503 0.9926516
#> X2   0.175453485 0.3232145       0 26.11690  0.6570228     0.1285641 0.9930559
#> X3  -0.418013052 0.3113504       0 24.23477  0.6103011     0.1239723 0.9937620
#> X4   0.046323885 0.2896792       0 20.97851  0.5213130     0.1138181 0.9953662
#> X5   0.112501515 0.3243489       0 26.30055  0.6496307     0.1266730 0.9923416
#> X6   0.962236572 0.3033833       0 23.01035  0.5696458     0.1187526 0.9903251
#> X7   0.008016932 0.3193075       0 25.48931  0.6405764     0.1268796 0.9976101
#> X8  -0.167485557 0.3183241       0 25.33255  0.6361593     0.1263940 0.9953232
#> X9   0.038224263 0.2954998       0 21.83004  0.5445156     0.1165422 0.9944220
#> X10  0.281961204 0.3272948       0 26.78048  0.6687367     0.1292248 0.9939170
#> X11 -0.004586368 0.3338220       0 27.85928  0.6982568     0.1322910 0.9950954
#> X12  0.052857682 0.3070649       0 23.57222  0.5886430     0.1212416 0.9966118
#> X13 -0.290897589 0.3224354       0 25.99114  0.6500522     0.1275074 0.9917636
#> X14  0.189067463 0.3316864       0 27.50398  0.6937817     0.1322894 0.9960280
#> X15 -0.591724925 0.3165954       0 25.05817  0.6304805     0.1259497 0.9905689
#> X16 -0.054976502 0.3287746       0 27.02319  0.6747962     0.1298089 0.9955203
#> X17 -0.042518151 0.3256990       0 26.51997  0.6620895     0.1285672 0.9950354
#> X18 -0.056905663 0.3098202       0 23.99714  0.6003067     0.1225444 0.9943880
#> X19 -0.569087691 0.3131652       0 24.51811  0.6059841     0.1223820 0.9913250
#> X20  0.138129457 0.3015624       0 22.73497  0.5653493     0.1185686 0.9938098
#>       log10_W
#> X1  0.9779148
#> X2  0.9915171
#> X3  0.9920987
#> X4  0.9819215
#> X5  0.9775257
#> X6  0.9851844
#> X7  0.9894249
#> X8  0.9830045
#> X9  0.9780595
#> X10 0.9858927
#> X11 0.9864587
#> X12 0.9896904
#> X13 0.9931952
#> X14 0.9838967
#> X15 0.9822725
#> X16 0.9928386
#> X17 0.9873941
#> X18 0.9841642
#> X19 0.9894722
#> X20 0.9956899