Retrieve Feature Summary Data

Extracts feature-level summary statistics from a Metabolites object for a specified data layer, with optional exclusions and subsetting.

Usage

get_feature_summary(
  metabolites,
  layer,
  feature_ids = NULL,
  drop = FALSE,
  as_df = FALSE
)

Arguments

metabolites

A Metabolites object containing feature summary data.

layer

A character string specifying the data layer from which to extract feature summary data.

feature_ids

Optional. A vector of row names (features) to extract. Defaults to NULL, meaning all features are included.

drop

Logical. Whether to drop all NA (excluded features)

as_df

Logical. Whether to return the result as a data.frame. Defaults to FALSE, meaning the result is returned as a matrix.

Value

A matrix or data.frame (if as_df = TRUE) containing the feature summary data for the specified layer.

Examples

file <- system.file("extdata", "metabolon_v1_example.xlsx", package = "metaboprep2")
metabolites <- Metabolites(project_name = "MYPROJECT",
                           format       = "metabolon_v1",
                           filepath     = file)
metabolites <- import_data(metabolites)
#> Warning: 86 unmapped features:
#>  - 123, 124, 125, 126, 127, 128, 129, 130, 131, 132, 133, 134, 135, 137, 138, 139, 140, 141, 142, 143, 145, 146, 147, 148, 149, 150, 151, 152, 153, 154, 155, 156, 157, 159, 160, 161, 162, 163, 164, 165, 166, 167, 168, 169, 170, 172, 173, 174, 175, 176, 177, 178, 183, 184, 185, 186, 187, 188, 190, 191, 193, 195, 196, 198, 199, 200, 201, 202, 204, 205, 206, 207, 208, 209, 210, 211, 213, 214, 215, 216, 217, 218, 219, 220, 221, 222
metabolites <- metabolite_qc(metabolites, source="raw")
#> 
#> ── Starting Metabolite QC Process ──────────────────────────────────────────────
#> ℹ Validating input parameters...
#> ✔ Input validation complete.
#> ℹ Applying exclusion of xenobiotic features...
#> ✔ Xenobiotic features exclusions assessment complete. (3 features excluded)
#> ℹ Analysing exclusions...
#> ℹ Applying exclusion of derived features...
#> ✔ Derived features exclusions assessment complete. (0 features excluded)
#> ℹ Sample & Feature Summary Statistics for raw data...
#> ✔ Sample & Feature Summary Statistics for raw data...
#> ℹ Copying raw data to new 'qcing' data layer...
#> ✔ qcing data ready for QC.
#> ℹ Assessing for extreme sample missingness >=80%...
#> ✔ Extreme sample missingness exclusions assessment complete. (0 samples excluded)
#> ℹ Assessing for extreme feature missingness >=80%...
#> ✔ Extreme feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Assessing for sample missingness at specified level of >=20%...
#> ✔ Sample missingness exclusions assessment complete. (2 samples excluded)
#> ℹ Assessing for feature missingness at specified level of >=20%...
#> ✔ Feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Calculating total peak abundance outliers at +/- 5 Sdev...
#> ✔ Total peak abundance outlier exclusions assessment complete. (0 samples excluded)
#> ℹ Running data outlier analysis at +/- 5 Sdev...
#> ℹ Data outlier analysis complete.
#> ℹ Re-identify feature independence and PC outliers...
#> ℹ Re-running sample summary using current exclusions...
#> ℹ QC sample summary complete.
#> ℹ PCA outlier analysis 2 complete.
#> ℹ Creating final QC dataset...
#> ℹ Rerunning Sample & Feature Summary Statistics for QC data...
#> ✔ Sample & Feature Summary Statistics for QC data complete.
#> 
#> ── Metabolite QC Process Completed ─────────────────────────────────────────────
get_feature_summary(metabolites, layer = "post_qc")
#>                             unmapped_123 unmapped_124 unmapped_125 unmapped_126
#> missingness                    0.1836735   0.11224490    0.1530612    0.0000000
#> missingness_w_exclusions              NA           NA           NA           NA
#> outlier_count                  0.0000000   0.00000000    0.0000000    0.0000000
#> n                             80.0000000  87.00000000   83.0000000   98.0000000
#> mean                          10.5077287  10.33218592   10.7465703   10.6445208
#> sd                             2.2538597   2.14547292    2.1011821    2.1189599
#> median                        10.7264724  10.47049066   11.1452368   10.4034864
#> min                            3.8287814   5.29085576    4.7641198    6.5683390
#> max                           14.8689519  15.74854493   15.9922763   17.1487104
#> range                         11.0401704  10.45768917   11.2281564   10.5803715
#> skew                          -0.4756323  -0.03672953   -0.2227355    0.2201510
#> kurtosis                       0.1091215  -0.51013300   -0.2263649   -0.2781731
#> se                             0.2519892   0.23001884    0.2306347    0.2140473
#> missing                       18.0000000  11.00000000   15.0000000    0.0000000
#> var                            5.0798837   4.60305406    4.4149661    4.4899912
#> disp_index                     0.4834426   0.44550631    0.4108256    0.4218124
#> coef_variance                  0.2144954   0.20764947    0.1955212    0.1990658
#> W                              0.9786328   0.99037370    0.9859793    0.9863157
#> log10_W                        0.9079980   0.97054379    0.9506945    0.9839707
#> k                              1.0000000   2.00000000    3.0000000    4.0000000
#> independent_features_binary    1.0000000   1.00000000    1.0000000    1.0000000
#>                             unmapped_127 unmapped_128 unmapped_129 unmapped_130
#> missingness                   0.00000000   0.00000000    0.0000000    0.0000000
#> missingness_w_exclusions              NA           NA           NA           NA
#> outlier_count                 0.00000000   0.00000000    0.0000000    0.0000000
#> n                            98.00000000  98.00000000   98.0000000   98.0000000
#> mean                         10.64216437  10.70203801   10.8860833   10.7956052
#> sd                            2.19035693   2.08861591    1.9345899    2.1991246
#> median                       10.52837286  10.55455243   10.8198548   10.7295236
#> min                           4.34325228   4.69088342    6.2658342    5.9792798
#> max                          15.54996513  14.96578311   16.8847603   16.3042954
#> range                        11.20671285  10.27489969   10.6189260   10.3250156
#> skew                         -0.06705904  -0.23854730    0.4570077    0.2534129
#> kurtosis                      0.13377778  -0.08435797    0.6052234   -0.5612584
#> se                            0.22125946   0.21098207    0.1954231    0.2221451
#> missing                       0.00000000   0.00000000    0.0000000    0.0000000
#> var                           4.79766347   4.36231642    3.7426379    4.8361489
#> disp_index                    0.45081652   0.40761549    0.3438002    0.4479739
#> coef_variance                 0.20581875   0.19516058    0.1777122    0.2037055
#> W                             0.98674321   0.98686448    0.9749241    0.9871983
#> log10_W                       0.94942580   0.94564578    0.9834747    0.9905338
#> k                             5.00000000   3.00000000    6.0000000    7.0000000
#> independent_features_binary   1.00000000   0.00000000    1.0000000    1.0000000
#>                             unmapped_131 unmapped_132 unmapped_133 unmapped_134
#> missingness                   0.00000000    0.0000000   0.00000000    0.0000000
#> missingness_w_exclusions              NA           NA           NA           NA
#> outlier_count                 0.00000000    0.0000000   0.00000000    0.0000000
#> n                            98.00000000   98.0000000  98.00000000   98.0000000
#> mean                         10.50792555   10.3889029  10.75326893   10.7297003
#> sd                            2.06408544    2.2457716   2.03084218    2.3155607
#> median                       10.49239313   10.2085580  10.84400258   10.5967850
#> min                           6.53139311    3.4548927   5.92832517    4.8675912
#> max                          15.05211296   16.7942017  15.40040839   16.2539313
#> range                         8.52071985   13.3393090   9.47208322   11.3863400
#> skew                          0.07404147    0.1143258   0.08249936    0.1999519
#> kurtosis                     -0.77100525    0.7445899  -0.35336946   -0.3289992
#> se                            0.20850412    0.2268572   0.20514604    0.2339070
#> missing                       0.00000000    0.0000000   0.00000000    0.0000000
#> var                           4.26044870    5.0434901   4.12431997    5.3618212
#> disp_index                    0.40545098    0.4854690   0.38354104    0.4997177
#> coef_variance                 0.19643130    0.2161702   0.18885812    0.2158085
#> W                             0.98426352    0.9854525   0.99031590    0.9916672
#> log10_W                       0.97804526    0.9362547   0.98163859    0.9852082
#> k                             8.00000000    9.0000000  10.00000000   11.0000000
#> independent_features_binary   1.00000000    1.0000000   1.00000000    1.0000000
#>                             unmapped_135 chem_id_136 unmapped_137 unmapped_138
#> missingness                   0.00000000   0.0000000    0.0000000   0.00000000
#> missingness_w_exclusions              NA          NA           NA           NA
#> outlier_count                 0.00000000   0.0000000    0.0000000   0.00000000
#> n                            98.00000000  98.0000000   98.0000000  98.00000000
#> mean                         10.56429254  10.6183766   10.7304221  10.71511415
#> sd                            1.96629085   1.8969665    2.0336149   1.91795624
#> median                       10.60485045  10.6224801   10.3994157  10.70097583
#> min                           5.24359421   4.2405620    6.2191162   6.96935968
#> max                          15.72849318  14.3772791   16.0132654  14.75567126
#> range                        10.48489897  10.1367171    9.7941492   7.78631158
#> skew                          0.04872579  -0.2809391    0.2187702   0.09534678
#> kurtosis                      0.06552789   0.3601381   -0.3582697  -0.59744767
#> se                            0.19862537   0.1916226    0.2054261   0.19374284
#> missing                       0.00000000   0.0000000    0.0000000   0.00000000
#> var                           3.86629972   3.5984820    4.1355897   3.67855613
#> disp_index                    0.36597810   0.3388919    0.3854079   0.34330536
#> coef_variance                 0.18612613   0.1786494    0.1895186   0.17899541
#> W                             0.99055403   0.9853113    0.9880107   0.98484421
#> log10_W                       0.96726313   0.9323444    0.9868603   0.97995682
#> k                            12.00000000  13.0000000   14.0000000  15.00000000
#> independent_features_binary   1.00000000   1.0000000    1.0000000   1.00000000
#>                             unmapped_139 unmapped_140 unmapped_141 unmapped_142
#> missingness                    0.0000000    0.0000000   0.00000000    0.0000000
#> missingness_w_exclusions              NA           NA           NA           NA
#> outlier_count                  0.0000000    0.0000000   0.00000000    0.0000000
#> n                             98.0000000   98.0000000  98.00000000   98.0000000
#> mean                          10.6111153   10.9943017  10.81126507   10.5499313
#> sd                             2.1684357    2.1158927   1.89078354    2.0060625
#> median                        10.5608178   10.9969822  10.88128331   10.4943965
#> min                            4.4665809    5.3632425   6.79673039    5.5574180
#> max                           16.2120501   15.7680079  14.90656733   15.0168278
#> range                         11.7454692   10.4047654   8.10983694    9.4594098
#> skew                          -0.2844336   -0.1061747  -0.05196721   -0.1438616
#> kurtosis                       0.2855917   -0.1940355  -0.54837825   -0.4111697
#> se                             0.2190451    0.2137374   0.19099798    0.2026429
#> missing                        0.0000000    0.0000000   0.00000000    0.0000000
#> var                            4.7021133    4.4770019   3.57506240    4.0242868
#> disp_index                     0.4431309    0.4072111   0.33067938    0.3814515
#> coef_variance                  0.2043551    0.1924536   0.17489013    0.1901493
#> W                              0.9883286    0.9934861   0.98948985    0.9917156
#> log10_W                        0.9336281    0.9656854   0.97522825    0.9678830
#> k                             16.0000000   17.0000000  18.00000000   19.0000000
#> independent_features_binary    1.0000000    1.0000000   1.00000000    1.0000000
#>                             unmapped_143 chem_id_144 unmapped_145 unmapped_146
#> missingness                    0.0000000   0.0000000  0.000000000  0.000000000
#> missingness_w_exclusions              NA          NA           NA           NA
#> outlier_count                  0.0000000   0.0000000  0.000000000  0.000000000
#> n                             98.0000000  98.0000000 98.000000000 98.000000000
#> mean                          10.6574681  10.2865441 10.516111317 10.222098951
#> sd                             2.3368523   2.1276229  2.215633978  2.018082021
#> median                        10.7120792  10.1762878 10.554019810 10.144485780
#> min                            5.0217014   6.1561222  4.477736055  5.037569933
#> max                           15.8154256  16.0392988 16.096074600 15.244401990
#> range                         10.7937242   9.8831766 11.618338545 10.206832057
#> skew                          -0.1173240   0.1780056  0.008411218 -0.007082688
#> kurtosis                      -0.2800203  -0.3637923 -0.284715249 -0.152847679
#> se                             0.2360577   0.2149224  0.223812830  0.203857069
#> missing                        0.0000000   0.0000000  0.000000000  0.000000000
#> var                            5.4608788   4.5267792  4.909033925  4.072655043
#> disp_index                     0.5123993   0.4400680  0.466810761  0.398416711
#> coef_variance                  0.2192690   0.2068355  0.210689476  0.197423448
#> W                              0.9910907   0.9864291  0.995662759  0.994555296
#> log10_W                        0.9570277   0.9813342  0.972195082  0.972794656
#> k                             20.0000000  21.0000000 22.000000000 23.000000000
#> independent_features_binary    1.0000000   1.0000000  1.000000000  1.000000000
#>                             unmapped_147 unmapped_148 unmapped_149 unmapped_150
#> missingness                   0.00000000    0.0000000    0.0000000   0.00000000
#> missingness_w_exclusions              NA           NA           NA           NA
#> outlier_count                 0.00000000    0.0000000    0.0000000   0.00000000
#> n                            98.00000000   98.0000000   98.0000000  98.00000000
#> mean                         10.59668374   10.3370743   10.6264755  10.59575261
#> sd                            1.83476117    2.2310873    2.0487604   2.14499764
#> median                       10.42567942   10.7523293   10.4905206  10.49338506
#> min                           6.10170123    4.4105415    5.4461628   5.27566433
#> max                          14.76676904   15.7200911   17.7333444  15.28460200
#> range                         8.66506781   11.3095496   12.2871817  10.00893767
#> skew                          0.08475963   -0.2352069    0.2605932  -0.05466947
#> kurtosis                     -0.30377102   -0.3219217    0.7683451  -0.38435213
#> se                            0.18533887    0.2253738    0.2069561   0.21667748
#> missing                       0.00000000    0.0000000    0.0000000   0.00000000
#> var                           3.36634854    4.9777504    4.1974192   4.60101485
#> disp_index                    0.31767944    0.4815434    0.3949964   0.43423200
#> coef_variance                 0.17314485    0.2158335    0.1927977   0.20243938
#> W                             0.98946015    0.9879279    0.9837205   0.99375678
#> log10_W                       0.98031688    0.9486279    0.9757500   0.97241385
#> k                            24.00000000   25.0000000   26.0000000  27.00000000
#> independent_features_binary   1.00000000    1.0000000    1.0000000   1.00000000
#>                             unmapped_151 unmapped_152 unmapped_153 unmapped_154
#> missingness                   0.00000000    0.0000000    0.0000000    0.0000000
#> missingness_w_exclusions              NA           NA           NA           NA
#> outlier_count                 0.00000000    0.0000000    0.0000000    0.0000000
#> n                            98.00000000   98.0000000   98.0000000   98.0000000
#> mean                         10.47395734   10.6489251   10.3223429   10.3533419
#> sd                            2.16485901    1.8847842    2.1594421    2.0662076
#> median                       10.40246886   10.7267478   10.0890734   10.4393002
#> min                           5.13347402    6.3640055    6.4240422    3.9972996
#> max                          15.88856933   15.4535081   16.9522004   15.5786971
#> range                        10.75509531    9.0895026   10.5281583   11.5813975
#> skew                         -0.03081478    0.1957928    0.3383719   -0.2927062
#> kurtosis                     -0.40886687   -0.3691141   -0.2927418    0.3062670
#> se                            0.21868378    0.1903920    0.2181366    0.2087185
#> missing                       0.00000000    0.0000000    0.0000000    0.0000000
#> var                           4.68661452    3.5524115    4.6631901    4.2692137
#> disp_index                    0.44745404    0.3335934    0.4517569    0.4123513
#> coef_variance                 0.20668969    0.1769929    0.2092008    0.1995691
#> W                             0.99643042    0.9904972    0.9788304    0.9892743
#> log10_W                       0.97629453    0.9894935    0.9844859    0.9332983
#> k                            28.00000000   29.0000000   30.0000000    1.0000000
#> independent_features_binary   1.00000000    1.0000000    1.0000000    0.0000000
#>                             unmapped_155 unmapped_156 unmapped_157 chem_id_158
#> missingness                  0.000000000    0.0000000   0.00000000   0.0000000
#> missingness_w_exclusions              NA           NA           NA          NA
#> outlier_count                0.000000000    0.0000000   0.00000000   0.0000000
#> n                           98.000000000   98.0000000  98.00000000  98.0000000
#> mean                        10.581203849   10.5496892  10.53205533  10.4760992
#> sd                           2.346372937    2.1268286   1.92607928   2.1422102
#> median                      10.818150155   10.5566898  10.62518509  10.4352561
#> min                          3.043604577    4.8685353   5.99490222   5.0288968
#> max                         17.265832600   17.0435726  15.10241457  16.4431173
#> range                       14.222228023   12.1750373   9.10751235  11.4142205
#> skew                        -0.005263263    0.1634667   0.09426847   0.1395251
#> kurtosis                     0.631729952    0.4860732  -0.27568701  -0.1711254
#> se                           0.237019459    0.2148421   0.19456339   0.2163959
#> missing                      0.000000000    0.0000000   0.00000000   0.0000000
#> var                          5.505465958    4.5234000   3.70978141   4.5890647
#> disp_index                   0.520306199    0.4287709   0.35223717   0.4380509
#> coef_variance                0.221749148    0.2016011   0.18287782   0.2044855
#> W                            0.987892465    0.9862320   0.99192984   0.9944659
#> log10_W                      0.919860958    0.9612469   0.98305129   0.9832274
#> k                           31.000000000   32.0000000  33.00000000  16.0000000
#> independent_features_binary  1.000000000    1.0000000   1.00000000   0.0000000
#>                             unmapped_159 unmapped_160 unmapped_161 unmapped_162
#> missingness                    0.0000000    0.0000000    0.0000000    0.0000000
#> missingness_w_exclusions              NA           NA           NA           NA
#> outlier_count                  0.0000000    0.0000000    0.0000000    0.0000000
#> n                             98.0000000   98.0000000   98.0000000   98.0000000
#> mean                          10.8366106   10.6769531   10.3587046   10.6814030
#> sd                             2.3653804    2.1175900    2.0374511    1.9893292
#> median                        10.7928851   10.8960580   10.0828377   10.5076633
#> min                            5.8143788    6.1217538    5.2665056    6.3705674
#> max                           15.8743659   16.0471834   16.0558511   15.4495632
#> range                         10.0599871    9.9254296   10.7893455    9.0789958
#> skew                           0.1125421   -0.1941829    0.2145792    0.1446747
#> kurtosis                      -0.7778044   -0.4928252    0.7990594   -0.5675739
#> se                             0.2389395    0.2139089    0.2058136    0.2009526
#> missing                        0.0000000    0.0000000    0.0000000    0.0000000
#> var                            5.5950243    4.4841873    4.1512070    3.9574305
#> disp_index                     0.5163076    0.4199875    0.4007458    0.3704973
#> coef_variance                  0.2182768    0.1983328    0.1966898    0.1862423
#> W                              0.9836757    0.9847702    0.9760596    0.9895660
#> log10_W                        0.9795734    0.9566843    0.9548016    0.9866458
#> k                             21.0000000   34.0000000   35.0000000   36.0000000
#> independent_features_binary    0.0000000    1.0000000    1.0000000    1.0000000
#>                             unmapped_163 unmapped_164 unmapped_165 unmapped_166
#> missingness                   0.00000000    0.0000000    0.0000000    0.0000000
#> missingness_w_exclusions              NA           NA           NA           NA
#> outlier_count                 0.00000000    0.0000000    0.0000000    0.0000000
#> n                            98.00000000   98.0000000   98.0000000   98.0000000
#> mean                         10.70789869   10.5398807   10.4479747   10.6547487
#> sd                            2.14661485    1.9625016    2.1008292    2.1388041
#> median                       10.88924753   10.7648836   10.5623599   10.8771079
#> min                           6.22201580    5.8004820    3.2543822    5.1230639
#> max                          14.87566035   14.9815896   15.6579668   14.5931670
#> range                         8.65364455    9.1811076   12.4035846    9.4701031
#> skew                         -0.09189465   -0.1092942   -0.2541996   -0.2722993
#> kurtosis                     -1.00202302   -0.6026189    0.3833767   -0.4623233
#> se                            0.21684085    0.1982426    0.2122158    0.2160518
#> missing                       0.00000000    0.0000000    0.0000000    0.0000000
#> var                           4.60795533    3.8514126    4.4134833    4.5744829
#> disp_index                    0.43033236    0.3654133    0.4224248    0.4293375
#> coef_variance                 0.20047022    0.1861977    0.2010753    0.2007372
#> W                             0.97494255    0.9890195    0.9861168    0.9838733
#> log10_W                       0.96304713    0.9719281    0.9112032    0.9508298
#> k                             4.00000000   37.0000000   38.0000000   39.0000000
#> independent_features_binary   0.00000000    1.0000000    1.0000000    1.0000000
#>                             unmapped_167 unmapped_168 unmapped_169 unmapped_170
#> missingness                    0.0000000   0.00000000    0.0000000    0.0000000
#> missingness_w_exclusions              NA           NA           NA           NA
#> outlier_count                  0.0000000   0.00000000    0.0000000    0.0000000
#> n                             98.0000000  98.00000000   98.0000000   98.0000000
#> mean                          10.6721561  10.66080797   10.6221655   10.4610811
#> sd                             2.1645133   2.20684544    2.2430177    2.1263963
#> median                        10.5343591  10.71432830   10.7127545   10.5066474
#> min                            6.7176719   4.33652944    4.8581078    5.6935272
#> max                           17.2249913  16.40866973   16.1000447   15.1421038
#> range                         10.5073194  12.07214029   11.2419369    9.4485766
#> skew                           0.4211277   0.02401142   -0.1599331    0.1292541
#> kurtosis                      -0.2155873   0.08936438   -0.3296538   -0.6508847
#> se                             0.2186489   0.22292505    0.2265790    0.2147985
#> missing                        0.0000000   0.00000000    0.0000000    0.0000000
#> var                            4.6851177   4.87016678    5.0311283    4.5215612
#> disp_index                     0.4390039   0.45682905    0.4736443    0.4322270
#> coef_variance                  0.2028187   0.20700546    0.2111639    0.2032674
#> W                              0.9810500   0.99286554    0.9894272    0.9864548
#> log10_W                        0.9917721   0.96400695    0.9591631    0.9834128
#> k                             10.0000000  40.00000000   34.0000000   25.0000000
#> independent_features_binary    0.0000000   1.00000000    0.0000000    0.0000000
#>                             chem_id_171 unmapped_172 unmapped_173 unmapped_174
#> missingness                   0.0000000    0.0000000    0.0000000    0.0000000
#> missingness_w_exclusions             NA           NA           NA           NA
#> outlier_count                 0.0000000    0.0000000    0.0000000    0.0000000
#> n                            98.0000000   98.0000000   98.0000000   98.0000000
#> mean                         10.3027145   10.4525467    9.6035110   10.1440795
#> sd                            2.1414218    2.1125229    2.0555047    2.4888351
#> median                       10.0107216   10.3498835    9.6440956   10.1337933
#> min                           5.8141787    6.0096646    5.0666758    5.3613410
#> max                          15.5514994   16.5188674   14.8779197   22.7845433
#> range                         9.7373207   10.5092028    9.8112439   17.4232023
#> skew                          0.3573120    0.3899260    0.0683601    1.3120968
#> kurtosis                     -0.3182159   -0.1621296    0.2055001    5.2804590
#> se                            0.2163163    0.2133970    0.2076373    0.2514103
#> missing                       0.0000000    0.0000000    0.0000000    0.0000000
#> var                           4.5856872    4.4627529    4.2250997    6.1943002
#> disp_index                    0.4450950    0.4269536    0.4399537    0.6106321
#> coef_variance                 0.2078502    0.2021060    0.2140368    0.2453485
#> W                             0.9823750    0.9792805    0.9804163    0.9227785
#> log10_W                       0.9904952    0.9891518    0.9560395    0.9808063
#> k                            41.0000000   42.0000000   43.0000000   44.0000000
#> independent_features_binary   1.0000000    1.0000000    1.0000000    1.0000000
#>                             unmapped_175 unmapped_176 unmapped_177 unmapped_178
#> missingness                    0.0000000    0.0000000   0.00000000    0.0000000
#> missingness_w_exclusions              NA           NA           NA           NA
#> outlier_count                  0.0000000    0.0000000   0.00000000    0.0000000
#> n                             98.0000000   98.0000000  98.00000000   98.0000000
#> mean                          10.0038980   10.3412998  10.14719712   10.1821726
#> sd                             2.0603859    2.2282017   2.30268383    2.3865042
#> median                         9.9245639   10.2953770  10.35774283   10.0774144
#> min                            4.3624048    3.4309803   4.66376175    4.0724943
#> max                           19.0543767   17.1072576  15.60236652   17.7474711
#> range                         14.6919719   13.6762772  10.93860477   13.6749768
#> skew                           1.1659306    0.1214373  -0.06813588    0.1045125
#> kurtosis                       4.6133012    0.8253542  -0.38453358    0.1585321
#> se                             0.2081304    0.2250824   0.23260619    0.2410733
#> missing                        0.0000000    0.0000000   0.00000000    0.0000000
#> var                            4.2451902    4.9648830   5.30235284    5.6954025
#> disp_index                     0.4243536    0.4801024   0.52254359    0.5593504
#> coef_variance                  0.2059583    0.2154663   0.22692807    0.2343807
#> W                              0.9131019    0.9882403   0.99199542    0.9884564
#> log10_W                        0.9498913    0.9395862   0.96497909    0.9659827
#> k                             45.0000000   46.0000000  47.00000000   48.0000000
#> independent_features_binary    1.0000000    1.0000000   1.00000000    1.0000000
#>                             chem_id_179 chem_id_180 chem_id_181 chem_id_182
#> missingness                  0.00000000   0.0000000  0.00000000   0.0000000
#> missingness_w_exclusions             NA          NA          NA          NA
#> outlier_count                0.00000000   0.0000000  0.00000000   0.0000000
#> n                           98.00000000  98.0000000 98.00000000  98.0000000
#> mean                        10.08958927  10.3823477  9.90540540  10.0557349
#> sd                           2.22497608   2.3704140  2.09858960   2.4976811
#> median                      10.08117360  10.3112536 10.17250639   9.8026383
#> min                          5.69285226   5.3298024  5.09802906   2.7609170
#> max                         16.95121798  18.4919064 16.26199286  18.2120819
#> range                       11.25836572  13.1621039 11.16396380  15.4511649
#> skew                         0.22649912   0.2640977 -0.06002203   0.1737799
#> kurtosis                    -0.01418983   0.3153370  0.25322505   1.4185632
#> se                           0.22475652   0.2394480  0.21198956   0.2523039
#> missing                      0.00000000   0.0000000  0.00000000   0.0000000
#> var                          4.95051854   5.6188627  4.40407829   6.2384109
#> disp_index                   0.49065610   0.5411939  0.44461363   0.6203834
#> coef_variance                0.22052197   0.2283119  0.21186307   0.2483837
#> W                            0.98643234   0.9859205  0.97224245   0.9701981
#> log10_W                      0.97897333   0.9782932  0.94370133   0.8918646
#> k                           49.00000000  50.0000000 51.00000000  49.0000000
#> independent_features_binary  1.00000000   1.0000000  1.00000000   0.0000000
#>                             unmapped_183 unmapped_184 unmapped_185 unmapped_186
#> missingness                    0.0000000    0.0000000    0.0000000   0.00000000
#> missingness_w_exclusions              NA           NA           NA           NA
#> outlier_count                  0.0000000    0.0000000    0.0000000   0.00000000
#> n                             98.0000000   98.0000000   98.0000000  98.00000000
#> mean                          10.4244553   10.5756534   10.0774694   9.87989807
#> sd                             2.4347427    2.7786893    2.2154269   2.23849898
#> median                        10.2867149   10.0904585   10.1615477   9.89787854
#> min                            3.7300532    5.0784576    2.3829405   3.42866404
#> max                           16.7965550   24.2419261   16.2122970  16.06164157
#> range                         13.0665018   19.1634685   13.8293565  12.63297753
#> skew                           0.1736296    1.4444331   -0.3737524  -0.06077615
#> kurtosis                       0.1126305    4.7052896    1.2641860   0.16798966
#> se                             0.2459462    0.2806900    0.2237919   0.22612254
#> missing                        0.0000000    0.0000000    0.0000000   0.00000000
#> var                            5.9279722    7.7211141    4.9081165   5.01087766
#> disp_index                     0.5686601    0.7300839    0.4870386   0.50717909
#> coef_variance                  0.2335607    0.2627440    0.2198396   0.22657106
#> W                              0.9911449    0.9143018    0.9771091   0.99439928
#> log10_W                        0.9659660    0.9878595    0.8402097   0.94664808
#> k                             52.0000000   53.0000000   54.0000000  55.00000000
#> independent_features_binary    1.0000000    1.0000000    1.0000000   1.00000000
#>                             unmapped_187 unmapped_188 chem_id_189 unmapped_190
#> missingness                    0.0000000   0.00000000   0.0000000    0.0000000
#> missingness_w_exclusions              NA           NA          NA           NA
#> outlier_count                  0.0000000   0.00000000   0.0000000    0.0000000
#> n                             98.0000000  98.00000000  98.0000000   98.0000000
#> mean                          10.2697137   9.86701173  10.3516972    9.8703503
#> sd                             2.0787822   2.18978199   2.4261962    2.3328699
#> median                        10.1805461   9.74083477  10.2509258    9.7359161
#> min                            5.5487302   4.62429129   5.0359416    4.2584568
#> max                           14.3737642  16.08675343  20.8905566   17.6110248
#> range                          8.8250340  11.46246214  15.8546150   13.3525680
#> skew                          -0.1792282   0.23849923   0.8435299    0.3132117
#> kurtosis                      -0.5265802   0.03576841   2.2855156    0.4559636
#> se                             0.2099887   0.22120139   0.2450828    0.2356554
#> missing                        0.0000000   0.00000000   0.0000000    0.0000000
#> var                            4.3213355   4.79514518   5.8864279    5.4422821
#> disp_index                     0.4207844   0.48597745   0.5686438    0.5513768
#> coef_variance                  0.2024187   0.22192960   0.2343767    0.2363513
#> W                              0.9817248   0.99027465   0.9575812    0.9880106
#> log10_W                        0.9547933   0.98218174   0.9882156    0.9761967
#> k                             56.0000000  57.00000000  58.0000000   59.0000000
#> independent_features_binary    1.0000000   1.00000000   1.0000000    1.0000000
#>                             unmapped_191 chem_id_192 unmapped_193 chem_id_194
#> missingness                    0.0000000   0.0000000    0.0000000   0.0000000
#> missingness_w_exclusions              NA          NA           NA          NA
#> outlier_count                  0.0000000   0.0000000    1.0000000   0.0000000
#> n                             98.0000000  98.0000000   98.0000000  98.0000000
#> mean                          10.1301217   9.9358165   10.1172656  10.6771117
#> sd                             2.3857164   2.1174926    2.6018950   2.5511777
#> median                        10.0204111  10.0246707    9.9638025  10.4613522
#> min                            4.0611064   3.0441432    4.3238970   5.1158987
#> max                           18.3833658  17.4755027   24.7215305  19.5447316
#> range                         14.3222594  14.4313595   20.3976335  14.4288329
#> skew                           0.4850897   0.1408904    1.7573748   0.7243388
#> kurtosis                       1.3142505   1.7135094    8.6096600   0.6993475
#> se                             0.2409938   0.2138991    0.2628311   0.2577079
#> missing                        0.0000000   0.0000000    0.0000000   0.0000000
#> var                            5.6916429   4.4837751    6.7698578   6.5085078
#> disp_index                     0.5618534   0.4512739    0.6691391   0.6095757
#> coef_variance                  0.2355072   0.2131171    0.2571737   0.2389389
#> W                              0.9726119   0.9736626    0.8854242   0.9637304
#> log10_W                        0.9637587   0.9136494    0.9669361   0.9907161
#> k                             53.0000000  60.0000000   47.0000000  61.0000000
#> independent_features_binary    0.0000000   1.0000000    0.0000000   1.0000000
#>                             unmapped_195 unmapped_196 chem_id_197 unmapped_198
#> missingness                    0.0000000   0.00000000   0.0000000    0.0000000
#> missingness_w_exclusions              NA           NA          NA           NA
#> outlier_count                  0.0000000   0.00000000   0.0000000    0.0000000
#> n                             98.0000000  98.00000000  98.0000000   98.0000000
#> mean                           9.9167628  10.11424118   9.9285607   10.0397572
#> sd                             2.2153442   2.10954687   2.2823284    2.3066514
#> median                         9.9111488  10.08779918  10.1783251   10.0959462
#> min                            5.0285511   4.11670564   4.7539321    4.9695857
#> max                           17.1639027  17.44203262  15.4771130   14.5195265
#> range                         12.1353516  13.32532698  10.7231808    9.5499408
#> skew                           0.5367906   0.08751949  -0.1145303   -0.1121376
#> kurtosis                       0.7704379   0.82100302  -0.3310632   -0.5709693
#> se                             0.2237836   0.21309641   0.2305500    0.2330070
#> missing                        0.0000000   0.00000000   0.0000000    0.0000000
#> var                            4.9077500   4.45018799   5.2090228    5.3206408
#> disp_index                     0.4948944   0.43999228   0.5246503    0.5299571
#> coef_variance                  0.2233939   0.20857194   0.2298750    0.2297517
#> W                              0.9768343   0.98840734   0.9891089    0.9868166
#> log10_W                        0.9900198   0.95773268   0.9558242    0.9588789
#> k                             62.0000000  56.00000000  63.0000000   64.0000000
#> independent_features_binary    1.0000000   0.00000000   1.0000000    1.0000000
#>                             unmapped_199 unmapped_200 unmapped_201 unmapped_202
#> missingness                    0.0000000    0.0000000    0.0000000    0.0000000
#> missingness_w_exclusions              NA           NA           NA           NA
#> outlier_count                  0.0000000    0.0000000    0.0000000    0.0000000
#> n                             98.0000000   98.0000000   98.0000000   98.0000000
#> mean                           9.8414919   10.1882972    9.8346476    9.8847242
#> sd                             1.9172482    2.4020207    2.4658845    1.9617541
#> median                         9.8744366   10.0419147    9.7578528    9.8872217
#> min                            5.7191754    4.7123055    2.7925263    5.6686602
#> max                           16.3754948   17.8453907   20.3770604   17.7062994
#> range                         10.6563193   13.1330852   17.5845341   12.0376392
#> skew                           0.3490506    0.5944353    0.6466371    0.5460555
#> kurtosis                       0.1579179    0.5545365    2.5117330    1.2819056
#> se                             0.1936713    0.2426407    0.2490920    0.1981671
#> missing                        0.0000000    0.0000000    0.0000000    0.0000000
#> var                            3.6758405    5.7697033    6.0805864    3.8484791
#> disp_index                     0.3735044    0.5663069    0.6182821    0.3893360
#> coef_variance                  0.1948128    0.2357627    0.2507344    0.1984632
#> W                              0.9863771    0.9709731    0.9646315    0.9737354
#> log10_W                        0.9917607    0.9894457    0.9447968    0.9885303
#> k                             65.0000000   66.0000000   67.0000000   45.0000000
#> independent_features_binary    1.0000000    1.0000000    1.0000000    0.0000000
#>                             chem_id_203 unmapped_204 unmapped_205 unmapped_206
#> missingness                   0.0000000    0.0000000    0.0000000   0.00000000
#> missingness_w_exclusions             NA           NA           NA           NA
#> outlier_count                 0.0000000    0.0000000    0.0000000   0.00000000
#> n                            98.0000000   98.0000000   98.0000000  98.00000000
#> mean                         10.4988196   10.5518762   10.0191891   9.95622009
#> sd                            2.1455915    2.5564921    2.0470155   2.32038871
#> median                       10.4721576   10.3730681   10.2398553   9.87654087
#> min                           5.7165722    3.4692435    4.7251392   4.18662524
#> max                          15.1587184   18.8550782   14.4313509  16.47731253
#> range                         9.4421463   15.3858347    9.7062117  12.29068729
#> skew                          0.1504186    0.4094812   -0.1766162   0.34821301
#> kurtosis                     -0.3763365    0.6768486   -0.1895491  -0.03248259
#> se                            0.2167375    0.2582447    0.2067798   0.23439466
#> missing                       0.0000000    0.0000000    0.0000000   0.00000000
#> var                           4.6035629    6.5356518    4.1902726   5.38420376
#> disp_index                    0.4384839    0.6193829    0.4182247   0.54078794
#> coef_variance                 0.2043650    0.2422784    0.2043095   0.23305920
#> W                             0.9834705    0.9830054    0.9870500   0.98639701
#> log10_W                       0.9757169    0.9644719    0.9508229   0.98537813
#> k                            68.0000000   46.0000000   69.0000000  70.00000000
#> independent_features_binary   1.0000000    0.0000000    1.0000000   1.00000000
#>                             unmapped_207 unmapped_208 unmapped_209 unmapped_210
#> missingness                    0.0000000   0.00000000    0.0000000    0.0000000
#> missingness_w_exclusions              NA           NA           NA           NA
#> outlier_count                  0.0000000   0.00000000    0.0000000    1.0000000
#> n                             98.0000000  98.00000000   98.0000000   98.0000000
#> mean                          10.0400801  10.13730051    9.7808249    9.8634453
#> sd                             2.7570304   2.30061949    2.2230331    2.5251522
#> median                         9.7673083   9.88712994    9.9366658    9.6142769
#> min                            3.8658018   4.21090940    4.6720925    5.2217042
#> max                           22.0649134  16.13837257   15.4184478   24.6653608
#> range                         18.1991116  11.92746317   10.7463553   19.4436566
#> skew                           0.8439699  -0.04646507   -0.0195292    2.0383512
#> kurtosis                       2.6417288   0.06785008   -0.4374552   10.0990194
#> se                             0.2785021   0.23239766    0.2245603    0.2550789
#> missing                        0.0000000   0.00000000    0.0000000    0.0000000
#> var                            7.6012166   5.29285004    4.9418760    6.3763938
#> disp_index                     0.7570872   0.52211632    0.5052617    0.6464672
#> coef_variance                  0.2746024   0.22694597    0.2272848    0.2560112
#> W                              0.9507812   0.99266613    0.9918351    0.8637020
#> log10_W                        0.9684927   0.95145442    0.9688747    0.9700409
#> k                             71.0000000  72.00000000   73.0000000   49.0000000
#> independent_features_binary    1.0000000   1.00000000    1.0000000    0.0000000
#>                             unmapped_211 chem_id_212 unmapped_213 unmapped_214
#> missingness                    0.0000000   0.0000000   0.00000000   0.00000000
#> missingness_w_exclusions              NA          NA           NA           NA
#> outlier_count                  0.0000000   0.0000000   0.00000000   0.00000000
#> n                             98.0000000  98.0000000  98.00000000  98.00000000
#> mean                          10.0232840   9.8152608  10.29222314  10.02801964
#> sd                             2.3292309   2.0720916   2.25368427   2.31204834
#> median                         9.9115049   9.8684922  10.24472285  10.27199739
#> min                            4.9527574   2.9990323   5.12998896   4.92361159
#> max                           18.0545517  16.0004936  15.22476136  14.57647197
#> range                         13.1017944  13.0014613  10.09477240   9.65286038
#> skew                           0.7391058  -0.4045024   0.04050389  -0.05656143
#> kurtosis                       1.1920295   1.1625248  -0.64015869  -0.68033658
#> se                             0.2352878   0.2093129   0.22765649   0.23355215
#> missing                        0.0000000   0.0000000   0.00000000   0.00000000
#> var                            5.4253164   4.2935638   5.07909281   5.34556751
#> disp_index                     0.5412713   0.4374376   0.49348841   0.53306313
#> coef_variance                  0.2323820   0.2111092   0.21896963   0.23055882
#> W                              0.9658605   0.9731735   0.98994378   0.98553314
#> log10_W                        0.9915976   0.8809632   0.97802913   0.96507537
#> k                             56.0000000  74.0000000  61.00000000  63.00000000
#> independent_features_binary    0.0000000   1.0000000   0.00000000   0.00000000
#>                             unmapped_215 unmapped_216 unmapped_217 unmapped_218
#> missingness                    0.0000000    0.0000000    0.0000000    0.0000000
#> missingness_w_exclusions              NA           NA           NA           NA
#> outlier_count                  0.0000000    0.0000000    0.0000000    0.0000000
#> n                             98.0000000   98.0000000   98.0000000   98.0000000
#> mean                          10.1137186   10.4678940   10.1009156   10.2280847
#> sd                             2.5184156    2.6521404    2.3025373    2.4127177
#> median                        10.0470286   10.1657584   10.2576757   10.0210977
#> min                            4.9229712    5.2598816    4.1023416    5.3441774
#> max                           20.8076499   21.5882858   19.8033239   16.4046289
#> range                         15.8846787   16.3284042   15.7009823   11.0604515
#> skew                           0.9559761    1.1328220    0.5339491    0.2336359
#> kurtosis                       2.3774416    2.8229303    2.1604915   -0.4098178
#> se                             0.2543984    0.2679066    0.2325914    0.2437213
#> missing                        0.0000000    0.0000000    0.0000000    0.0000000
#> var                            6.3424169    7.0338486    5.3016781    5.8212069
#> disp_index                     0.6271103    0.6719450    0.5248710    0.5691395
#> coef_variance                  0.2490099    0.2533595    0.2279533    0.2358915
#> W                              0.9524372    0.9361872    0.9664129    0.9874544
#> log10_W                        0.9893684    0.9896489    0.9652891    0.9837448
#> k                             72.0000000   61.0000000   47.0000000   75.0000000
#> independent_features_binary    0.0000000    0.0000000    0.0000000    1.0000000
#>                             unmapped_219 unmapped_220 unmapped_221 unmapped_222
#> missingness                    0.0000000  0.000000000   0.00000000   0.00000000
#> missingness_w_exclusions              NA           NA           NA           NA
#> outlier_count                  0.0000000  0.000000000   0.00000000   0.00000000
#> n                             98.0000000 98.000000000  98.00000000  98.00000000
#> mean                           9.9903254  9.872490518  10.20322791   9.66283284
#> sd                             2.2299573  2.332819287   2.04118385   2.40558959
#> median                         9.7107253  9.786662249  10.39698620   9.77523702
#> min                            5.5040388  4.595012831   5.85548991   4.33541373
#> max                           16.7112714 15.637964240  15.71087146  16.30274092
#> range                         11.2072326 11.042951409   9.85538155  11.96732719
#> skew                           0.6476204 -0.009230954   0.06288604   0.05672121
#> kurtosis                       0.3900279 -0.455048392  -0.28959135  -0.24691983
#> se                             0.2252597  0.235650334   0.20619071   0.24300124
#> missing                        0.0000000  0.000000000   0.00000000   0.00000000
#> var                            4.9727097  5.442045826   4.16643149   5.78686126
#> disp_index                     0.4977525  0.551233330   0.40834445   0.59887834
#> coef_variance                  0.2232117  0.236294913   0.20005275   0.24895283
#> W                              0.9710423  0.994446721   0.98397676   0.99089116
#> log10_W                        0.9960236  0.971161779   0.97024136   0.96770312
#> k                             76.0000000 77.000000000  58.00000000  71.00000000
#> independent_features_binary    1.0000000  1.000000000   0.00000000   0.00000000