Retrieve Feature Data

Extracts feature-level data from a Metabolites object for a specified data layer, with optional exclusions and subsetting.

Usage

get_features(metabolites, layer, feature_ids = NULL, as_df = FALSE)

Arguments

metabolites

A Metabolites object containing feature data.

layer

A character string specifying the data layer from which to extract features.

feature_ids

Optional. A vector of row names to extract. Defaults to NULL, meaning all rows are included.

as_df

Logical. Whether to return the result as a data.frame. Defaults to FALSE, returning a matrix.

Value

A data.table or data.frame (if as_df = TRUE) containing feature data for the specified layer.

Examples

file <- system.file("extdata", "metabolon_v1_example.xlsx", package = "metaboprep2")
metabolites <- Metabolites(project_name = "MYPROJECT",
                           format       = "metabolon_v1",
                           filepath     = file)
metabolites <- import_data(metabolites)
#> Using cached MetaboAnalyst DB.
#> Attempting join [features$hmdb] ~ [annotation$hmdb_id] 
#> Attempting join [features$kegg] ~ [annotation$kegg_id] 
#> Attempting join [features$metabolite_id] ~ [annotation$name] 
#> Attempting join [features$comp_id] ~ [annotation$comp_id] 
#> Warning: 100 unannotated metabolites remain after merging with metabolomics database
metabolites <- metabolite_qc(metabolites, source="raw")
#> 
#> ── Starting Metabolite QC Process ──────────────────────────────────────────────
#> ℹ Validating input parameters...
#> ✔ Input validation complete.
#> ℹ Batch normalising raw data...
#> ✔ Normalised raw data.
#> ℹ Applying exclusion of xenobiotic features...
#> ✔ Xenobiotic features exclusions assessment complete. (3 features excluded)
#> ℹ Analysing exclusions...
#> ℹ Applying exclusion of derived features...
#> ✔ Derived features exclusions assessment complete. (0 features excluded)
#> ℹ Sample & Feature Summary Statistics for raw data...
#> ✔ Sample & Feature Summary Statistics for raw data...
#> ℹ Copying batch_normalised data to new 'qcing' data layer...
#> ✔ qcing data ready for QC.
#> ℹ Assessing for extreme sample missingness >=80%...
#> ✔ Extreme sample missingness exclusions assessment complete. (0 samples excluded)
#> ℹ Assessing for extreme feature missingness >=80%...
#> ✔ Extreme feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Assessing for total feature missingness (100%) of xenobiotics...
#> ✔ Total xenobiotic feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Assessing for sample missingness at specified level of >=20%...
#> ✔ Sample missingness exclusions assessment complete. (2 samples excluded)
#> ℹ Assessing for feature missingness at specified level of >=20%...
#> ✔ Feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Calculating total peak abundance outliers at +/- 5 Sdev...
#> ✔ Total peak abundance outlier exclusions assessment complete. (0 samples excluded)
#> ℹ Running data outlier analysis at +/- 5 Sdev...
#> ℹ Data outlier analysis complete.
#> ℹ Re-identify feature independence and PC outliers...
#> ℹ Re-running sample summary using current exclusions...
#> ℹ QC sample summary complete.
#> ℹ PCA outlier analysis 2 complete.
#> ℹ Creating final QC dataset...
#> ℹ Rerunning Sample & Feature Summary Statistics for QC data...
#> ✔ Sample & Feature Summary Statistics for QC data complete.
#> 
#> ── Metabolite QC Process Completed ─────────────────────────────────────────────
get_features(metabolites, layer = "post_qc")
#>      metabolite_id comp_id platform                pathway   kegg    hmdb
#>             <char>  <char>   <char>                 <char> <char>  <char>
#>   1:          met1     123      neg             xenobiotic   K100   HMDB1
#>   2:          met2     124      neg             xenobiotic   K100   HMDB2
#>   3:          met3     125      neg             xenobiotic   K100   HMDB3
#>   4:          met4     126      neg                  lipid   K200   HMDB4
#>   5:          met5     127      neg                  lipid   K200   HMDB5
#>   6:          met6     128      neg                  lipid   K200   HMDB6
#>   7:          met7     129      neg                  lipid   K200   HMDB7
#>   8:          met8     130      neg                  lipid   K200   HMDB8
#>   9:          met9     131      neg                  lipid   K200   HMDB9
#>  10:         met10     132      neg                  lipid   K200  HMDB10
#>  11:         met11     133      neg                  lipid   K200  HMDB11
#>  12:         met12     134      neg                  lipid   K200  HMDB12
#>  13:         met13     135      neg                  lipid   K200  HMDB13
#>  14:         met14     136      neg                  lipid   K200  HMDB14
#>  15:         met15     137      neg                  lipid   K200  HMDB15
#>  16:         met16     138      neg                  lipid   K200  HMDB16
#>  17:         met17     139      neg                  lipid   K200  HMDB17
#>  18:         met18     140      neg                  lipid   K200  HMDB18
#>  19:         met19     141      neg                  lipid   K200  HMDB19
#>  20:         met20     142      neg                  lipid   K200  HMDB20
#>  21:         met21     143      neg                  lipid   K200  HMDB21
#>  22:         met22     144      neg                  lipid   K200  HMDB22
#>  23:         met23     145      neg                  lipid   K200  HMDB23
#>  24:         met24     146      neg                  lipid   K200  HMDB24
#>  25:         met25     147      neg                  lipid   K200  HMDB25
#>  26:         met26     148      neg                  lipid   K200  HMDB26
#>  27:         met27     149      neg                  lipid   K200  HMDB27
#>  28:         met28     150      neg                  lipid   K200  HMDB28
#>  29:         met29     151      neg                  lipid   K200  HMDB29
#>  30:         met30     152      neg                  lipid   K200  HMDB30
#>  31:         met31     153      neg                  lipid   K200  HMDB31
#>  32:         met32     154      neg                  lipid   K200  HMDB32
#>  33:         met33     155      neg                  lipid   K200  HMDB33
#>  34:         met34     156      neg                  lipid   K200  HMDB34
#>  35:         met35     157      neg                  lipid   K200  HMDB35
#>  36:         met36     158      neg                  lipid   K200  HMDB36
#>  37:         met37     159      neg                  lipid   K200  HMDB37
#>  38:         met38     160      neg                  lipid   K200  HMDB38
#>  39:         met39     161      neg                  lipid   K200  HMDB39
#>  40:         met40     162      neg                  lipid   K200  HMDB40
#>  41:         met41     163      neg                  lipid   K200  HMDB41
#>  42:         met42     164      neg                  lipid   K200  HMDB42
#>  43:         met43     165      neg                  lipid   K200  HMDB43
#>  44:         met44     166      neg                  lipid   K200  HMDB44
#>  45:         met45     167      neg                  lipid   K200  HMDB45
#>  46:         met46     168      neg                  lipid   K200  HMDB46
#>  47:         met47     169      neg                  lipid   K200  HMDB47
#>  48:         met48     170      neg                  lipid   K200  HMDB48
#>  49:         met49     171      neg             amino acid   K300  HMDB49
#>  50:         met50     172      pos             amino acid   K300  HMDB50
#>  51:         met51     173      pos             amino acid   K300  HMDB51
#>  52:         met52     174      pos             amino acid   K300  HMDB52
#>  53:         met53     175      pos             amino acid   K300  HMDB53
#>  54:         met54     176      pos             amino acid   K300  HMDB54
#>  55:         met55     177      pos             amino acid   K300  HMDB55
#>  56:         met56     178      pos             amino acid   K300  HMDB56
#>  57:         met57     179      pos             amino acid   K300  HMDB57
#>  58:         met58     180      pos             amino acid   K300  HMDB58
#>  59:         met59     181      pos             amino acid   K300  HMDB59
#>  60:         met60     182      pos           carbohydrate   K200  HMDB60
#>  61:         met61     183      pos           carbohydrate   K200  HMDB61
#>  62:         met62     184      pos           carbohydrate   K200  HMDB62
#>  63:         met63     185      pos           carbohydrate   K200  HMDB63
#>  64:         met64     186      pos           carbohydrate   K200  HMDB64
#>  65:         met65     187      pos           carbohydrate   K200  HMDB65
#>  66:         met66     188      pos           carbohydrate   K200  HMDB66
#>  67:         met67     189      pos           carbohydrate   K200  HMDB67
#>  68:         met68     190      pos           carbohydrate   K200  HMDB68
#>  69:         met69     191      pos           carbohydrate   K200  HMDB69
#>  70:         met70     192      pos cofactors and vitamins   K400  HMDB70
#>  71:         met71     193      pos cofactors and vitamins   K400  HMDB71
#>  72:         met72     194      pos cofactors and vitamins   K400  HMDB72
#>  73:         met73     195      pos cofactors and vitamins   K400  HMDB73
#>  74:         met74     196      pos cofactors and vitamins   K400  HMDB74
#>  75:         met75     197      pos cofactors and vitamins   K400  HMDB75
#>  76:         met76     198      pos cofactors and vitamins   K400  HMDB76
#>  77:         met77     199      pos cofactors and vitamins   K400  HMDB77
#>  78:         met78     200      pos cofactors and vitamins   K400  HMDB78
#>  79:         met79     201      pos cofactors and vitamins   K400  HMDB79
#>  80:         met80     202      pos cofactors and vitamins   K400  HMDB80
#>  81:         met81     203      pos cofactors and vitamins   K400  HMDB81
#>  82:         met82     204      pos cofactors and vitamins   K400  HMDB82
#>  83:         met83     205      pos cofactors and vitamins   K400  HMDB83
#>  84:         met84     206      pos cofactors and vitamins   K400  HMDB84
#>  85:         met85     207      pos cofactors and vitamins   K400  HMDB85
#>  86:         met86     208      pos cofactors and vitamins   K400  HMDB86
#>  87:         met87     209      pos cofactors and vitamins   K400  HMDB87
#>  88:         met88     210      pos cofactors and vitamins   K400  HMDB88
#>  89:         met89     211      pos cofactors and vitamins   K400  HMDB89
#>  90:         met90     212      pos                 energy   K400  HMDB90
#>  91:         met91     213      pos                 energy   K400  HMDB91
#>  92:         met92     214      pos                 energy   K400  HMDB92
#>  93:         met93     215      pos                 energy   K400  HMDB93
#>  94:         met94     216      pos                 energy   K400  HMDB94
#>  95:         met95     217      pos                 energy   K400  HMDB95
#>  96:         met96     218      pos                 energy   K400  HMDB96
#>  97:         met97     219      pos                 energy   K400  HMDB97
#>  98:         met98     220      pos                 energy   K400  HMDB98
#>  99:         met99     221      pos                 energy   K400  HMDB99
#> 100:        met100     222      pos                 energy   K400 HMDB100
#>      metabolite_id comp_id platform                pathway   kegg    hmdb
#>      join_key anno_hmdb_id anno_name anno_pubchem_id anno_chebi_id anno_kegg_id
#>        <char>       <char>    <char>          <char>        <char>       <char>
#>   1:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>   2:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>   3:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>   4:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>   5:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>   6:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>   7:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>   8:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>   9:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  10:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  11:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  12:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  13:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  14:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  15:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  16:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  17:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  18:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  19:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  20:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  21:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  22:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  23:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  24:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  25:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  26:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  27:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  28:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  29:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  30:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  31:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  32:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  33:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  34:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  35:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  36:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  37:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  38:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  39:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  40:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  41:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  42:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  43:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  44:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  45:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  46:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  47:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  48:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  49:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  50:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  51:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  52:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  53:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  54:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  55:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  56:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  57:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  58:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  59:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  60:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  61:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  62:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  63:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  64:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  65:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  66:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  67:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  68:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  69:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  70:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  71:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  72:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  73:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  74:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  75:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  76:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  77:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  78:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  79:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  80:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  81:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  82:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  83:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  84:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  85:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  86:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  87:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  88:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  89:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  90:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  91:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  92:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  93:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  94:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  95:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  96:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  97:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  98:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>  99:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#> 100:     <NA>         <NA>      <NA>            <NA>          <NA>         <NA>
#>      join_key anno_hmdb_id anno_name anno_pubchem_id anno_chebi_id anno_kegg_id
#>      anno_metlin_id anno_lipid anno_smiles anno_exactmass anno_formula
#>              <char>     <char>      <char>         <char>       <char>
#>   1:           <NA>       <NA>        <NA>           <NA>         <NA>
#>   2:           <NA>       <NA>        <NA>           <NA>         <NA>
#>   3:           <NA>       <NA>        <NA>           <NA>         <NA>
#>   4:           <NA>       <NA>        <NA>           <NA>         <NA>
#>   5:           <NA>       <NA>        <NA>           <NA>         <NA>
#>   6:           <NA>       <NA>        <NA>           <NA>         <NA>
#>   7:           <NA>       <NA>        <NA>           <NA>         <NA>
#>   8:           <NA>       <NA>        <NA>           <NA>         <NA>
#>   9:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  10:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  11:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  12:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  13:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  14:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  15:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  16:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  17:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  18:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  19:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  20:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  21:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  22:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  23:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  24:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  25:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  26:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  27:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  28:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  29:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  30:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  31:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  32:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  33:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  34:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  35:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  36:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  37:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  38:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  39:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  40:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  41:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  42:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  43:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  44:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  45:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  46:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  47:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  48:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  49:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  50:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  51:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  52:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  53:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  54:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  55:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  56:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  57:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  58:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  59:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  60:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  61:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  62:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  63:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  64:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  65:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  66:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  67:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  68:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  69:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  70:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  71:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  72:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  73:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  74:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  75:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  76:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  77:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  78:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  79:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  80:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  81:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  82:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  83:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  84:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  85:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  86:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  87:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  88:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  89:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  90:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  91:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  92:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  93:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  94:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  95:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  96:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  97:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  98:           <NA>       <NA>        <NA>           <NA>         <NA>
#>  99:           <NA>       <NA>        <NA>           <NA>         <NA>
#> 100:           <NA>       <NA>        <NA>           <NA>         <NA>
#>      anno_metlin_id anno_lipid anno_smiles anno_exactmass anno_formula
#>      anno_inchi_key anno_super_class anno_main_class anno_sub_class
#>              <char>           <char>          <char>         <char>
#>   1:           <NA>             <NA>            <NA>           <NA>
#>   2:           <NA>             <NA>            <NA>           <NA>
#>   3:           <NA>             <NA>            <NA>           <NA>
#>   4:           <NA>             <NA>            <NA>           <NA>
#>   5:           <NA>             <NA>            <NA>           <NA>
#>   6:           <NA>             <NA>            <NA>           <NA>
#>   7:           <NA>             <NA>            <NA>           <NA>
#>   8:           <NA>             <NA>            <NA>           <NA>
#>   9:           <NA>             <NA>            <NA>           <NA>
#>  10:           <NA>             <NA>            <NA>           <NA>
#>  11:           <NA>             <NA>            <NA>           <NA>
#>  12:           <NA>             <NA>            <NA>           <NA>
#>  13:           <NA>             <NA>            <NA>           <NA>
#>  14:           <NA>             <NA>            <NA>           <NA>
#>  15:           <NA>             <NA>            <NA>           <NA>
#>  16:           <NA>             <NA>            <NA>           <NA>
#>  17:           <NA>             <NA>            <NA>           <NA>
#>  18:           <NA>             <NA>            <NA>           <NA>
#>  19:           <NA>             <NA>            <NA>           <NA>
#>  20:           <NA>             <NA>            <NA>           <NA>
#>  21:           <NA>             <NA>            <NA>           <NA>
#>  22:           <NA>             <NA>            <NA>           <NA>
#>  23:           <NA>             <NA>            <NA>           <NA>
#>  24:           <NA>             <NA>            <NA>           <NA>
#>  25:           <NA>             <NA>            <NA>           <NA>
#>  26:           <NA>             <NA>            <NA>           <NA>
#>  27:           <NA>             <NA>            <NA>           <NA>
#>  28:           <NA>             <NA>            <NA>           <NA>
#>  29:           <NA>             <NA>            <NA>           <NA>
#>  30:           <NA>             <NA>            <NA>           <NA>
#>  31:           <NA>             <NA>            <NA>           <NA>
#>  32:           <NA>             <NA>            <NA>           <NA>
#>  33:           <NA>             <NA>            <NA>           <NA>
#>  34:           <NA>             <NA>            <NA>           <NA>
#>  35:           <NA>             <NA>            <NA>           <NA>
#>  36:           <NA>             <NA>            <NA>           <NA>
#>  37:           <NA>             <NA>            <NA>           <NA>
#>  38:           <NA>             <NA>            <NA>           <NA>
#>  39:           <NA>             <NA>            <NA>           <NA>
#>  40:           <NA>             <NA>            <NA>           <NA>
#>  41:           <NA>             <NA>            <NA>           <NA>
#>  42:           <NA>             <NA>            <NA>           <NA>
#>  43:           <NA>             <NA>            <NA>           <NA>
#>  44:           <NA>             <NA>            <NA>           <NA>
#>  45:           <NA>             <NA>            <NA>           <NA>
#>  46:           <NA>             <NA>            <NA>           <NA>
#>  47:           <NA>             <NA>            <NA>           <NA>
#>  48:           <NA>             <NA>            <NA>           <NA>
#>  49:           <NA>             <NA>            <NA>           <NA>
#>  50:           <NA>             <NA>            <NA>           <NA>
#>  51:           <NA>             <NA>            <NA>           <NA>
#>  52:           <NA>             <NA>            <NA>           <NA>
#>  53:           <NA>             <NA>            <NA>           <NA>
#>  54:           <NA>             <NA>            <NA>           <NA>
#>  55:           <NA>             <NA>            <NA>           <NA>
#>  56:           <NA>             <NA>            <NA>           <NA>
#>  57:           <NA>             <NA>            <NA>           <NA>
#>  58:           <NA>             <NA>            <NA>           <NA>
#>  59:           <NA>             <NA>            <NA>           <NA>
#>  60:           <NA>             <NA>            <NA>           <NA>
#>  61:           <NA>             <NA>            <NA>           <NA>
#>  62:           <NA>             <NA>            <NA>           <NA>
#>  63:           <NA>             <NA>            <NA>           <NA>
#>  64:           <NA>             <NA>            <NA>           <NA>
#>  65:           <NA>             <NA>            <NA>           <NA>
#>  66:           <NA>             <NA>            <NA>           <NA>
#>  67:           <NA>             <NA>            <NA>           <NA>
#>  68:           <NA>             <NA>            <NA>           <NA>
#>  69:           <NA>             <NA>            <NA>           <NA>
#>  70:           <NA>             <NA>            <NA>           <NA>
#>  71:           <NA>             <NA>            <NA>           <NA>
#>  72:           <NA>             <NA>            <NA>           <NA>
#>  73:           <NA>             <NA>            <NA>           <NA>
#>  74:           <NA>             <NA>            <NA>           <NA>
#>  75:           <NA>             <NA>            <NA>           <NA>
#>  76:           <NA>             <NA>            <NA>           <NA>
#>  77:           <NA>             <NA>            <NA>           <NA>
#>  78:           <NA>             <NA>            <NA>           <NA>
#>  79:           <NA>             <NA>            <NA>           <NA>
#>  80:           <NA>             <NA>            <NA>           <NA>
#>  81:           <NA>             <NA>            <NA>           <NA>
#>  82:           <NA>             <NA>            <NA>           <NA>
#>  83:           <NA>             <NA>            <NA>           <NA>
#>  84:           <NA>             <NA>            <NA>           <NA>
#>  85:           <NA>             <NA>            <NA>           <NA>
#>  86:           <NA>             <NA>            <NA>           <NA>
#>  87:           <NA>             <NA>            <NA>           <NA>
#>  88:           <NA>             <NA>            <NA>           <NA>
#>  89:           <NA>             <NA>            <NA>           <NA>
#>  90:           <NA>             <NA>            <NA>           <NA>
#>  91:           <NA>             <NA>            <NA>           <NA>
#>  92:           <NA>             <NA>            <NA>           <NA>
#>  93:           <NA>             <NA>            <NA>           <NA>
#>  94:           <NA>             <NA>            <NA>           <NA>
#>  95:           <NA>             <NA>            <NA>           <NA>
#>  96:           <NA>             <NA>            <NA>           <NA>
#>  97:           <NA>             <NA>            <NA>           <NA>
#>  98:           <NA>             <NA>            <NA>           <NA>
#>  99:           <NA>             <NA>            <NA>           <NA>
#> 100:           <NA>             <NA>            <NA>           <NA>
#>      anno_inchi_key anno_super_class anno_main_class anno_sub_class
#>      anno_synonym anno_chem_id anno_comp_id anno_derived_feature
#>            <char>       <char>       <char>               <char>
#>   1:         <NA>         <NA>         <NA>                 <NA>
#>   2:         <NA>         <NA>         <NA>                 <NA>
#>   3:         <NA>         <NA>         <NA>                 <NA>
#>   4:         <NA>         <NA>         <NA>                 <NA>
#>   5:         <NA>         <NA>         <NA>                 <NA>
#>   6:         <NA>         <NA>         <NA>                 <NA>
#>   7:         <NA>         <NA>         <NA>                 <NA>
#>   8:         <NA>         <NA>         <NA>                 <NA>
#>   9:         <NA>         <NA>         <NA>                 <NA>
#>  10:         <NA>         <NA>         <NA>                 <NA>
#>  11:         <NA>         <NA>         <NA>                 <NA>
#>  12:         <NA>         <NA>         <NA>                 <NA>
#>  13:         <NA>         <NA>         <NA>                 <NA>
#>  14:         <NA>         <NA>         <NA>                 <NA>
#>  15:         <NA>         <NA>         <NA>                 <NA>
#>  16:         <NA>         <NA>         <NA>                 <NA>
#>  17:         <NA>         <NA>         <NA>                 <NA>
#>  18:         <NA>         <NA>         <NA>                 <NA>
#>  19:         <NA>         <NA>         <NA>                 <NA>
#>  20:         <NA>         <NA>         <NA>                 <NA>
#>  21:         <NA>         <NA>         <NA>                 <NA>
#>  22:         <NA>         <NA>         <NA>                 <NA>
#>  23:         <NA>         <NA>         <NA>                 <NA>
#>  24:         <NA>         <NA>         <NA>                 <NA>
#>  25:         <NA>         <NA>         <NA>                 <NA>
#>  26:         <NA>         <NA>         <NA>                 <NA>
#>  27:         <NA>         <NA>         <NA>                 <NA>
#>  28:         <NA>         <NA>         <NA>                 <NA>
#>  29:         <NA>         <NA>         <NA>                 <NA>
#>  30:         <NA>         <NA>         <NA>                 <NA>
#>  31:         <NA>         <NA>         <NA>                 <NA>
#>  32:         <NA>         <NA>         <NA>                 <NA>
#>  33:         <NA>         <NA>         <NA>                 <NA>
#>  34:         <NA>         <NA>         <NA>                 <NA>
#>  35:         <NA>         <NA>         <NA>                 <NA>
#>  36:         <NA>         <NA>         <NA>                 <NA>
#>  37:         <NA>         <NA>         <NA>                 <NA>
#>  38:         <NA>         <NA>         <NA>                 <NA>
#>  39:         <NA>         <NA>         <NA>                 <NA>
#>  40:         <NA>         <NA>         <NA>                 <NA>
#>  41:         <NA>         <NA>         <NA>                 <NA>
#>  42:         <NA>         <NA>         <NA>                 <NA>
#>  43:         <NA>         <NA>         <NA>                 <NA>
#>  44:         <NA>         <NA>         <NA>                 <NA>
#>  45:         <NA>         <NA>         <NA>                 <NA>
#>  46:         <NA>         <NA>         <NA>                 <NA>
#>  47:         <NA>         <NA>         <NA>                 <NA>
#>  48:         <NA>         <NA>         <NA>                 <NA>
#>  49:         <NA>         <NA>         <NA>                 <NA>
#>  50:         <NA>         <NA>         <NA>                 <NA>
#>  51:         <NA>         <NA>         <NA>                 <NA>
#>  52:         <NA>         <NA>         <NA>                 <NA>
#>  53:         <NA>         <NA>         <NA>                 <NA>
#>  54:         <NA>         <NA>         <NA>                 <NA>
#>  55:         <NA>         <NA>         <NA>                 <NA>
#>  56:         <NA>         <NA>         <NA>                 <NA>
#>  57:         <NA>         <NA>         <NA>                 <NA>
#>  58:         <NA>         <NA>         <NA>                 <NA>
#>  59:         <NA>         <NA>         <NA>                 <NA>
#>  60:         <NA>         <NA>         <NA>                 <NA>
#>  61:         <NA>         <NA>         <NA>                 <NA>
#>  62:         <NA>         <NA>         <NA>                 <NA>
#>  63:         <NA>         <NA>         <NA>                 <NA>
#>  64:         <NA>         <NA>         <NA>                 <NA>
#>  65:         <NA>         <NA>         <NA>                 <NA>
#>  66:         <NA>         <NA>         <NA>                 <NA>
#>  67:         <NA>         <NA>         <NA>                 <NA>
#>  68:         <NA>         <NA>         <NA>                 <NA>
#>  69:         <NA>         <NA>         <NA>                 <NA>
#>  70:         <NA>         <NA>         <NA>                 <NA>
#>  71:         <NA>         <NA>         <NA>                 <NA>
#>  72:         <NA>         <NA>         <NA>                 <NA>
#>  73:         <NA>         <NA>         <NA>                 <NA>
#>  74:         <NA>         <NA>         <NA>                 <NA>
#>  75:         <NA>         <NA>         <NA>                 <NA>
#>  76:         <NA>         <NA>         <NA>                 <NA>
#>  77:         <NA>         <NA>         <NA>                 <NA>
#>  78:         <NA>         <NA>         <NA>                 <NA>
#>  79:         <NA>         <NA>         <NA>                 <NA>
#>  80:         <NA>         <NA>         <NA>                 <NA>
#>  81:         <NA>         <NA>         <NA>                 <NA>
#>  82:         <NA>         <NA>         <NA>                 <NA>
#>  83:         <NA>         <NA>         <NA>                 <NA>
#>  84:         <NA>         <NA>         <NA>                 <NA>
#>  85:         <NA>         <NA>         <NA>                 <NA>
#>  86:         <NA>         <NA>         <NA>                 <NA>
#>  87:         <NA>         <NA>         <NA>                 <NA>
#>  88:         <NA>         <NA>         <NA>                 <NA>
#>  89:         <NA>         <NA>         <NA>                 <NA>
#>  90:         <NA>         <NA>         <NA>                 <NA>
#>  91:         <NA>         <NA>         <NA>                 <NA>
#>  92:         <NA>         <NA>         <NA>                 <NA>
#>  93:         <NA>         <NA>         <NA>                 <NA>
#>  94:         <NA>         <NA>         <NA>                 <NA>
#>  95:         <NA>         <NA>         <NA>                 <NA>
#>  96:         <NA>         <NA>         <NA>                 <NA>
#>  97:         <NA>         <NA>         <NA>                 <NA>
#>  98:         <NA>         <NA>         <NA>                 <NA>
#>  99:         <NA>         <NA>         <NA>                 <NA>
#> 100:         <NA>         <NA>         <NA>                 <NA>
#>      anno_synonym anno_chem_id anno_comp_id anno_derived_feature
#>      anno_split_synonym  feature_id derived_feature
#>                  <char>      <char>          <char>
#>   1:               <NA> comp_id_123            <NA>
#>   2:               <NA> comp_id_124            <NA>
#>   3:               <NA> comp_id_125            <NA>
#>   4:               <NA> comp_id_126            <NA>
#>   5:               <NA> comp_id_127            <NA>
#>   6:               <NA> comp_id_128            <NA>
#>   7:               <NA> comp_id_129            <NA>
#>   8:               <NA> comp_id_130            <NA>
#>   9:               <NA> comp_id_131            <NA>
#>  10:               <NA> comp_id_132            <NA>
#>  11:               <NA> comp_id_133            <NA>
#>  12:               <NA> comp_id_134            <NA>
#>  13:               <NA> comp_id_135            <NA>
#>  14:               <NA> comp_id_136            <NA>
#>  15:               <NA> comp_id_137            <NA>
#>  16:               <NA> comp_id_138            <NA>
#>  17:               <NA> comp_id_139            <NA>
#>  18:               <NA> comp_id_140            <NA>
#>  19:               <NA> comp_id_141            <NA>
#>  20:               <NA> comp_id_142            <NA>
#>  21:               <NA> comp_id_143            <NA>
#>  22:               <NA> comp_id_144            <NA>
#>  23:               <NA> comp_id_145            <NA>
#>  24:               <NA> comp_id_146            <NA>
#>  25:               <NA> comp_id_147            <NA>
#>  26:               <NA> comp_id_148            <NA>
#>  27:               <NA> comp_id_149            <NA>
#>  28:               <NA> comp_id_150            <NA>
#>  29:               <NA> comp_id_151            <NA>
#>  30:               <NA> comp_id_152            <NA>
#>  31:               <NA> comp_id_153            <NA>
#>  32:               <NA> comp_id_154            <NA>
#>  33:               <NA> comp_id_155            <NA>
#>  34:               <NA> comp_id_156            <NA>
#>  35:               <NA> comp_id_157            <NA>
#>  36:               <NA> comp_id_158            <NA>
#>  37:               <NA> comp_id_159            <NA>
#>  38:               <NA> comp_id_160            <NA>
#>  39:               <NA> comp_id_161            <NA>
#>  40:               <NA> comp_id_162            <NA>
#>  41:               <NA> comp_id_163            <NA>
#>  42:               <NA> comp_id_164            <NA>
#>  43:               <NA> comp_id_165            <NA>
#>  44:               <NA> comp_id_166            <NA>
#>  45:               <NA> comp_id_167            <NA>
#>  46:               <NA> comp_id_168            <NA>
#>  47:               <NA> comp_id_169            <NA>
#>  48:               <NA> comp_id_170            <NA>
#>  49:               <NA> comp_id_171            <NA>
#>  50:               <NA> comp_id_172            <NA>
#>  51:               <NA> comp_id_173            <NA>
#>  52:               <NA> comp_id_174            <NA>
#>  53:               <NA> comp_id_175            <NA>
#>  54:               <NA> comp_id_176            <NA>
#>  55:               <NA> comp_id_177            <NA>
#>  56:               <NA> comp_id_178            <NA>
#>  57:               <NA> comp_id_179            <NA>
#>  58:               <NA> comp_id_180            <NA>
#>  59:               <NA> comp_id_181            <NA>
#>  60:               <NA> comp_id_182            <NA>
#>  61:               <NA> comp_id_183            <NA>
#>  62:               <NA> comp_id_184            <NA>
#>  63:               <NA> comp_id_185            <NA>
#>  64:               <NA> comp_id_186            <NA>
#>  65:               <NA> comp_id_187            <NA>
#>  66:               <NA> comp_id_188            <NA>
#>  67:               <NA> comp_id_189            <NA>
#>  68:               <NA> comp_id_190            <NA>
#>  69:               <NA> comp_id_191            <NA>
#>  70:               <NA> comp_id_192            <NA>
#>  71:               <NA> comp_id_193            <NA>
#>  72:               <NA> comp_id_194            <NA>
#>  73:               <NA> comp_id_195            <NA>
#>  74:               <NA> comp_id_196            <NA>
#>  75:               <NA> comp_id_197            <NA>
#>  76:               <NA> comp_id_198            <NA>
#>  77:               <NA> comp_id_199            <NA>
#>  78:               <NA> comp_id_200            <NA>
#>  79:               <NA> comp_id_201            <NA>
#>  80:               <NA> comp_id_202            <NA>
#>  81:               <NA> comp_id_203            <NA>
#>  82:               <NA> comp_id_204            <NA>
#>  83:               <NA> comp_id_205            <NA>
#>  84:               <NA> comp_id_206            <NA>
#>  85:               <NA> comp_id_207            <NA>
#>  86:               <NA> comp_id_208            <NA>
#>  87:               <NA> comp_id_209            <NA>
#>  88:               <NA> comp_id_210            <NA>
#>  89:               <NA> comp_id_211            <NA>
#>  90:               <NA> comp_id_212            <NA>
#>  91:               <NA> comp_id_213            <NA>
#>  92:               <NA> comp_id_214            <NA>
#>  93:               <NA> comp_id_215            <NA>
#>  94:               <NA> comp_id_216            <NA>
#>  95:               <NA> comp_id_217            <NA>
#>  96:               <NA> comp_id_218            <NA>
#>  97:               <NA> comp_id_219            <NA>
#>  98:               <NA> comp_id_220            <NA>
#>  99:               <NA> comp_id_221            <NA>
#> 100:               <NA> comp_id_222            <NA>
#>      anno_split_synonym  feature_id derived_feature