Retrieve Feature Data

Extracts feature-level data from a Metabolites object for a specified data layer, with optional exclusions and subsetting.

Usage

get_features(metabolites, layer, feature_ids = NULL, as_df = FALSE)

Arguments

metabolites

A Metabolites object containing feature data.

layer

A character string specifying the data layer from which to extract features.

feature_ids

Optional. A vector of row names to extract. Defaults to NULL, meaning all rows are included.

as_df

Logical. Whether to return the result as a data.frame. Defaults to FALSE, returning a matrix.

Value

A data.table or data.frame (if as_df = TRUE) containing feature data for the specified layer.

Examples

file <- system.file("extdata", "metabolon_v1_example.xlsx", package = "metaboprep2")
metabolites <- Metabolites(project_name = "MYPROJECT",
                           format       = "metabolon_v1",
                           filepath     = file)
metabolites <- import_data(metabolites)
#> Warning: 86 unmapped features:
#>  - 123, 124, 125, 126, 127, 128, 129, 130, 131, 132, 133, 134, 135, 137, 138, 139, 140, 141, 142, 143, 145, 146, 147, 148, 149, 150, 151, 152, 153, 154, 155, 156, 157, 159, 160, 161, 162, 163, 164, 165, 166, 167, 168, 169, 170, 172, 173, 174, 175, 176, 177, 178, 183, 184, 185, 186, 187, 188, 190, 191, 193, 195, 196, 198, 199, 200, 201, 202, 204, 205, 206, 207, 208, 209, 210, 211, 213, 214, 215, 216, 217, 218, 219, 220, 221, 222
metabolites <- metabolite_qc(metabolites, source="raw")
#> 
#> ── Starting Metabolite QC Process ──────────────────────────────────────────────
#> ℹ Validating input parameters...
#> ✔ Input validation complete.
#> ℹ Sample & Feature Summary Statistics for raw data...
#> ✔ Sample & Feature Summary Statistics for raw data...
#> ℹ Copying raw data to new 'qcing' data layer...
#> ✔ qcing data ready for QC.
#> ℹ Analysing exclusions...
#> ℹ Applying exclusion of derived features...
#> ✔ Derived features exclusions assessment complete. (0 features excluded)
#> ℹ Applying exclusion of xenobiotic features...
#> ✔ Xenobiotic features exclusions assessment complete. (3 features excluded)
#> ℹ Assessing for extreme sample missingness >=80%...
#> ✔ Extreme sample missingness exclusions assessment complete. (0 samples excluded)
#> ℹ Assessing for extreme feature missingness >=80%...
#> ✔ Extreme feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Assessing for sample missingness at specified level of >=20%...
#> ✔ Sample missingness exclusions assessment complete. (2 samples excluded)
#> ℹ Assessing for feature missingness at specified level of >=20%...
#> ✔ Feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Calculating total peak abundance outliers at +/- 5 Sdev...
#> ✔ Total peak abundance outlier exclusions assessment complete. (0 samples excluded)
#> ℹ Running principal component outlier analysis at +/- 5 Sdev...
#> ℹ PCA outlier analysis complete.
#> ℹ Re-running sample summary using current exclusions...
#> ℹ QC sample summary complete.
#> ℹ Re-identify feature independence and PC outliers...
#> ℹ PCA outlier analysis 2 complete.
#> ℹ Creating final QC dataset...
#> ℹ Rerunning Sample & Feature Summary Statistics for QC data...
#> ✔ Sample & Feature Summary Statistics for QC data complete.
#> 
#> ── Metabolite QC Process Completed ─────────────────────────────────────────────
get_features(metabolites, layer = "post_qc")
#> Index: <derived_feature>
#>        feature_id comp_id metabolite_id platform                pathway   kegg
#>            <char>  <char>        <char>   <char>                 <char> <char>
#>   1: unmapped_123     123          met1     <NA>             xenobiotic   K100
#>   2: unmapped_124     124          met2     <NA>             xenobiotic   K100
#>   3: unmapped_125     125          met3     <NA>             xenobiotic   K100
#>   4: unmapped_126     126          met4     <NA>                  lipid   K200
#>   5: unmapped_127     127          met5     <NA>                  lipid   K200
#>   6: unmapped_128     128          met6     <NA>                  lipid   K200
#>   7: unmapped_129     129          met7     <NA>                  lipid   K200
#>   8: unmapped_130     130          met8     <NA>                  lipid   K200
#>   9: unmapped_131     131          met9     <NA>                  lipid   K200
#>  10: unmapped_132     132         met10     <NA>                  lipid   K200
#>  11: unmapped_133     133         met11     <NA>                  lipid   K200
#>  12: unmapped_134     134         met12     <NA>                  lipid   K200
#>  13: unmapped_135     135         met13     <NA>                  lipid   K200
#>  14:  chem_id_136     136         met14     <NA>                  lipid   K200
#>  15: unmapped_137     137         met15     <NA>                  lipid   K200
#>  16: unmapped_138     138         met16     <NA>                  lipid   K200
#>  17: unmapped_139     139         met17     <NA>                  lipid   K200
#>  18: unmapped_140     140         met18     <NA>                  lipid   K200
#>  19: unmapped_141     141         met19     <NA>                  lipid   K200
#>  20: unmapped_142     142         met20     <NA>                  lipid   K200
#>  21: unmapped_143     143         met21     <NA>                  lipid   K200
#>  22:  chem_id_144     144         met22     <NA>                  lipid   K200
#>  23: unmapped_145     145         met23     <NA>                  lipid   K200
#>  24: unmapped_146     146         met24     <NA>                  lipid   K200
#>  25: unmapped_147     147         met25     <NA>                  lipid   K200
#>  26: unmapped_148     148         met26     <NA>                  lipid   K200
#>  27: unmapped_149     149         met27     <NA>                  lipid   K200
#>  28: unmapped_150     150         met28     <NA>                  lipid   K200
#>  29: unmapped_151     151         met29     <NA>                  lipid   K200
#>  30: unmapped_152     152         met30     <NA>                  lipid   K200
#>  31: unmapped_153     153         met31     <NA>                  lipid   K200
#>  32: unmapped_154     154         met32     <NA>                  lipid   K200
#>  33: unmapped_155     155         met33     <NA>                  lipid   K200
#>  34: unmapped_156     156         met34     <NA>                  lipid   K200
#>  35: unmapped_157     157         met35     <NA>                  lipid   K200
#>  36:  chem_id_158     158         met36     <NA>                  lipid   K200
#>  37: unmapped_159     159         met37     <NA>                  lipid   K200
#>  38: unmapped_160     160         met38     <NA>                  lipid   K200
#>  39: unmapped_161     161         met39     <NA>                  lipid   K200
#>  40: unmapped_162     162         met40     <NA>                  lipid   K200
#>  41: unmapped_163     163         met41     <NA>                  lipid   K200
#>  42: unmapped_164     164         met42     <NA>                  lipid   K200
#>  43: unmapped_165     165         met43     <NA>                  lipid   K200
#>  44: unmapped_166     166         met44     <NA>                  lipid   K200
#>  45: unmapped_167     167         met45     <NA>                  lipid   K200
#>  46: unmapped_168     168         met46     <NA>                  lipid   K200
#>  47: unmapped_169     169         met47     <NA>                  lipid   K200
#>  48: unmapped_170     170         met48     <NA>                  lipid   K200
#>  49:  chem_id_171     171         met49     <NA>             amino acid   K300
#>  50: unmapped_172     172         met50     <NA>             amino acid   K300
#>  51: unmapped_173     173         met51     <NA>             amino acid   K300
#>  52: unmapped_174     174         met52     <NA>             amino acid   K300
#>  53: unmapped_175     175         met53     <NA>             amino acid   K300
#>  54: unmapped_176     176         met54     <NA>             amino acid   K300
#>  55: unmapped_177     177         met55     <NA>             amino acid   K300
#>  56: unmapped_178     178         met56     <NA>             amino acid   K300
#>  57:  chem_id_179     179         met57     <NA>             amino acid   K300
#>  58:  chem_id_180     180         met58     <NA>             amino acid   K300
#>  59:  chem_id_181     181         met59     <NA>             amino acid   K300
#>  60:  chem_id_182     182         met60     <NA>           carbohydrate   K200
#>  61: unmapped_183     183         met61     <NA>           carbohydrate   K200
#>  62: unmapped_184     184         met62     <NA>           carbohydrate   K200
#>  63: unmapped_185     185         met63     <NA>           carbohydrate   K200
#>  64: unmapped_186     186         met64     <NA>           carbohydrate   K200
#>  65: unmapped_187     187         met65     <NA>           carbohydrate   K200
#>  66: unmapped_188     188         met66     <NA>           carbohydrate   K200
#>  67:  chem_id_189     189         met67     <NA>           carbohydrate   K200
#>  68: unmapped_190     190         met68     <NA>           carbohydrate   K200
#>  69: unmapped_191     191         met69     <NA>           carbohydrate   K200
#>  70:  chem_id_192     192         met70     <NA> cofactors and vitamins   K400
#>  71: unmapped_193     193         met71     <NA> cofactors and vitamins   K400
#>  72:  chem_id_194     194         met72     <NA> cofactors and vitamins   K400
#>  73: unmapped_195     195         met73     <NA> cofactors and vitamins   K400
#>  74: unmapped_196     196         met74     <NA> cofactors and vitamins   K400
#>  75:  chem_id_197     197         met75     <NA> cofactors and vitamins   K400
#>  76: unmapped_198     198         met76     <NA> cofactors and vitamins   K400
#>  77: unmapped_199     199         met77     <NA> cofactors and vitamins   K400
#>  78: unmapped_200     200         met78     <NA> cofactors and vitamins   K400
#>  79: unmapped_201     201         met79     <NA> cofactors and vitamins   K400
#>  80: unmapped_202     202         met80     <NA> cofactors and vitamins   K400
#>  81:  chem_id_203     203         met81     <NA> cofactors and vitamins   K400
#>  82: unmapped_204     204         met82     <NA> cofactors and vitamins   K400
#>  83: unmapped_205     205         met83     <NA> cofactors and vitamins   K400
#>  84: unmapped_206     206         met84     <NA> cofactors and vitamins   K400
#>  85: unmapped_207     207         met85     <NA> cofactors and vitamins   K400
#>  86: unmapped_208     208         met86     <NA> cofactors and vitamins   K400
#>  87: unmapped_209     209         met87     <NA> cofactors and vitamins   K400
#>  88: unmapped_210     210         met88     <NA> cofactors and vitamins   K400
#>  89: unmapped_211     211         met89     <NA> cofactors and vitamins   K400
#>  90:  chem_id_212     212         met90     <NA>                 energy   K400
#>  91: unmapped_213     213         met91     <NA>                 energy   K400
#>  92: unmapped_214     214         met92     <NA>                 energy   K400
#>  93: unmapped_215     215         met93     <NA>                 energy   K400
#>  94: unmapped_216     216         met94     <NA>                 energy   K400
#>  95: unmapped_217     217         met95     <NA>                 energy   K400
#>  96: unmapped_218     218         met96     <NA>                 energy   K400
#>  97: unmapped_219     219         met97     <NA>                 energy   K400
#>  98: unmapped_220     220         met98     <NA>                 energy   K400
#>  99: unmapped_221     221         met99     <NA>                 energy   K400
#> 100: unmapped_222     222        met100     <NA>                 energy   K400
#>        feature_id comp_id metabolite_id platform                pathway   kegg
#>         hmdb derived_feature id_source                chemical_name
#>       <char>          <lgcl>    <char>                       <char>
#>   1:   HMDB1           FALSE      <NA>                         <NA>
#>   2:   HMDB2           FALSE      <NA>                         <NA>
#>   3:   HMDB3           FALSE      <NA>                         <NA>
#>   4:   HMDB4           FALSE      <NA>                         <NA>
#>   5:   HMDB5           FALSE      <NA>                         <NA>
#>   6:   HMDB6           FALSE      <NA>                         <NA>
#>   7:   HMDB7           FALSE      <NA>                         <NA>
#>   8:   HMDB8           FALSE      <NA>                         <NA>
#>   9:   HMDB9           FALSE      <NA>                         <NA>
#>  10:  HMDB10           FALSE      <NA>                         <NA>
#>  11:  HMDB11           FALSE      <NA>                         <NA>
#>  12:  HMDB12           FALSE      <NA>                         <NA>
#>  13:  HMDB13           FALSE      <NA>                         <NA>
#>  14:  HMDB14           FALSE   chem_id                      cholate
#>  15:  HMDB15           FALSE      <NA>                         <NA>
#>  16:  HMDB16           FALSE      <NA>                         <NA>
#>  17:  HMDB17           FALSE      <NA>                         <NA>
#>  18:  HMDB18           FALSE      <NA>                         <NA>
#>  19:  HMDB19           FALSE      <NA>                         <NA>
#>  20:  HMDB20           FALSE      <NA>                         <NA>
#>  21:  HMDB21           FALSE      <NA>                         <NA>
#>  22:  HMDB22           FALSE   chem_id       4-hydroxyphenylacetate
#>  23:  HMDB23           FALSE      <NA>                         <NA>
#>  24:  HMDB24           FALSE      <NA>                         <NA>
#>  25:  HMDB25           FALSE      <NA>                         <NA>
#>  26:  HMDB26           FALSE      <NA>                         <NA>
#>  27:  HMDB27           FALSE      <NA>                         <NA>
#>  28:  HMDB28           FALSE      <NA>                         <NA>
#>  29:  HMDB29           FALSE      <NA>                         <NA>
#>  30:  HMDB30           FALSE      <NA>                         <NA>
#>  31:  HMDB31           FALSE      <NA>                         <NA>
#>  32:  HMDB32           FALSE      <NA>                         <NA>
#>  33:  HMDB33           FALSE      <NA>                         <NA>
#>  34:  HMDB34           FALSE      <NA>                         <NA>
#>  35:  HMDB35           FALSE      <NA>                         <NA>
#>  36:  HMDB36           FALSE   chem_id           5,6-dihydrothymine
#>  37:  HMDB37           FALSE      <NA>                         <NA>
#>  38:  HMDB38           FALSE      <NA>                         <NA>
#>  39:  HMDB39           FALSE      <NA>                         <NA>
#>  40:  HMDB40           FALSE      <NA>                         <NA>
#>  41:  HMDB41           FALSE      <NA>                         <NA>
#>  42:  HMDB42           FALSE      <NA>                         <NA>
#>  43:  HMDB43           FALSE      <NA>                         <NA>
#>  44:  HMDB44           FALSE      <NA>                         <NA>
#>  45:  HMDB45           FALSE      <NA>                         <NA>
#>  46:  HMDB46           FALSE      <NA>                         <NA>
#>  47:  HMDB47           FALSE      <NA>                         <NA>
#>  48:  HMDB48           FALSE      <NA>                         <NA>
#>  49:  HMDB49           FALSE   chem_id                 hypoxanthine
#>  50:  HMDB50           FALSE      <NA>                         <NA>
#>  51:  HMDB51           FALSE      <NA>                         <NA>
#>  52:  HMDB52           FALSE      <NA>                         <NA>
#>  53:  HMDB53           FALSE      <NA>                         <NA>
#>  54:  HMDB54           FALSE      <NA>                         <NA>
#>  55:  HMDB55           FALSE      <NA>                         <NA>
#>  56:  HMDB56           FALSE      <NA>                         <NA>
#>  57:  HMDB57           FALSE   chem_id                  9,10-DiHOME
#>  58:  HMDB58           FALSE   chem_id           linoleate (18:2n6)
#>  59:  HMDB59           FALSE   chem_id               laurate (12:0)
#>  60:  HMDB60           FALSE   chem_id                  quinolinate
#>  61:  HMDB61           FALSE      <NA>                         <NA>
#>  62:  HMDB62           FALSE      <NA>                         <NA>
#>  63:  HMDB63           FALSE      <NA>                         <NA>
#>  64:  HMDB64           FALSE      <NA>                         <NA>
#>  65:  HMDB65           FALSE      <NA>                         <NA>
#>  66:  HMDB66           FALSE      <NA>                         <NA>
#>  67:  HMDB67           FALSE   chem_id     N6,N6,N6-trimethyllysine
#>  68:  HMDB68           FALSE      <NA>                         <NA>
#>  69:  HMDB69           FALSE      <NA>                         <NA>
#>  70:  HMDB70           FALSE   chem_id           N-acetylputrescine
#>  71:  HMDB71           FALSE      <NA>                         <NA>
#>  72:  HMDB72           FALSE   chem_id           N-formylmethionine
#>  73:  HMDB73           FALSE      <NA>                         <NA>
#>  74:  HMDB74           FALSE      <NA>                         <NA>
#>  75:  HMDB75           FALSE   chem_id S-adenosylhomocysteine (SAH)
#>  76:  HMDB76           FALSE      <NA>                         <NA>
#>  77:  HMDB77           FALSE      <NA>                         <NA>
#>  78:  HMDB78           FALSE      <NA>                         <NA>
#>  79:  HMDB79           FALSE      <NA>                         <NA>
#>  80:  HMDB80           FALSE      <NA>                         <NA>
#>  81:  HMDB81           FALSE   chem_id   glycolate (hydroxyacetate)
#>  82:  HMDB82           FALSE      <NA>                         <NA>
#>  83:  HMDB83           FALSE      <NA>                         <NA>
#>  84:  HMDB84           FALSE      <NA>                         <NA>
#>  85:  HMDB85           FALSE      <NA>                         <NA>
#>  86:  HMDB86           FALSE      <NA>                         <NA>
#>  87:  HMDB87           FALSE      <NA>                         <NA>
#>  88:  HMDB88           FALSE      <NA>                         <NA>
#>  89:  HMDB89           FALSE      <NA>                         <NA>
#>  90:  HMDB90           FALSE   chem_id  5-methylthioadenosine (MTA)
#>  91:  HMDB91           FALSE      <NA>                         <NA>
#>  92:  HMDB92           FALSE      <NA>                         <NA>
#>  93:  HMDB93           FALSE      <NA>                         <NA>
#>  94:  HMDB94           FALSE      <NA>                         <NA>
#>  95:  HMDB95           FALSE      <NA>                         <NA>
#>  96:  HMDB96           FALSE      <NA>                         <NA>
#>  97:  HMDB97           FALSE      <NA>                         <NA>
#>  98:  HMDB98           FALSE      <NA>                         <NA>
#>  99:  HMDB99           FALSE      <NA>                         <NA>
#> 100: HMDB100           FALSE      <NA>                         <NA>
#>         hmdb derived_feature id_source                chemical_name