Retrieve Feature Data

Extracts feature-level data from a Metabolites object for a specified data layer, with optional exclusions and subsetting.

Usage

get_features(metabolites, layer, feature_ids = NULL, as_df = FALSE)

Arguments

metabolites

A Metabolites object containing feature data.

layer

A character string specifying the data layer from which to extract features.

feature_ids

Optional. A vector of row names to extract. Defaults to NULL, meaning all rows are included.

as_df

Logical. Whether to return the result as a data.frame. Defaults to FALSE, returning a matrix.

Value

A data.table or data.frame (if as_df = TRUE) containing feature data for the specified layer.

Examples

file <- system.file("extdata", "metabolon_v1_example.xlsx", package = "metaboprep2")
metabolites <- Metabolites(project_name = "MYPROJECT",
                           format       = "metabolon_v1",
                           filepath     = file)
metabolites <- import_data(metabolites)
#> Warning: 86 unmapped features:
#>  - 123, 124, 125, 126, 127, 128, 129, 130, 131, 132, 133, 134, 135, 137, 138, 139, 140, 141, 142, 143, 145, 146, 147, 148, 149, 150, 151, 152, 153, 154, 155, 156, 157, 159, 160, 161, 162, 163, 164, 165, 166, 167, 168, 169, 170, 172, 173, 174, 175, 176, 177, 178, 183, 184, 185, 186, 187, 188, 190, 191, 193, 195, 196, 198, 199, 200, 201, 202, 204, 205, 206, 207, 208, 209, 210, 211, 213, 214, 215, 216, 217, 218, 219, 220, 221, 222
metabolites <- metabolite_qc(metabolites, source="raw")
#> 
#> ── Starting Metabolite QC Process ──────────────────────────────────────────────
#> ℹ Validating input parameters...
#> ✔ Input validation complete.
#> ℹ Sample & Feature Summary Statistics for raw data...
#> ✔ Sample & Feature Summary Statistics for raw data...
#> ℹ Copying raw data to new 'qcing' data layer...
#> ✔ qcing data ready for QC.
#> ℹ Analysing exclusions...
#> ℹ Applying exclusion of derived features...
#> ✔ Derived features exclusions assessment complete. (0 features excluded)
#> ℹ Applying exclusion of xenobiotic features...
#> ✔ Xenobiotic features exclusions assessment complete. (3 features excluded)
#> ℹ Assessing for extreme sample missingness >=80%...
#> ✔ Extreme sample missingness exclusions assessment complete. (0 samples excluded)
#> ℹ Assessing for extreme feature missingness >=80%...
#> ✔ Extreme feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Assessing for sample missingness at specified level of >=20%...
#> ✔ Sample missingness exclusions assessment complete. (2 samples excluded)
#> ℹ Assessing for feature missingness at specified level of >=20%...
#> ✔ Feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Calculating total peak abundance outliers at +/- 5 Sdev...
#> ✔ Total peak abundance outlier exclusions assessment complete. (0 samples excluded)
#> ℹ Running principal component outlier analysis at +/- 5 Sdev...
#> ℹ PCA outlier analysis complete.
#> ℹ Re-running sample summary using current exclusions...
#> ℹ QC sample summary complete.
#> ℹ Re-identify feature independence and PC outliers...
#> ℹ PCA outlier analysis 2 complete.
#> ℹ Creating final QC dataset...
#> ℹ Rerunning Sample & Feature Summary Statistics for QC data...
#> ✔ Sample & Feature Summary Statistics for QC data complete.
#> 
#> ── Metabolite QC Process Completed ─────────────────────────────────────────────
get_features(metabolites, layer = "post_qc")
#> Index: <derived_feature>
#>        feature_id metabolite_id comp_id platform                pathway   kegg
#>            <char>        <char>  <char>   <char>                 <char> <char>
#>   1: unmapped_123          met1     123     <NA>             xenobiotic   K100
#>   2: unmapped_124          met2     124     <NA>             xenobiotic   K100
#>   3: unmapped_125          met3     125     <NA>             xenobiotic   K100
#>   4: unmapped_126          met4     126     <NA>                  lipid   K200
#>   5: unmapped_127          met5     127     <NA>                  lipid   K200
#>   6: unmapped_128          met6     128     <NA>                  lipid   K200
#>   7: unmapped_129          met7     129     <NA>                  lipid   K200
#>   8: unmapped_130          met8     130     <NA>                  lipid   K200
#>   9: unmapped_131          met9     131     <NA>                  lipid   K200
#>  10: unmapped_132         met10     132     <NA>                  lipid   K200
#>  11: unmapped_133         met11     133     <NA>                  lipid   K200
#>  12: unmapped_134         met12     134     <NA>                  lipid   K200
#>  13: unmapped_135         met13     135     <NA>                  lipid   K200
#>  14:  chem_id_136         met14     136     <NA>                  lipid   K200
#>  15: unmapped_137         met15     137     <NA>                  lipid   K200
#>  16: unmapped_138         met16     138     <NA>                  lipid   K200
#>  17: unmapped_139         met17     139     <NA>                  lipid   K200
#>  18: unmapped_140         met18     140     <NA>                  lipid   K200
#>  19: unmapped_141         met19     141     <NA>                  lipid   K200
#>  20: unmapped_142         met20     142     <NA>                  lipid   K200
#>  21: unmapped_143         met21     143     <NA>                  lipid   K200
#>  22:  chem_id_144         met22     144     <NA>                  lipid   K200
#>  23: unmapped_145         met23     145     <NA>                  lipid   K200
#>  24: unmapped_146         met24     146     <NA>                  lipid   K200
#>  25: unmapped_147         met25     147     <NA>                  lipid   K200
#>  26: unmapped_148         met26     148     <NA>                  lipid   K200
#>  27: unmapped_149         met27     149     <NA>                  lipid   K200
#>  28: unmapped_150         met28     150     <NA>                  lipid   K200
#>  29: unmapped_151         met29     151     <NA>                  lipid   K200
#>  30: unmapped_152         met30     152     <NA>                  lipid   K200
#>  31: unmapped_153         met31     153     <NA>                  lipid   K200
#>  32: unmapped_154         met32     154     <NA>                  lipid   K200
#>  33: unmapped_155         met33     155     <NA>                  lipid   K200
#>  34: unmapped_156         met34     156     <NA>                  lipid   K200
#>  35: unmapped_157         met35     157     <NA>                  lipid   K200
#>  36:  chem_id_158         met36     158     <NA>                  lipid   K200
#>  37: unmapped_159         met37     159     <NA>                  lipid   K200
#>  38: unmapped_160         met38     160     <NA>                  lipid   K200
#>  39: unmapped_161         met39     161     <NA>                  lipid   K200
#>  40: unmapped_162         met40     162     <NA>                  lipid   K200
#>  41: unmapped_163         met41     163     <NA>                  lipid   K200
#>  42: unmapped_164         met42     164     <NA>                  lipid   K200
#>  43: unmapped_165         met43     165     <NA>                  lipid   K200
#>  44: unmapped_166         met44     166     <NA>                  lipid   K200
#>  45: unmapped_167         met45     167     <NA>                  lipid   K200
#>  46: unmapped_168         met46     168     <NA>                  lipid   K200
#>  47: unmapped_169         met47     169     <NA>                  lipid   K200
#>  48: unmapped_170         met48     170     <NA>                  lipid   K200
#>  49:  chem_id_171         met49     171     <NA>             amino acid   K300
#>  50: unmapped_172         met50     172     <NA>             amino acid   K300
#>  51: unmapped_173         met51     173     <NA>             amino acid   K300
#>  52: unmapped_174         met52     174     <NA>             amino acid   K300
#>  53: unmapped_175         met53     175     <NA>             amino acid   K300
#>  54: unmapped_176         met54     176     <NA>             amino acid   K300
#>  55: unmapped_177         met55     177     <NA>             amino acid   K300
#>  56: unmapped_178         met56     178     <NA>             amino acid   K300
#>  57:  chem_id_179         met57     179     <NA>             amino acid   K300
#>  58:  chem_id_180         met58     180     <NA>             amino acid   K300
#>  59:  chem_id_181         met59     181     <NA>             amino acid   K300
#>  60:  chem_id_182         met60     182     <NA>           carbohydrate   K200
#>  61: unmapped_183         met61     183     <NA>           carbohydrate   K200
#>  62: unmapped_184         met62     184     <NA>           carbohydrate   K200
#>  63: unmapped_185         met63     185     <NA>           carbohydrate   K200
#>  64: unmapped_186         met64     186     <NA>           carbohydrate   K200
#>  65: unmapped_187         met65     187     <NA>           carbohydrate   K200
#>  66: unmapped_188         met66     188     <NA>           carbohydrate   K200
#>  67:  chem_id_189         met67     189     <NA>           carbohydrate   K200
#>  68: unmapped_190         met68     190     <NA>           carbohydrate   K200
#>  69: unmapped_191         met69     191     <NA>           carbohydrate   K200
#>  70:  chem_id_192         met70     192     <NA> cofactors and vitamins   K400
#>  71: unmapped_193         met71     193     <NA> cofactors and vitamins   K400
#>  72:  chem_id_194         met72     194     <NA> cofactors and vitamins   K400
#>  73: unmapped_195         met73     195     <NA> cofactors and vitamins   K400
#>  74: unmapped_196         met74     196     <NA> cofactors and vitamins   K400
#>  75:  chem_id_197         met75     197     <NA> cofactors and vitamins   K400
#>  76: unmapped_198         met76     198     <NA> cofactors and vitamins   K400
#>  77: unmapped_199         met77     199     <NA> cofactors and vitamins   K400
#>  78: unmapped_200         met78     200     <NA> cofactors and vitamins   K400
#>  79: unmapped_201         met79     201     <NA> cofactors and vitamins   K400
#>  80: unmapped_202         met80     202     <NA> cofactors and vitamins   K400
#>  81:  chem_id_203         met81     203     <NA> cofactors and vitamins   K400
#>  82: unmapped_204         met82     204     <NA> cofactors and vitamins   K400
#>  83: unmapped_205         met83     205     <NA> cofactors and vitamins   K400
#>  84: unmapped_206         met84     206     <NA> cofactors and vitamins   K400
#>  85: unmapped_207         met85     207     <NA> cofactors and vitamins   K400
#>  86: unmapped_208         met86     208     <NA> cofactors and vitamins   K400
#>  87: unmapped_209         met87     209     <NA> cofactors and vitamins   K400
#>  88: unmapped_210         met88     210     <NA> cofactors and vitamins   K400
#>  89: unmapped_211         met89     211     <NA> cofactors and vitamins   K400
#>  90:  chem_id_212         met90     212     <NA>                 energy   K400
#>  91: unmapped_213         met91     213     <NA>                 energy   K400
#>  92: unmapped_214         met92     214     <NA>                 energy   K400
#>  93: unmapped_215         met93     215     <NA>                 energy   K400
#>  94: unmapped_216         met94     216     <NA>                 energy   K400
#>  95: unmapped_217         met95     217     <NA>                 energy   K400
#>  96: unmapped_218         met96     218     <NA>                 energy   K400
#>  97: unmapped_219         met97     219     <NA>                 energy   K400
#>  98: unmapped_220         met98     220     <NA>                 energy   K400
#>  99: unmapped_221         met99     221     <NA>                 energy   K400
#> 100: unmapped_222        met100     222     <NA>                 energy   K400
#>        feature_id metabolite_id comp_id platform                pathway   kegg
#>         hmdb derived_feature                chemical_name
#>       <char>          <lgcl>                       <char>
#>   1:   HMDB1           FALSE                         <NA>
#>   2:   HMDB2           FALSE                         <NA>
#>   3:   HMDB3           FALSE                         <NA>
#>   4:   HMDB4           FALSE                         <NA>
#>   5:   HMDB5           FALSE                         <NA>
#>   6:   HMDB6           FALSE                         <NA>
#>   7:   HMDB7           FALSE                         <NA>
#>   8:   HMDB8           FALSE                         <NA>
#>   9:   HMDB9           FALSE                         <NA>
#>  10:  HMDB10           FALSE                         <NA>
#>  11:  HMDB11           FALSE                         <NA>
#>  12:  HMDB12           FALSE                         <NA>
#>  13:  HMDB13           FALSE                         <NA>
#>  14:  HMDB14           FALSE                      cholate
#>  15:  HMDB15           FALSE                         <NA>
#>  16:  HMDB16           FALSE                         <NA>
#>  17:  HMDB17           FALSE                         <NA>
#>  18:  HMDB18           FALSE                         <NA>
#>  19:  HMDB19           FALSE                         <NA>
#>  20:  HMDB20           FALSE                         <NA>
#>  21:  HMDB21           FALSE                         <NA>
#>  22:  HMDB22           FALSE       4-hydroxyphenylacetate
#>  23:  HMDB23           FALSE                         <NA>
#>  24:  HMDB24           FALSE                         <NA>
#>  25:  HMDB25           FALSE                         <NA>
#>  26:  HMDB26           FALSE                         <NA>
#>  27:  HMDB27           FALSE                         <NA>
#>  28:  HMDB28           FALSE                         <NA>
#>  29:  HMDB29           FALSE                         <NA>
#>  30:  HMDB30           FALSE                         <NA>
#>  31:  HMDB31           FALSE                         <NA>
#>  32:  HMDB32           FALSE                         <NA>
#>  33:  HMDB33           FALSE                         <NA>
#>  34:  HMDB34           FALSE                         <NA>
#>  35:  HMDB35           FALSE                         <NA>
#>  36:  HMDB36           FALSE           5,6-dihydrothymine
#>  37:  HMDB37           FALSE                         <NA>
#>  38:  HMDB38           FALSE                         <NA>
#>  39:  HMDB39           FALSE                         <NA>
#>  40:  HMDB40           FALSE                         <NA>
#>  41:  HMDB41           FALSE                         <NA>
#>  42:  HMDB42           FALSE                         <NA>
#>  43:  HMDB43           FALSE                         <NA>
#>  44:  HMDB44           FALSE                         <NA>
#>  45:  HMDB45           FALSE                         <NA>
#>  46:  HMDB46           FALSE                         <NA>
#>  47:  HMDB47           FALSE                         <NA>
#>  48:  HMDB48           FALSE                         <NA>
#>  49:  HMDB49           FALSE                 hypoxanthine
#>  50:  HMDB50           FALSE                         <NA>
#>  51:  HMDB51           FALSE                         <NA>
#>  52:  HMDB52           FALSE                         <NA>
#>  53:  HMDB53           FALSE                         <NA>
#>  54:  HMDB54           FALSE                         <NA>
#>  55:  HMDB55           FALSE                         <NA>
#>  56:  HMDB56           FALSE                         <NA>
#>  57:  HMDB57           FALSE                  9,10-DiHOME
#>  58:  HMDB58           FALSE           linoleate (18:2n6)
#>  59:  HMDB59           FALSE               laurate (12:0)
#>  60:  HMDB60           FALSE                  quinolinate
#>  61:  HMDB61           FALSE                         <NA>
#>  62:  HMDB62           FALSE                         <NA>
#>  63:  HMDB63           FALSE                         <NA>
#>  64:  HMDB64           FALSE                         <NA>
#>  65:  HMDB65           FALSE                         <NA>
#>  66:  HMDB66           FALSE                         <NA>
#>  67:  HMDB67           FALSE     N6,N6,N6-trimethyllysine
#>  68:  HMDB68           FALSE                         <NA>
#>  69:  HMDB69           FALSE                         <NA>
#>  70:  HMDB70           FALSE           N-acetylputrescine
#>  71:  HMDB71           FALSE                         <NA>
#>  72:  HMDB72           FALSE           N-formylmethionine
#>  73:  HMDB73           FALSE                         <NA>
#>  74:  HMDB74           FALSE                         <NA>
#>  75:  HMDB75           FALSE S-adenosylhomocysteine (SAH)
#>  76:  HMDB76           FALSE                         <NA>
#>  77:  HMDB77           FALSE                         <NA>
#>  78:  HMDB78           FALSE                         <NA>
#>  79:  HMDB79           FALSE                         <NA>
#>  80:  HMDB80           FALSE                         <NA>
#>  81:  HMDB81           FALSE   glycolate (hydroxyacetate)
#>  82:  HMDB82           FALSE                         <NA>
#>  83:  HMDB83           FALSE                         <NA>
#>  84:  HMDB84           FALSE                         <NA>
#>  85:  HMDB85           FALSE                         <NA>
#>  86:  HMDB86           FALSE                         <NA>
#>  87:  HMDB87           FALSE                         <NA>
#>  88:  HMDB88           FALSE                         <NA>
#>  89:  HMDB89           FALSE                         <NA>
#>  90:  HMDB90           FALSE  5-methylthioadenosine (MTA)
#>  91:  HMDB91           FALSE                         <NA>
#>  92:  HMDB92           FALSE                         <NA>
#>  93:  HMDB93           FALSE                         <NA>
#>  94:  HMDB94           FALSE                         <NA>
#>  95:  HMDB95           FALSE                         <NA>
#>  96:  HMDB96           FALSE                         <NA>
#>  97:  HMDB97           FALSE                         <NA>
#>  98:  HMDB98           FALSE                         <NA>
#>  99:  HMDB99           FALSE                         <NA>
#> 100: HMDB100           FALSE                         <NA>
#>         hmdb derived_feature                chemical_name