Retrieve Principal Components

Extracts the principal components (PCs) from a Metabolites object for a specified data layer, with optional exclusions and subsetting.

Usage

get_pcs(metabolites, layer, sample_ids = NULL, drop = FALSE, as_df = FALSE)

Arguments

metabolites

A Metabolites object containing the principal component data.

layer

A character string specifying the data layer from which to extract the PCs.

sample_ids

Optional. A vector of column names to extract. Defaults to NULL, meaning all columns are included.

drop

Logical. Whether to drop completely NA rows (i.e. excluded)

as_df

Logical. Whether to return the result as a data.frame. Defaults to FALSE, returning a matrix.

Value

A matrix or data.frame (if as_df = TRUE) containing the principal components.

Examples

file <- system.file("extdata", "metabolon_v1_example.xlsx", package = "metaboprep2")
metabolites <- Metabolites(project_name = "MYPROJECT",
                           format       = "metabolon_v1",
                           filepath     = file)
metabolites <- import_data(metabolites)
#> Warning: 86 unmapped features:
#>  - 123, 124, 125, 126, 127, 128, 129, 130, 131, 132, 133, 134, 135, 137, 138, 139, 140, 141, 142, 143, 145, 146, 147, 148, 149, 150, 151, 152, 153, 154, 155, 156, 157, 159, 160, 161, 162, 163, 164, 165, 166, 167, 168, 169, 170, 172, 173, 174, 175, 176, 177, 178, 183, 184, 185, 186, 187, 188, 190, 191, 193, 195, 196, 198, 199, 200, 201, 202, 204, 205, 206, 207, 208, 209, 210, 211, 213, 214, 215, 216, 217, 218, 219, 220, 221, 222
metabolites <- metabolite_qc(metabolites, source="raw")
#> 
#> ── Starting Metabolite QC Process ──────────────────────────────────────────────
#> ℹ Validating input parameters...
#> ✔ Input validation complete.
#> ℹ Sample & Feature Summary Statistics for raw data...
#> ✔ Sample & Feature Summary Statistics for raw data...
#> ℹ Copying raw data to new 'qcing' data layer...
#> ✔ qcing data ready for QC.
#> ℹ Analysing exclusions...
#> ℹ Applying exclusion of derived features...
#> ✔ Derived features exclusions assessment complete. (0 features excluded)
#> ℹ Applying exclusion of xenobiotic features...
#> ✔ Xenobiotic features exclusions assessment complete. (3 features excluded)
#> ℹ Assessing for extreme sample missingness >=80%...
#> ✔ Extreme sample missingness exclusions assessment complete. (0 samples excluded)
#> ℹ Assessing for extreme feature missingness >=80%...
#> ✔ Extreme feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Assessing for sample missingness at specified level of >=20%...
#> ✔ Sample missingness exclusions assessment complete. (2 samples excluded)
#> ℹ Assessing for feature missingness at specified level of >=20%...
#> ✔ Feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Calculating total peak abundance outliers at +/- 5 Sdev...
#> ✔ Total peak abundance outlier exclusions assessment complete. (0 samples excluded)
#> ℹ Running data outlier analysis at +/- 5 Sdev...
#> ℹ Data outlier analysis complete.
#> ℹ Re-running sample summary using current exclusions...
#> ℹ QC sample summary complete.
#> ℹ Re-identify feature independence and PC outliers...
#> ℹ PCA outlier analysis 2 complete.
#> ℹ Creating final QC dataset...
#> ℹ Rerunning Sample & Feature Summary Statistics for QC data...
#> ✔ Sample & Feature Summary Statistics for QC data complete.
#> 
#> ── Metabolite QC Process Completed ─────────────────────────────────────────────
get_pcs(metabolites, layer = "post_qc")
#>                 pc1         pc2          pc3          pc4          pc5
#> ind1   -12.61227940 -2.49340731 -1.934802715 -3.459430584  3.383695442
#> ind2   -11.35663338  0.68815580  1.681522028 -0.912322414 -3.338356891
#> ind3    -0.23159022 -4.76180717  0.936182189 -2.917076680 -1.264524262
#> ind4    -1.28823552 -2.17582108  0.197168884  0.117496280  1.985268185
#> ind5    -2.81537623 -1.76631934 -1.352481945 -3.151306779 -1.134557414
#> ind6     0.12104751 -2.91565037  0.718993676  2.739419752  0.095539046
#> ind7     0.26363723 -0.72370733 -0.674300759  1.294511146 -1.393704168
#> ind8    -0.10571832 -0.75895831 -3.400384636 -1.798072913  0.083047443
#> ind9             NA          NA           NA           NA           NA
#> ind10    0.05993891  0.97301249 -1.089551196  1.762482998  1.408038626
#> ind11    0.05608374  1.53602419 -0.049093190  0.009824096 -0.766804766
#> ind12    2.36202089 -0.44276494  0.740121884  0.425062140 -1.096785530
#> ind13    2.06838626 -3.97293149 -1.982204291  1.112276356  0.234459259
#> ind14   -1.52239024 -1.59948374  0.945060352  0.031207443  0.109918836
#> ind15    0.22347660 -3.87593956  0.502771529  2.700294649 -0.207303440
#> ind16    3.67619432 -3.02101609  2.487263196  2.545473877  1.808427725
#> ind17    1.80647377  3.57013966 -1.290430229 -0.012355907 -2.627878038
#> ind18   -1.51469921 -0.61146571 -0.567151533  3.055356875  1.030123941
#> ind19    0.23889472 -3.65579770 -1.928763972  2.161142146 -0.377281881
#> ind20    0.18049194 -2.26926044 -1.688342534  0.725092750  2.071568338
#> ind21   -2.82607638 -0.68622566  0.836300658  1.584050821  2.729419215
#> ind22   -0.94474120 -1.79764990  0.285240774  1.518434927  1.063598841
#> ind23   -1.82825603 -1.98223281  0.003994129  3.124773779 -0.267605854
#> ind24    0.98936136 -2.02346020 -2.458364018  0.901031570  1.724968127
#> ind25   -1.90962264  0.71068991  0.744912989  1.632047438  0.304871359
#> ind26    0.60025503  0.49507768  0.150181995  1.670685666  0.171525738
#> ind27   -1.38259117 -3.43506760 -3.707303646 -0.611181242  1.724360281
#> ind28   -0.26603472 -4.84415500  1.451247407 -0.379003787 -1.152047205
#> ind29    0.92981289 -0.86379046 -1.589776399 -0.506799470 -2.197374778
#> ind30   -2.63080120  0.49582998  1.234705112  0.922952651  0.155643558
#> ind31   -1.39364738 -2.79458715  0.833881050 -0.321430951 -1.265175026
#> ind32    1.64979490 -2.42891232  2.667698726 -0.910949416 -2.009796298
#> ind33    0.52796706 -2.56192340 -0.972188885 -0.672598710  0.503097021
#> ind34   -0.71114212  0.41233762 -0.178442157  1.689543649  1.600850362
#> ind35    1.08651889 -0.24726082  0.079172562  2.785484356 -4.512951859
#> ind36   -2.13792980 -0.26172982 -0.203905671  2.450340055 -1.935260027
#> ind37    0.64726602 -5.90629210  2.896312795 -1.315573648  0.876080182
#> ind38    2.70226734 -3.25503194 -2.155778241 -0.413729459 -1.226759071
#> ind39    2.96765831 -1.46779742  3.069202276 -2.219135359  3.738514533
#> ind40    3.27330384 -2.69103000 -1.849828144 -2.864102300 -0.623242557
#> ind41    2.68327675 -2.56756578 -2.056129847 -0.919728705 -3.865498681
#> ind42    0.42865197 -3.34578502 -0.993097308  0.880594345  0.607468563
#> ind43   -0.45210286 -3.65991546 -1.402350723 -3.066152557  0.232217998
#> ind44   -1.90037900 -3.42033416  1.147909805 -0.396049313 -0.962187351
#> ind45   -0.97570497 -1.66766644  0.869136412  2.268087942 -0.969233716
#> ind46    0.36017684 -2.03848391 -3.719738835 -0.765050030 -0.999091418
#> ind47   -2.14006078 -3.42307033 -2.094671721 -0.895916824 -1.175321507
#> ind48    2.70969381 -3.71198995  0.792891051  1.477321012  2.079333632
#> ind49   -0.53152847 -1.53710028 -2.442345831 -1.266063045  1.176281342
#> ind50   -0.56515826  1.87598731 -0.691970907 -1.481768097  3.872187872
#> ind51   -4.16932811  3.17494555  1.977355881  1.803018931  0.903096137
#> ind52   -1.40458982  0.35779036  1.713339779 -0.639837829 -1.121397824
#> ind53   -2.11717626  5.06732861 -2.322308317  0.200826319  0.641232649
#> ind54   -4.62071886  3.06304961  0.670282177 -1.589487152 -0.633187903
#> ind55   -5.04277733  0.00445237  1.778689948  1.613368531 -0.601690517
#> ind56   -1.49499053  0.65750048  3.160815177  1.493311925  0.909779198
#> ind57   -1.20688995  1.55913813 -0.733247648 -1.429864617 -3.246558397
#> ind58   -1.66998503  0.63508239 -2.826529049 -0.187906722 -0.223808200
#> ind59            NA          NA           NA           NA           NA
#> ind60   -0.30149038  3.68825740  1.527112202 -0.263655668  3.230898480
#> ind61   -3.01698270  3.41559893 -2.085442926  1.797824602 -0.776064228
#> ind62   -2.18001916 -0.13291560  0.750761527 -3.596338027  1.828559568
#> ind63   -2.55941483  2.41483520 -1.412220450 -0.065303426 -2.009608594
#> ind64   -1.31835172  2.44928464  1.377050522 -3.210355682 -0.920127162
#> ind65   -1.54143713  0.91489067 -0.625658810 -1.789283551 -0.175569503
#> ind66   -6.23816553  1.06870050  1.624142875 -1.400562546 -0.002523621
#> ind67   -0.26133338  4.73012459 -1.742871596  2.241496109  1.476324660
#> ind68   -4.29978452  2.19572691  1.059928194  1.889558337 -0.319341008
#> ind69   -0.61200973  3.37903967 -4.232475747 -1.981194154  1.376018582
#> ind70   -3.96220240  4.74439167  1.738611286  0.216142305 -2.108062298
#> ind71   -2.04711541  3.91235542 -1.397686871 -3.055615338  1.000359439
#> ind72   -5.13929491 -1.24743273  2.812675254  1.288508513 -1.417203782
#> ind73   -1.44278775  1.73831355 -0.261457004  2.862936252  3.272727962
#> ind74   -0.62164625  2.99725723 -1.467384257 -0.223466608  2.421852531
#> ind75   -2.84324891  1.80543383  4.120363342 -0.921973264 -0.630671541
#> ind76    4.58585201  2.98039733  1.322018654 -0.083876068 -2.116462517
#> ind77    1.28549405  2.50443163  3.158848434 -2.299257878 -1.313801017
#> ind78    6.06119943  5.81700238 -1.485612238  0.362970565  2.229707330
#> ind79    1.72024865  2.13267294  1.580142449  1.038736348 -2.863408759
#> ind80    3.35393779 -0.74110960 -3.838359322  0.387824210  0.363787650
#> ind81    1.39172743  1.24864146  1.036189122 -2.169957448  0.227006466
#> ind82    3.82722735 -1.33947505  0.565165551 -1.025140634 -0.249695137
#> ind83    3.29783435 -1.13859484 -0.445651475  0.859947994  0.325468489
#> ind84    3.89206944  3.96267746  0.758700423  1.595827591 -1.713935279
#> ind85    2.20856425 -1.88415077 -2.309839909  2.277323222 -0.408454853
#> ind86    1.76440482  1.26008184  2.022095275 -0.538803245  0.961832312
#> ind87    5.61315399  3.53392837  0.697990078 -2.822366261 -1.937043364
#> ind88    4.31251804  1.48624406 -0.171372966  1.005421263  0.275600901
#> ind89    2.06840872  2.62696193 -1.297765712  2.326982376 -1.099816565
#> ind90    3.79254818  0.47379983  0.109344294 -0.785345548  0.911158136
#> ind91    2.15229466  2.62105630 -1.344835058 -0.330453370 -0.519005940
#> ind92    5.91183926 -2.13804135  4.248210291 -4.781305450  1.598036166
#> ind93    6.35984083  1.15149007  1.841600611 -2.660992892 -0.422748630
#> ind94    2.66716110 -1.84800155  3.646636756  2.389335499  0.853179832
#> ind95    2.38416527  0.25558414  3.060530108  0.517923962  2.172828139
#> ind96    1.73632224  3.16259617  1.750703968 -0.102691309  0.729370372
#> ind97    2.47039853  1.66855286 -3.074465891 -0.720245932 -2.977120224
#> ind98    3.36612733  4.50820524 -0.331939663 -0.121620196  1.962863864
#> ind99    3.42953301 -1.30788585  1.338562181  0.232703706 -0.910640058
#> ind100   1.88891851  3.31591953 -0.839213599  0.065721727  1.644494329
#>                pc6         pc7          pc8          pc9         pc10
#> ind1    0.19654921  1.80117032 -2.622562448 -0.008317296  3.662380875
#> ind2    1.66431265  1.26873881  0.243208300 -0.352981944  2.670483990
#> ind3   -0.73537124  0.79042731 -1.723739110 -2.562634085 -0.683886969
#> ind4    0.10476686 -2.71326586 -0.429900837 -0.235378734  0.819890223
#> ind5    0.35300537 -0.60330822  1.331463335 -0.021650331 -2.791489166
#> ind6    1.10506878 -0.22160366 -1.109372014  0.954008822  0.904315500
#> ind7    1.70687139  0.57101242 -1.230807549 -1.952125129  0.569910344
#> ind8   -0.10424862 -1.26853621 -1.146918848  0.816025161 -2.024457324
#> ind9            NA          NA           NA           NA           NA
#> ind10  -0.48481461  0.65416232  0.999876555  1.831362639 -0.880534851
#> ind11   1.78677756  1.00604282 -0.053696337  2.581517038  2.341471205
#> ind12  -0.73959707 -0.60091037  2.482614136 -0.679676516 -0.343543031
#> ind13   1.10882462  0.52371614 -0.178736955  1.067775968  1.338328081
#> ind14  -0.67980523 -0.37022853  0.036944390  0.840269534 -1.498432090
#> ind15   0.79849967  0.21407382  2.014438396 -0.961689170 -0.849877444
#> ind16  -2.16466387  0.70893789 -1.198258663 -0.918429530 -0.429934876
#> ind17  -2.69055046  1.13564699  1.591053220 -2.211400194  1.289434290
#> ind18   0.94182134  2.15437720  0.768492853 -0.719499425 -0.304817850
#> ind19  -0.51995979  0.21737986 -0.418526516 -2.008555578  1.165490296
#> ind20   1.05281648  1.00770374 -0.369041111  1.410177912 -0.171488086
#> ind21   1.15855432 -2.88791464 -1.078376619  0.700890506  0.513986815
#> ind22   0.61166187 -0.93036322  2.316832408  2.293146059 -0.285763291
#> ind23  -0.10635239 -0.06716427 -0.115961974 -0.973483671  2.065931435
#> ind24   2.49845690  1.00575980 -1.090561074  1.565321259 -1.407525515
#> ind25   3.14975325  0.54892467 -0.987257207 -0.341754537 -0.877993319
#> ind26  -2.14360078  0.31202354  1.117361446 -0.887861167  0.954315640
#> ind27  -0.69276381 -1.13043259  0.261154334  0.493669180 -3.623035435
#> ind28   0.32921849  0.04654728 -0.428875921  1.488001676 -1.936074201
#> ind29  -1.83231362 -0.43454311  0.431453569 -1.255614932 -0.447505751
#> ind30   2.33803562 -2.06428527  2.118111250 -2.663996988 -2.493425404
#> ind31  -1.58549629 -0.65025270  1.478438723  1.982612556  1.813242160
#> ind32  -0.96077791  0.14735734  1.082796134 -0.520434034  0.606034795
#> ind33   1.07435491 -0.08531323 -1.216411367  0.199044036 -2.029754332
#> ind34   0.93835612 -0.97980369 -0.858026019 -1.391637901 -1.153168942
#> ind35   1.12791966  6.35854650 -0.662013187  2.381023964  0.110288759
#> ind36  -0.92769189 -0.66102927 -0.011435780  1.470794782  1.220505122
#> ind37   2.41974210  2.96326467  2.375773207 -0.460855243  1.128062396
#> ind38  -0.18368275  3.18637036 -1.340060728  0.723625559  1.195019963
#> ind39   0.51749970  0.39951794  1.819928158  1.887254905 -0.234737870
#> ind40  -2.97528887 -1.32618377 -1.714912252  1.819964869 -1.192266901
#> ind41  -0.41723031  1.04012098 -1.394763211 -1.241341204 -2.821615881
#> ind42  -0.16765296 -0.39444159  0.956365124 -0.278536173  0.198230488
#> ind43  -1.33115017 -2.45296069  1.243250097  0.070865098  1.932285291
#> ind44  -0.45233478 -1.25188811  0.989480404  2.541277572 -0.449608034
#> ind45  -1.48460300  1.07085126  0.775876132 -0.927440184 -1.411962705
#> ind46  -0.42577631 -1.48697231  0.281293122 -1.602421045  2.127639192
#> ind47   1.94034687 -0.82816463  1.603474001  0.135417409 -0.065293191
#> ind48  -0.84898483  2.05760554  1.279830506  1.877086145 -0.596927606
#> ind49   1.37790054 -0.89455418  1.094808359 -0.687612037  0.397303259
#> ind50  -4.62337886 -0.48962079 -0.490732107 -1.144762059  0.315931259
#> ind51  -1.44435103 -2.12278300 -3.501219649  1.400586376 -1.998460576
#> ind52   0.03739242 -0.60857745  0.185395806  0.383793526  0.985486558
#> ind53  -0.29714647  2.64807692  0.469940391 -1.913670056  0.450421456
#> ind54   0.71910079  1.00676850 -1.565710686 -0.481152139 -2.331622530
#> ind55  -4.77553438  0.89120798  1.026383635  0.874221360 -2.631759128
#> ind56  -0.60991697  2.32601569 -1.925689264  0.546025985 -0.225555932
#> ind57  -0.13892721  1.96256833  0.905240872  2.286717336 -0.677891714
#> ind58  -1.37679458 -1.23305428 -0.989770074  0.833391028  0.983294993
#> ind59           NA          NA           NA           NA           NA
#> ind60  -1.13851660  2.22958861 -0.005695987  2.044326007 -1.038315293
#> ind61  -0.01508511  2.30837971 -0.609715079 -1.499895162  0.296862040
#> ind62   2.24779938 -0.99819427 -2.404902084  1.343293536  0.586163467
#> ind63   1.04425539 -0.93326947 -2.103673536 -0.135844655 -0.256166724
#> ind64   0.86037239  2.06937708  3.273591302  1.290971990 -0.463112199
#> ind65  -0.13725956 -0.57282895 -0.106678509 -1.909070364 -3.042392254
#> ind66  -0.83215892 -0.54501837 -0.754989450  0.513467172  0.685703287
#> ind67   1.49749430  0.34557147  1.704937834 -1.866656676 -1.035436884
#> ind68   2.78588480 -1.97716342  2.540676942 -1.321635968 -2.838359209
#> ind69   1.04831583  0.75178507 -0.545043879  2.733497990 -0.006306743
#> ind70   0.06314145 -2.37426158  0.694816829  1.485530074 -0.392919441
#> ind71   0.02127935  2.31868092  2.114923794 -2.237605491 -0.389190468
#> ind72  -1.67243426 -2.47634311 -0.358986702 -1.125384779  2.898968569
#> ind73  -2.44359267  0.76105158  3.292067007  0.063789986 -0.500053900
#> ind74  -2.18766881  2.41208981  0.466993157 -2.707386778  0.021176953
#> ind75   0.16075725  0.06223826 -0.596236293 -1.123056952  0.724624861
#> ind76   0.75869817 -1.71384671  4.254462429  1.374750160  0.293678508
#> ind77   1.83137487 -1.58440263 -1.153257418  1.243069864 -1.354726903
#> ind78   3.50015773 -0.07063447 -0.268764254 -0.354961705  0.833279930
#> ind79   0.62095054 -3.01966227 -2.438346056 -0.321262093 -0.531490288
#> ind80   1.77686033 -0.26030610  1.090472561  0.584609846  0.485567578
#> ind81  -1.81613212 -1.17299327  2.567096901 -3.219726532  0.879016342
#> ind82   2.39479733  0.35246213 -1.271798938 -2.062187317 -0.948833360
#> ind83   1.40826573 -0.07488597 -1.008888774 -0.954757570  1.344300619
#> ind84  -0.19359848 -1.17118375 -0.437536928  1.651690786  0.265699789
#> ind85  -2.48685954 -3.49036742 -0.337793168  0.494772172  0.072210799
#> ind86  -0.09789832  0.94403435 -1.938792539 -0.054267495 -0.347458586
#> ind87  -0.76992678  1.08428901 -1.154930192  1.834041843  0.245557983
#> ind88   1.85389374 -1.49535733 -0.724834296  0.039488928  2.067258236
#> ind89  -1.28586451  0.87503219 -1.161273663  0.745993994  0.637091426
#> ind90  -0.50914413  1.04052561 -3.327672232 -2.805293817 -1.077429011
#> ind91  -2.49286022 -0.62238602 -0.231118519  1.834193291  2.626509462
#> ind92   1.18150552  0.71516553  2.238897027 -1.532954211  1.954824008
#> ind93  -2.12316300  0.55434436 -2.135493838 -0.158938211 -1.809888632
#> ind94   1.68728388 -0.82824413 -2.382995078 -1.786000916  0.828700365
#> ind95  -1.24692447  1.18690268 -0.215205097  0.939873004 -0.294330669
#> ind96   1.80473534 -1.28099888 -0.722538861 -0.482902752  0.201009075
#> ind97   1.27473397 -1.69836094  0.840781550 -1.188790532  0.929311290
#> ind98  -0.45879177 -0.87165399 -0.971644182  0.295789316  2.762567883
#> ind99  -1.28971178 -0.02680687 -1.202732533 -1.719062628  1.344699823
#> ind100  0.23818739 -2.01910976  2.063879393  2.003535682  0.452353831
#>        pc1_3_sd_outlier pc2_3_sd_outlier pc1_4_sd_outlier pc2_4_sd_outlier
#> ind1                  1                0                0                0
#> ind2                  1                0                0                0
#> ind3                  0                0                0                0
#> ind4                  0                0                0                0
#> ind5                  0                0                0                0
#> ind6                  0                0                0                0
#> ind7                  0                0                0                0
#> ind8                  0                0                0                0
#> ind9                 NA               NA               NA               NA
#> ind10                 0                0                0                0
#> ind11                 0                0                0                0
#> ind12                 0                0                0                0
#> ind13                 0                0                0                0
#> ind14                 0                0                0                0
#> ind15                 0                0                0                0
#> ind16                 0                0                0                0
#> ind17                 0                0                0                0
#> ind18                 0                0                0                0
#> ind19                 0                0                0                0
#> ind20                 0                0                0                0
#> ind21                 0                0                0                0
#> ind22                 0                0                0                0
#> ind23                 0                0                0                0
#> ind24                 0                0                0                0
#> ind25                 0                0                0                0
#> ind26                 0                0                0                0
#> ind27                 0                0                0                0
#> ind28                 0                0                0                0
#> ind29                 0                0                0                0
#> ind30                 0                0                0                0
#> ind31                 0                0                0                0
#> ind32                 0                0                0                0
#> ind33                 0                0                0                0
#> ind34                 0                0                0                0
#> ind35                 0                0                0                0
#> ind36                 0                0                0                0
#> ind37                 0                0                0                0
#> ind38                 0                0                0                0
#> ind39                 0                0                0                0
#> ind40                 0                0                0                0
#> ind41                 0                0                0                0
#> ind42                 0                0                0                0
#> ind43                 0                0                0                0
#> ind44                 0                0                0                0
#> ind45                 0                0                0                0
#> ind46                 0                0                0                0
#> ind47                 0                0                0                0
#> ind48                 0                0                0                0
#> ind49                 0                0                0                0
#> ind50                 0                0                0                0
#> ind51                 0                0                0                0
#> ind52                 0                0                0                0
#> ind53                 0                0                0                0
#> ind54                 0                0                0                0
#> ind55                 0                0                0                0
#> ind56                 0                0                0                0
#> ind57                 0                0                0                0
#> ind58                 0                0                0                0
#> ind59                NA               NA               NA               NA
#> ind60                 0                0                0                0
#> ind61                 0                0                0                0
#> ind62                 0                0                0                0
#> ind63                 0                0                0                0
#> ind64                 0                0                0                0
#> ind65                 0                0                0                0
#> ind66                 0                0                0                0
#> ind67                 0                0                0                0
#> ind68                 0                0                0                0
#> ind69                 0                0                0                0
#> ind70                 0                0                0                0
#> ind71                 0                0                0                0
#> ind72                 0                0                0                0
#> ind73                 0                0                0                0
#> ind74                 0                0                0                0
#> ind75                 0                0                0                0
#> ind76                 0                0                0                0
#> ind77                 0                0                0                0
#> ind78                 0                0                0                0
#> ind79                 0                0                0                0
#> ind80                 0                0                0                0
#> ind81                 0                0                0                0
#> ind82                 0                0                0                0
#> ind83                 0                0                0                0
#> ind84                 0                0                0                0
#> ind85                 0                0                0                0
#> ind86                 0                0                0                0
#> ind87                 0                0                0                0
#> ind88                 0                0                0                0
#> ind89                 0                0                0                0
#> ind90                 0                0                0                0
#> ind91                 0                0                0                0
#> ind92                 0                0                0                0
#> ind93                 0                0                0                0
#> ind94                 0                0                0                0
#> ind95                 0                0                0                0
#> ind96                 0                0                0                0
#> ind97                 0                0                0                0
#> ind98                 0                0                0                0
#> ind99                 0                0                0                0
#> ind100                0                0                0                0
#>        pc1_5_sd_outlier pc2_5_sd_outlier
#> ind1                  0                0
#> ind2                  0                0
#> ind3                  0                0
#> ind4                  0                0
#> ind5                  0                0
#> ind6                  0                0
#> ind7                  0                0
#> ind8                  0                0
#> ind9                 NA               NA
#> ind10                 0                0
#> ind11                 0                0
#> ind12                 0                0
#> ind13                 0                0
#> ind14                 0                0
#> ind15                 0                0
#> ind16                 0                0
#> ind17                 0                0
#> ind18                 0                0
#> ind19                 0                0
#> ind20                 0                0
#> ind21                 0                0
#> ind22                 0                0
#> ind23                 0                0
#> ind24                 0                0
#> ind25                 0                0
#> ind26                 0                0
#> ind27                 0                0
#> ind28                 0                0
#> ind29                 0                0
#> ind30                 0                0
#> ind31                 0                0
#> ind32                 0                0
#> ind33                 0                0
#> ind34                 0                0
#> ind35                 0                0
#> ind36                 0                0
#> ind37                 0                0
#> ind38                 0                0
#> ind39                 0                0
#> ind40                 0                0
#> ind41                 0                0
#> ind42                 0                0
#> ind43                 0                0
#> ind44                 0                0
#> ind45                 0                0
#> ind46                 0                0
#> ind47                 0                0
#> ind48                 0                0
#> ind49                 0                0
#> ind50                 0                0
#> ind51                 0                0
#> ind52                 0                0
#> ind53                 0                0
#> ind54                 0                0
#> ind55                 0                0
#> ind56                 0                0
#> ind57                 0                0
#> ind58                 0                0
#> ind59                NA               NA
#> ind60                 0                0
#> ind61                 0                0
#> ind62                 0                0
#> ind63                 0                0
#> ind64                 0                0
#> ind65                 0                0
#> ind66                 0                0
#> ind67                 0                0
#> ind68                 0                0
#> ind69                 0                0
#> ind70                 0                0
#> ind71                 0                0
#> ind72                 0                0
#> ind73                 0                0
#> ind74                 0                0
#> ind75                 0                0
#> ind76                 0                0
#> ind77                 0                0
#> ind78                 0                0
#> ind79                 0                0
#> ind80                 0                0
#> ind81                 0                0
#> ind82                 0                0
#> ind83                 0                0
#> ind84                 0                0
#> ind85                 0                0
#> ind86                 0                0
#> ind87                 0                0
#> ind88                 0                0
#> ind89                 0                0
#> ind90                 0                0
#> ind91                 0                0
#> ind92                 0                0
#> ind93                 0                0
#> ind94                 0                0
#> ind95                 0                0
#> ind96                 0                0
#> ind97                 0                0
#> ind98                 0                0
#> ind99                 0                0
#> ind100                0                0