Retrieve Probabilistic Principal Components

Extracts the probabilistic principal components (PCs) from a Metabolites object for a specified data layer, with optional exclusions applied.

Usage

get_prob_pcs(
  metabolites,
  layer,
  sample_ids = NULL,
  drop = FALSE,
  as_df = FALSE
)

Arguments

metabolites

A Metabolites object containing the probabilistic PC data.

layer

A character string specifying the data layer from which to extract the probabilistic PCs.

sample_ids

Optional. A vector of column names to extract. Defaults to NULL, meaning all columns are included.

drop

Logical. Whether to drop completely NA rows (i.e. excluded)

as_df

Logical. Whether to return the result as a data.frame. Defaults to FALSE, returning a matrix.

Value

A matrix or data.frame (if as_df = TRUE) containing the probabilistic principal components.

Examples

file <- system.file("extdata", "metabolon_v1_example.xlsx", package = "metaboprep2")
metabolites <- Metabolites(project_name = "MYPROJECT",
                           format       = "metabolon_v1",
                           filepath     = file)
metabolites <- import_data(metabolites)
#> Using cached MetaboAnalyst DB.
#> Attempting join [features$hmdb] ~ [annotation$hmdb_id] 
#> Attempting join [features$kegg] ~ [annotation$kegg_id] 
#> Attempting join [features$metabolite_id] ~ [annotation$name] 
#> Attempting join [features$comp_id] ~ [annotation$comp_id] 
#> Warning: 100 unannotated metabolites remain after merging with metabolomics database
metabolites <- metabolite_qc(metabolites, source="raw")
#> 
#> ── Starting Metabolite QC Process ──────────────────────────────────────────────
#> ℹ Validating input parameters...
#> ✔ Input validation complete.
#> ℹ Batch normalising raw data...
#> ✔ Normalised raw data.
#> ℹ Applying exclusion of xenobiotic features...
#> ✔ Xenobiotic features exclusions assessment complete. (3 features excluded)
#> ℹ Analysing exclusions...
#> ℹ Applying exclusion of derived features...
#> ✔ Derived features exclusions assessment complete. (0 features excluded)
#> ℹ Sample & Feature Summary Statistics for raw data...
#> ✔ Sample & Feature Summary Statistics for raw data...
#> ℹ Copying batch_normalised data to new 'qcing' data layer...
#> ✔ qcing data ready for QC.
#> ℹ Assessing for extreme sample missingness >=80%...
#> ✔ Extreme sample missingness exclusions assessment complete. (0 samples excluded)
#> ℹ Assessing for total feature missingness (100%)
#> ✔ Total feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Assessing for extreme feature missingness >=80%...
#> ✔ Extreme feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Assessing for sample missingness at specified level of >=20%...
#> ✔ Sample missingness exclusions assessment complete. (2 samples excluded)
#> ℹ Assessing for feature missingness at specified level of >=20%...
#> ✔ Feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Calculating total peak abundance outliers at +/- 5 Sdev...
#> ✔ Total peak abundance outlier exclusions assessment complete. (0 samples excluded)
#> ℹ Running data outlier analysis at +/- 5 Sdev...
#> ℹ Data outlier analysis complete.
#> ℹ Re-identify feature independence and PC outliers...
#> ℹ Re-running sample summary using current exclusions...
#> ℹ QC sample summary complete.
#> ℹ PCA outlier analysis 2 complete.
#> ℹ Creating final QC dataset...
#> ℹ Rerunning Sample & Feature Summary Statistics for QC data...
#> ✔ Sample & Feature Summary Statistics for QC data complete.
#> 
#> ── Metabolite QC Process Completed ─────────────────────────────────────────────
get_prob_pcs(metabolites, layer = "post_qc")
#>             prob_pc1    prob_pc2     prob_pc3     prob_pc4    prob_pc5
#> ind1              NA          NA           NA           NA          NA
#> ind2   -13.634850857 -4.47076606 -1.408076384 -1.394488619  4.03436707
#> ind3     3.124254537 -4.81168129 -1.862521036 -0.308189099  1.88118490
#> ind4    -0.862372813 -1.78298191 -0.679847761 -0.270280174 -3.04027556
#> ind5    -2.023253461 -1.17551943 -3.263693805 -2.262986992  1.28802973
#> ind6     1.104274396 -1.05053669  1.842029879  3.758248057  0.48926929
#> ind7    -0.525702000  1.20106120  0.661619477  0.417437570  1.10669270
#> ind8    -1.178708157  3.27864836 -4.820903167 -1.743616702 -0.06602194
#> ind9              NA          NA           NA           NA          NA
#> ind10   -1.383994546  3.21409924  1.505686492 -0.453792787 -1.62940307
#> ind11   -1.834188420  3.85612943  2.666904447 -1.100236722  0.77021497
#> ind12    0.956333513  1.74820671  2.946585648 -1.934975499  0.50932624
#> ind13    3.964663053  0.34582202 -1.167838468  3.581812490  0.41689022
#> ind14   -1.190197706 -1.20881232  0.059859732 -1.206126778 -0.95553208
#> ind15    1.885366600 -1.66751293  1.404730401  2.985606998 -0.59445401
#> ind16    4.553861003 -0.30148850  4.607452471  0.625314978 -1.42472252
#> ind17   -1.481487685  6.51066387  1.736646425 -2.764365514  2.56718036
#> ind18   -1.915510393  1.28618337  1.809167022  3.697511219 -0.80255674
#> ind19    0.804019543 -0.10716929 -1.031691564  3.026917776  1.09104507
#> ind20    0.933764460  0.87285274 -0.948208948  1.640965059 -0.96630320
#> ind21   -3.878243042 -0.51471136  0.278528783 -0.050556696 -4.32292502
#> ind22   -0.815364446 -0.15618409  1.799724881  1.827590621 -1.38895299
#> ind23   -1.980245468 -0.93995067  0.033106598  2.808238640 -0.25333026
#> ind24    1.899215452  1.79472999 -0.851820183  3.241746418 -1.03922906
#> ind25   -3.292599114  1.27013602  2.341536007  0.491475826 -1.18323038
#> ind26   -0.339697973  2.33960907  3.226355535  0.007881556 -0.34605074
#> ind27    0.693057444  0.27787420 -4.257574955  0.516005898 -1.35730288
#> ind28    1.605717825 -3.12950018  0.238499487 -0.465017640  0.62606520
#> ind29    0.972095720  1.64483989 -0.902998025 -1.078621308  2.24037460
#> ind30   -4.123540249 -0.25580978  1.352382549 -0.715008690 -1.27342924
#> ind31   -0.756882906 -2.45796885  0.474715988 -1.719130765  0.81648249
#> ind32    2.103801802 -1.37513082  1.774181495 -1.918004730  1.92638147
#> ind33    1.052656944  0.15352940 -0.330707767 -0.610473858  0.44930029
#> ind34   -2.093617534  1.79646165  0.091791276  0.612507973 -2.10586000
#> ind35    0.001487092  3.10577478  3.775627761  1.186678865  5.51668823
#> ind36   -3.702435209  0.56368024  1.834049081  0.926725262  1.09973200
#> ind37    3.028691410 -4.16153588  0.773141624 -1.168511173 -0.25977022
#> ind38    4.225998309  1.53689086 -0.187522032  1.545367502  3.30623419
#> ind39    2.732337899  0.71736392  1.629757661 -4.442662386 -4.49866844
#> ind40    4.069871265  1.27942901 -2.656474734 -3.528969526  1.79082771
#> ind41    3.251510641  2.27354225 -0.499762468 -0.380458122  5.51953135
#> ind42    1.582556185 -1.03306842 -0.236353221  1.858746224 -0.14108343
#> ind43    0.805835154 -2.01676108 -4.464424497 -2.632830944 -1.22682629
#> ind44   -0.557729404 -2.97530629  0.597689914 -0.475436172  0.36550485
#> ind45   -2.342288678  0.53532508  1.853606706 -0.614677845  1.47587269
#> ind46    0.614520724  1.21274398 -4.779258947  0.241753518  0.86612630
#> ind47   -1.080975727 -1.06883943 -2.807911715  0.057064074  1.36808023
#> ind48    4.559463447 -0.82897610  1.588254230  0.578928136 -2.59800549
#> ind49   -0.667419756  1.24639372 -1.903313821 -0.361027845 -0.62139785
#> ind50   -1.706656885  2.09405001 -1.415018987 -5.427417528 -3.83596640
#> ind51   -2.652642077 -1.32628070  1.719844006  0.838913370 -1.12382327
#> ind52    1.665337759 -2.32248294  0.965966653 -0.088459044  0.38933612
#> ind53   -1.748817318  3.74212806 -0.306027646  0.023567603 -0.58680015
#> ind54   -2.160858066 -1.22644353 -0.345688864 -0.882552790  0.62387571
#> ind55   -1.610351282 -3.06209787  1.538556782 -0.107476754 -0.01337344
#> ind56    1.085705622 -2.12269809  2.777164658  0.439546057 -0.53811424
#> ind57    1.042915155  0.39906388 -0.240831180  0.175830003  3.89757621
#> ind58    0.795264002 -0.05432981 -2.480291691  1.992271075  0.43980000
#> ind59             NA          NA           NA           NA          NA
#> ind60    1.953140336  1.01871304  2.322301567 -0.197384653 -2.81419980
#> ind61   -1.592876708  2.43314147 -0.117469688  1.583286105  0.97014665
#> ind62    1.389647886 -2.51955220 -2.783739531 -1.022648762 -1.69762120
#> ind63   -0.449255611  0.41953590 -0.003825412  1.521122816  1.89344808
#> ind64    0.969515071 -0.19405813  0.874672386 -2.925979089  0.22932180
#> ind65    1.631700851 -0.68019803 -0.487244270 -0.368545061  0.07901129
#> ind66   -2.760054833 -4.33465323 -1.925058809 -0.706528061 -0.84422313
#> ind67   -0.050082526  3.87491004  0.559771687  1.743558788 -1.78328986
#> ind68   -2.697926967 -1.89736178  0.559751999  2.783882312 -1.06301995
#> ind69    1.736024898  3.12165847 -2.354911907  0.859764555 -1.24705676
#> ind70   -3.598828271  0.48228204  2.828626259 -1.084016318  0.29390831
#> ind71   -0.568825546  1.48164900 -2.328636239 -1.834191987 -0.85382567
#> ind72   -1.361169257 -5.58454976  1.137518456  1.714796559 -0.13414516
#> ind73    0.989146617 -0.05182808  1.706282986  1.681710713 -3.42619517
#> ind74    1.517723873  2.20648517 -0.594261290 -0.347360101 -1.43232244
#> ind75    0.393907486 -3.22519428  3.711873058 -1.034147762  0.11299350
#> ind76    1.383852824  1.53337789  3.024399753 -0.178570123  2.20261523
#> ind77   -2.112162231 -1.83264204  1.164321249 -2.368235795  1.50967374
#> ind78    0.398647592  5.90775186 -1.173919793  0.573568040 -2.88052677
#> ind79   -2.002051963 -1.00417846  0.733674568  0.324276418  2.43399301
#> ind80    1.420669177  1.48332849 -3.158563723  4.027373065 -0.15669434
#> ind81   -1.508851903 -1.10253732 -0.809031896 -3.002114294 -0.87193094
#> ind82    2.709188553 -1.93638868 -1.988318982  0.443892673  0.94211538
#> ind83    1.682362188 -1.57086623 -2.971692001  2.372553648 -0.08579004
#> ind84   -0.796257739  1.66731794  1.928074030  0.156929005  1.01880592
#> ind85   -0.346903896 -1.90201949 -4.297438037  4.400867607 -0.20616961
#> ind86   -0.628922163 -1.91277302  1.530012562 -0.291111605 -0.34810538
#> ind87    2.062562135  2.67668955  0.281005510 -3.458667587  2.74218439
#> ind88    0.877031280  1.67715144  0.420160862  0.783394080 -0.68538169
#> ind89   -1.767844988  1.98195181  0.949620502  1.881392884  1.40831737
#> ind90    1.166216654  0.33065298 -2.586012454 -0.045069984 -0.04354356
#> ind91   -1.342682355  0.88520096 -2.261904698 -0.282052147 -0.59247092
#> ind92    6.000412756 -4.33573795  0.043563382 -5.113305293 -0.85670636
#> ind93    4.350499026  0.34214179  1.584984290 -3.016236842  1.67146110
#> ind94    1.604916857 -4.36130992  1.935098465  2.381076233 -0.31645469
#> ind95    0.360986690 -2.11439790  0.806069409 -0.308482081 -1.55141487
#> ind96   -2.054567662 -0.14096351 -0.199382287 -1.018842453 -1.56246248
#> ind97   -1.441429835  1.59945688 -4.904601837 -0.080107928  2.72373293
#> ind98   -0.477538910  2.94722112  0.161938785  0.138110747 -2.22447945
#> ind99    1.935129425 -2.14642545  1.121884063  0.109571302  2.00483456
#> ind100  -2.437872056  2.01681720 -0.607230049 -0.156659392 -3.01764111
#>           prob_pc6    prob_pc7   prob_pc8    prob_pc9   prob_pc10
#> ind1            NA          NA         NA          NA          NA
#> ind2    1.60809731  5.47957974 -1.5020688 -2.30238117  0.48760406
#> ind3   -0.95240578  1.23423268 -0.4346773 -5.04743744  3.30162923
#> ind4   -1.56699010 -1.84211565 -0.1281836 -2.89888147  1.01661399
#> ind5   -1.39721372  0.49096620  0.3080348  1.30744625  1.79272256
#> ind6   -0.12797602  0.17986316  0.6980802 -0.50198383  0.99998918
#> ind7    0.98864649  1.14587385  1.4411941  0.07783394  2.22264589
#> ind8   -0.86190161 -1.71962724  2.8935096 -0.59686867 -1.04347056
#> ind9            NA          NA         NA          NA          NA
#> ind10   1.07965688 -0.29181279 -0.2629679  1.60298506 -0.78164655
#> ind11  -2.62038999  0.72753871  1.2744819  0.20575032 -1.75046920
#> ind12  -1.71756399 -1.22592186 -2.9910651 -0.14208224  0.98452005
#> ind13  -1.11091234  1.32947186 -0.3165891 -2.31219759 -0.18870385
#> ind14   0.81762447 -0.75381616  1.3875168  0.74031976 -1.38841880
#> ind15  -0.81924807  0.41052832 -1.0627974  0.82696573  1.97099429
#> ind16   2.75227656 -1.62295561  1.0833665 -1.02198671  1.59370188
#> ind17   1.27244509 -0.63621810 -3.0660286 -1.76610815  1.75105588
#> ind18   0.31280416  2.62230690 -1.5597934 -1.30862902  2.23054899
#> ind19   3.20544593 -0.77097700 -0.6046576  0.99271784  1.34799210
#> ind20  -0.02456873  2.08638880  0.1670819 -0.79058329  0.77993663
#> ind21   0.44346329 -1.33211955  2.8512697  1.22746380 -1.11213291
#> ind22  -3.13807186 -1.05572186 -1.2558909 -0.13068262 -1.99876376
#> ind23   2.70751451  0.43948986 -1.2653387 -0.22366998  0.56825628
#> ind24  -1.99289269  1.53618536  1.7327826 -0.41770047 -0.44329218
#> ind25  -1.60026778  2.20072974  3.4385695 -0.17291629  0.49830712
#> ind26   1.51812482 -1.11804016 -2.2193205 -0.82147674  1.92480168
#> ind27  -1.44824814 -2.93524248  0.8912031  0.94752727  1.51071904
#> ind28  -1.37120419 -0.75071762  2.2721070 -0.21763524 -1.52983482
#> ind29   1.17173747 -2.12349137 -0.5647785  2.28152743  0.26318966
#> ind30  -2.45876167  0.59393118 -0.5398526  2.06110149  3.66088771
#> ind31   0.13273268 -1.07706201 -2.2852762 -0.15606442 -2.45190471
#> ind32   0.80264140  0.24147345 -0.9020145 -0.33218128  0.08821189
#> ind33  -0.52462143 -0.13257456  3.3337851  0.95629700  0.97475775
#> ind34   0.56006920  0.17999160  1.9458518 -0.28863975  2.56712059
#> ind35   2.03078145  3.40583280  1.6162738 -0.70608255 -4.97345683
#> ind36   0.31741247 -1.46197680 -0.8688482 -1.62565864 -1.49015290
#> ind37  -0.46260846  4.74641286 -1.3180651  1.75162442 -1.52192169
#> ind38   0.69430409  1.46517029 -0.3988092 -1.90532055 -0.58870669
#> ind39  -0.47501568  2.17148027  0.4395753  0.56780229 -0.95909575
#> ind40   1.74476407 -3.46071174  1.4435708 -0.62012087 -0.95066286
#> ind41  -0.44733251 -1.95371335  2.0574692  0.28349039  3.01777659
#> ind42  -0.02669622  0.05143142 -1.5589905 -0.23293592 -0.66297478
#> ind43  -1.37999370 -0.65827426 -2.6557535 -2.50504123 -0.21293676
#> ind44  -2.82267198 -1.56237647 -0.6596315 -1.72528267 -1.33847246
#> ind45   3.47541591  0.26174299  0.2080938  1.77356465  0.97219871
#> ind46   1.45227699  0.35932333 -1.6447886 -0.97413903  2.56488903
#> ind47  -2.94499705  0.23021974 -0.5161323  1.91012013  0.02731035
#> ind48   0.78690915  0.78659603 -1.1934960  0.05487126 -2.41774769
#> ind49  -1.71706938  0.14786245 -1.2964513  3.06991606  0.31482504
#> ind50   5.83628682 -2.24534802  0.5957090 -0.04331148  0.20609593
#> ind51   1.25713119 -2.09032472  2.9725534 -1.40689811  1.09990374
#> ind52  -1.33279956 -0.36599852 -0.7087473 -0.25807598 -0.29663539
#> ind53   1.91258398  1.35684482 -0.3760242  0.24201684  0.59707275
#> ind54  -0.50028837  0.64216178  1.9183088  0.22787975  0.47826372
#> ind55   2.56719411 -3.10865244 -0.4249778  0.61640523 -1.50529460
#> ind56   2.04726448  0.68104389  2.1967692 -0.73182650 -0.89092891
#> ind57  -1.11514718  0.74395448 -1.0656307  0.02539372 -1.49136467
#> ind58  -0.22731160 -2.01506255 -0.8679451 -1.17683093 -0.90241730
#> ind59           NA          NA         NA          NA          NA
#> ind60  -0.07871137  1.05143842  0.1397083 -1.93479248 -0.81171834
#> ind61   2.50634657  0.52386896  1.6541591  2.61481017 -0.39170848
#> ind62  -1.84746570  1.33545008  3.1052421 -0.54668973 -0.32905726
#> ind63  -1.33445333 -1.34638962  2.4680064  0.18181982  1.21636950
#> ind64  -1.79038846  1.68964250 -0.9466567  2.20310235 -1.60610974
#> ind65  -2.24434115 -1.32133783  0.9559765  0.29730654  3.04084416
#> ind66   0.75934306  0.06140612  0.3352470  0.21770684 -1.60159577
#> ind67  -0.13236090  0.90381840 -0.2832443  2.03668111  1.98333551
#> ind68  -2.35492040  0.44457651 -0.5135818  1.18683262  1.13742788
#> ind69  -2.96942602 -0.20605923  0.2642206 -0.74596241 -1.33982724
#> ind70  -3.88495957 -3.13277763  0.4642698 -2.39839928 -0.02269123
#> ind71  -0.07508857  3.22502646 -1.6754050 -1.17983369  1.04746401
#> ind72   0.20023374 -2.61507408 -1.5761835  0.02811606  0.04349372
#> ind73   1.67567880 -0.85583580 -3.1782999  1.44005610  0.27062163
#> ind74   2.68249065  0.13178751 -0.7273356  0.27016974  1.46756441
#> ind75  -1.60217132 -0.65930044 -0.1838584 -0.18702978  2.41942236
#> ind76  -2.44390400 -0.81975285 -4.0963803  4.49405158  0.42614087
#> ind77  -2.85752029 -0.49011387  2.2038855  1.57176790 -1.37425193
#> ind78  -1.49629432  3.91540308  0.5875496 -0.61560679 -0.35786090
#> ind79  -0.05095800 -3.14155549  2.2378946  2.61318931 -0.46871034
#> ind80  -1.67634578  0.65635340 -1.4571546  2.13374926 -1.46068371
#> ind81   1.00007663 -0.52921889 -3.1661441  3.06985980  1.50394765
#> ind82   1.17580440  1.98126096  2.6988964  2.95942837 -0.96240118
#> ind83   2.58724412  1.88621777  0.5849883  2.08606778 -0.64389641
#> ind84  -0.26101920 -0.66642017 -0.6494668 -0.58036228 -1.61864566
#> ind85   1.35718442 -3.68464257 -2.5779972 -0.92296652 -2.06576685
#> ind86   0.81683260  0.91404217  1.3709429 -1.18771611 -0.11074159
#> ind87  -1.76501551 -0.34483309  0.4337960 -1.87793608 -2.07514692
#> ind88  -0.61506335 -0.70612462  0.7672533  2.69739735 -1.34102657
#> ind89  -0.60718412 -0.96196437 -0.9810253 -1.88330556 -1.32702677
#> ind90   2.58996781  1.72807192  2.4621033 -2.44252946  0.94651957
#> ind91   2.86102252 -1.37865871 -1.5202764 -1.26406423 -3.70137261
#> ind92   0.10652986  3.78737213 -2.0910664  1.36135437  0.44686181
#> ind93   0.84013928 -0.86537570  1.0885864 -1.84634760  2.02919105
#> ind94   1.27940632 -0.11820727  2.0307169  1.36719928  0.35769056
#> ind95   2.57016186  1.02237665 -0.1406304 -1.01612568 -2.40947262
#> ind96  -1.05888958  0.55251068  0.6706102  0.06655287  0.45758907
#> ind97   0.39771646  0.46562781 -0.9610401  1.92381256 -0.13484667
#> ind98  -0.98224200  0.12000426 -0.9597286 -4.00666986 -0.43698900
#> ind99   1.69368259 -1.04931877 -0.4961318 -0.25736460  0.92227557
#> ind100 -1.51674216 -1.14116384 -1.2782164 -0.64953538 -1.99227020