Retrieve Sample Summary Data

Extracts sample-level summary statistics from a Metabolites object for a specified data layer, with optional exclusions and subsetting.

Usage

get_sample_summary(
  metabolites,
  layer,
  sample_ids = NULL,
  drop = FALSE,
  as_df = FALSE
)

Arguments

metabolites

A Metabolites object containing sample summary data.

layer

A character string specifying the data layer from which to extract sample summary data.

sample_ids

Optional. A vector of column names (samples) to extract. Defaults to NULL, meaning all samples are included.

drop

Logical. Whether to drop all NA (excluded samples)

as_df

Logical. Whether to return the result as a data.frame. Defaults to FALSE, meaning the result is returned as a matrix.

Value

A matrix or data.frame (if as_df = TRUE) containing the sample summary data for the specified layer.

Examples

file <- system.file("extdata", "metabolon_v1_example.xlsx", package = "metaboprep2")
metabolites <- Metabolites(project_name = "MYPROJECT",
                           format       = "metabolon_v1",
                           filepath     = file)
metabolites <- import_data(metabolites)
#> Using cached MetaboAnalyst DB.
#> Attempting join [features$hmdb] ~ [annotation$hmdb_id] 
#> Attempting join [features$kegg] ~ [annotation$kegg_id] 
#> Attempting join [features$metabolite_id] ~ [annotation$name] 
#> Attempting join [features$comp_id] ~ [annotation$comp_id] 
#> Warning: 100 unannotated metabolites remain after merging with metabolomics database
metabolites <- metabolite_qc(metabolites, source="raw")
#> 
#> ── Starting Metabolite QC Process ──────────────────────────────────────────────
#> ℹ Validating input parameters...
#> ✔ Input validation complete.
#> ℹ Batch normalising raw data...
#> ✔ Normalised raw data.
#> ℹ Applying exclusion of xenobiotic features...
#> ✔ Xenobiotic features exclusions assessment complete. (3 features excluded)
#> ℹ Analysing exclusions...
#> ℹ Applying exclusion of derived features...
#> ✔ Derived features exclusions assessment complete. (0 features excluded)
#> ℹ Sample & Feature Summary Statistics for raw data...
#> ✔ Sample & Feature Summary Statistics for raw data...
#> ℹ Copying batch_normalised data to new 'qcing' data layer...
#> ✔ qcing data ready for QC.
#> ℹ Assessing for extreme sample missingness >=80%...
#> ✔ Extreme sample missingness exclusions assessment complete. (0 samples excluded)
#> ℹ Assessing for total feature missingness (100%)
#> ✔ Total feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Assessing for extreme feature missingness >=80%...
#> ✔ Extreme feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Assessing for sample missingness at specified level of >=20%...
#> ✔ Sample missingness exclusions assessment complete. (2 samples excluded)
#> ℹ Assessing for feature missingness at specified level of >=20%...
#> ✔ Feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Calculating total peak abundance outliers at +/- 5 Sdev...
#> ✔ Total peak abundance outlier exclusions assessment complete. (0 samples excluded)
#> ℹ Running data outlier analysis at +/- 5 Sdev...
#> ℹ Data outlier analysis complete.
#> ℹ Re-identify feature independence and PC outliers...
#> ℹ Re-running sample summary using current exclusions...
#> ℹ QC sample summary complete.
#> ℹ PCA outlier analysis 2 complete.
#> ℹ Creating final QC dataset...
#> ℹ Rerunning Sample & Feature Summary Statistics for QC data...
#> ✔ Sample & Feature Summary Statistics for QC data complete.
#> 
#> ── Metabolite QC Process Completed ─────────────────────────────────────────────
get_sample_summary(metabolites, layer = "post_qc")
#>        missingness missingness_w_exclusions tpa_total tpa_complete_features
#> ind1            NA                       NA        NA                    NA
#> ind2          0.00                        0  508.7894              492.8786
#> ind3          0.02                        0  353.3467              346.9899
#> ind4          0.00                        0  381.7751              369.9670
#> ind5          0.00                        0  403.2351              391.6733
#> ind6          0.00                        0  373.1347              360.5262
#> ind7          0.00                        0  382.8729              371.2553
#> ind8          0.00                        0  401.7967              389.7508
#> ind9            NA                       NA        NA                    NA
#> ind10         0.00                        0  410.6914              399.8748
#> ind11         0.00                        0  401.4834              388.0937
#> ind12         0.01                        0  366.9062              360.2909
#> ind13         0.00                        0  347.2292              335.9061
#> ind14         0.00                        0  388.8691              377.9240
#> ind15         0.01                        0  347.5606              340.4478
#> ind16         0.00                        0  341.6616              330.0813
#> ind17         0.00                        0  417.3087              406.5103
#> ind18         0.01                        0  399.8966              392.8238
#> ind19         0.01                        0  361.1083              353.9565
#> ind20         0.00                        0  386.0828              374.4718
#> ind21         0.01                        0  400.4804              392.5644
#> ind22         0.01                        0  379.2166              371.2341
#> ind23         0.01                        0  389.6081              384.0383
#> ind24         0.01                        0  371.6540              364.8369
#> ind25         0.01                        0  401.0630              395.0934
#> ind26         0.02                        0  390.2171              386.9443
#> ind27         0.01                        0  367.1512              364.4205
#> ind28         0.01                        0  354.9185              347.5067
#> ind29         0.00                        0  381.2393              368.8625
#> ind30         0.00                        0  413.8298              400.3422
#> ind31         0.01                        0  376.4718              369.8023
#> ind32         0.01                        0  364.1226              354.5883
#> ind33         0.00                        0  370.7611              360.1066
#> ind34         0.01                        0  401.5134              393.0931
#> ind35         0.00                        0  391.2666              379.7223
#> ind36         0.00                        0  409.8471              396.9707
#> ind37         0.00                        0  349.3609              337.6469
#> ind38         0.01                        0  358.6038              350.3004
#> ind39         0.01                        0  370.0657              364.0988
#> ind40         0.01                        0  355.9358              349.2635
#> ind41         0.00                        0  359.9534              348.3273
#> ind42         0.00                        0  367.8827              355.7330
#> ind43         0.00                        0  374.3105              359.2484
#> ind44         0.01                        0  359.0955              354.2752
#> ind45         0.00                        0  394.4488              385.7001
#> ind46         0.01                        0  378.5787              373.4887
#> ind47         0.00                        0  374.2399              363.6766
#> ind48         0.00                        0  347.6765              336.9237
#> ind49         0.00                        0  390.7546              376.1562
#> ind50         0.00                        0  430.2432              417.6961
#> ind51         0.00                        0  400.8457              391.6795
#> ind52         0.01                        0  348.5621              340.2470
#> ind53         0.00                        0  403.9982              392.5695
#> ind54         0.01                        0  403.0875              397.6857
#> ind55         0.00                        0  380.6361              373.7841
#> ind56         0.00                        0  365.8299              356.3817
#> ind57         0.00                        0  377.1005              364.1852
#> ind58         0.00                        0  369.4984              359.4607
#> ind59           NA                       NA        NA                    NA
#> ind60         0.00                        0  376.5645              365.6023
#> ind61         0.00                        0  401.5256              390.5735
#> ind62         0.01                        0  354.9646              346.3308
#> ind63         0.00                        0  378.8121              368.8956
#> ind64         0.00                        0  369.2687              356.5206
#> ind65         0.00                        0  360.4123              350.3926
#> ind66         0.01                        0  390.3517              382.2087
#> ind67         0.00                        0  383.0492              373.3305
#> ind68         0.00                        0  384.5966              372.9140
#> ind69         0.00                        0  378.3576              366.4013
#> ind70         0.00                        0  392.7291              381.1438
#> ind71         0.01                        0  396.1722              385.4032
#> ind72         0.00                        0  365.8605              353.6817
#> ind73         0.01                        0  360.5525              353.4829
#> ind74         0.01                        0  387.0433              378.7725
#> ind75         0.00                        0  374.8655              361.5894
#> ind76         0.01                        0  372.5086              363.1284
#> ind77         0.00                        0  404.1672              391.4608
#> ind78         0.00                        0  398.6028              386.7687
#> ind79         0.00                        0  400.2317              389.6509
#> ind80         0.00                        0  363.4649              351.0852
#> ind81         0.02                        0  385.7676              381.0244
#> ind82         0.00                        0  349.4534              337.6808
#> ind83         0.01                        0  351.5077              344.1981
#> ind84         0.00                        0  394.1948              385.0049
#> ind85         0.00                        0  371.1594              359.2531
#> ind86         0.02                        0  384.6302              380.5426
#> ind87         0.00                        0  378.7804              368.0198
#> ind88         0.01                        0  363.1606              355.3499
#> ind89         0.01                        0  402.4704              395.2352
#> ind90         0.01                        0  372.2764              366.8107
#> ind91         0.00                        0  411.6871              397.4729
#> ind92         0.01                        0  326.8889              316.2764
#> ind93         0.00                        0  364.3073              352.7148
#> ind94         0.00                        0  350.6093              341.0099
#> ind95         0.01                        0  379.3109              371.9853
#> ind96         0.01                        0  401.8077              394.1858
#> ind97         0.00                        0  394.2284              381.1877
#> ind98         0.01                        0  403.1847              393.9233
#> ind99         0.01                        0  354.0465              345.6860
#> ind100        0.00                        0  401.4720              389.1460
#>        outlier_count
#> ind1              NA
#> ind2               1
#> ind3               0
#> ind4               0
#> ind5               0
#> ind6               0
#> ind7               0
#> ind8               0
#> ind9              NA
#> ind10              0
#> ind11              0
#> ind12              0
#> ind13              0
#> ind14              0
#> ind15              0
#> ind16              0
#> ind17              0
#> ind18              0
#> ind19              0
#> ind20              0
#> ind21              0
#> ind22              0
#> ind23              0
#> ind24              0
#> ind25              0
#> ind26              0
#> ind27              0
#> ind28              0
#> ind29              0
#> ind30              0
#> ind31              0
#> ind32              0
#> ind33              0
#> ind34              0
#> ind35              0
#> ind36              0
#> ind37              0
#> ind38              0
#> ind39              0
#> ind40              0
#> ind41              0
#> ind42              0
#> ind43              0
#> ind44              0
#> ind45              0
#> ind46              0
#> ind47              0
#> ind48              0
#> ind49              0
#> ind50              0
#> ind51              0
#> ind52              0
#> ind53              0
#> ind54              0
#> ind55              0
#> ind56              0
#> ind57              0
#> ind58              0
#> ind59             NA
#> ind60              0
#> ind61              0
#> ind62              0
#> ind63              0
#> ind64              0
#> ind65              0
#> ind66              0
#> ind67              0
#> ind68              0
#> ind69              0
#> ind70              0
#> ind71              0
#> ind72              0
#> ind73              0
#> ind74              0
#> ind75              0
#> ind76              0
#> ind77              0
#> ind78              0
#> ind79              0
#> ind80              0
#> ind81              0
#> ind82              0
#> ind83              0
#> ind84              0
#> ind85              0
#> ind86              0
#> ind87              0
#> ind88              0
#> ind89              0
#> ind90              0
#> ind91              0
#> ind92              0
#> ind93              0
#> ind94              0
#> ind95              0
#> ind96              0
#> ind97              0
#> ind98              0
#> ind99              0
#> ind100             0