Retrieve Variance Explained by PCs

Extracts the variance explained by each principal components (PCs) from a Metabolites object for a specified data layer, with optional exclusions and subsetting.

Usage

get_var_exp(metabolites, layer, as_df = FALSE)

Arguments

metabolites

A Metabolites object containing the principal component data.

layer

A character string specifying the data layer from which to extract the PCs.

as_df

Logical. Whether to return the result as a data.frame. Defaults to FALSE, returning a matrix.

Value

A matrix or data.frame (if as_df = TRUE) containing the principal components.

Examples

file <- system.file("extdata", "metabolon_v1_example.xlsx", package = "metaboprep2")
metabolites <- Metabolites(project_name = "MYPROJECT",
                           format       = "metabolon_v1",
                           filepath     = file)
metabolites <- import_data(metabolites)
#> Warning: 86 unmapped features:
#>  - 123, 124, 125, 126, 127, 128, 129, 130, 131, 132, 133, 134, 135, 137, 138, 139, 140, 141, 142, 143, 145, 146, 147, 148, 149, 150, 151, 152, 153, 154, 155, 156, 157, 159, 160, 161, 162, 163, 164, 165, 166, 167, 168, 169, 170, 172, 173, 174, 175, 176, 177, 178, 183, 184, 185, 186, 187, 188, 190, 191, 193, 195, 196, 198, 199, 200, 201, 202, 204, 205, 206, 207, 208, 209, 210, 211, 213, 214, 215, 216, 217, 218, 219, 220, 221, 222
metabolites <- metabolite_qc(metabolites, source="raw")
#> 
#> ── Starting Metabolite QC Process ──────────────────────────────────────────────
#> ℹ Validating input parameters...
#> ✔ Input validation complete.
#> ℹ Sample & Feature Summary Statistics for raw data...
#> ✔ Sample & Feature Summary Statistics for raw data...
#> ℹ Copying raw data to new 'qcing' data layer...
#> ✔ qcing data ready for QC.
#> ℹ Analysing exclusions...
#> ℹ Applying exclusion of derived features...
#> ✔ Derived features exclusions assessment complete. (0 features excluded)
#> ℹ Applying exclusion of xenobiotic features...
#> ✔ Xenobiotic features exclusions assessment complete. (3 features excluded)
#> ℹ Assessing for extreme sample missingness >=80%...
#> ✔ Extreme sample missingness exclusions assessment complete. (0 samples excluded)
#> ℹ Assessing for extreme feature missingness >=80%...
#> ✔ Extreme feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Assessing for sample missingness at specified level of >=20%...
#> ✔ Sample missingness exclusions assessment complete. (2 samples excluded)
#> ℹ Assessing for feature missingness at specified level of >=20%...
#> ✔ Feature missingness exclusions assessment complete. (0 features excluded)
#> ℹ Calculating total peak abundance outliers at +/- 5 Sdev...
#> ✔ Total peak abundance outlier exclusions assessment complete. (0 samples excluded)
#> ℹ Running data outlier analysis at +/- 5 Sdev...
#> ℹ Data outlier analysis complete.
#> ℹ Re-running sample summary using current exclusions...
#> ℹ QC sample summary complete.
#> ℹ Re-identify feature independence and PC outliers...
#> ℹ PCA outlier analysis 2 complete.
#> ℹ Creating final QC dataset...
#> ℹ Rerunning Sample & Feature Summary Statistics for QC data...
#> ✔ Sample & Feature Summary Statistics for QC data complete.
#> 
#> ── Metabolite QC Process Completed ─────────────────────────────────────────────
get_var_exp(metabolites, layer = "post_qc")
#>     PC1     PC2     PC3     PC4     PC5     PC6     PC7     PC8     PC9    PC10 
#> 0.12818 0.09026 0.04668 0.04097 0.03733 0.03248 0.03224 0.03017 0.02768 0.02702 
#>    PC11    PC12    PC13    PC14    PC15    PC16    PC17    PC18    PC19    PC20 
#> 0.02573 0.02435 0.02362 0.02202 0.02137 0.01896 0.01892 0.01779 0.01715 0.01638 
#>    PC21    PC22    PC23    PC24    PC25    PC26    PC27    PC28    PC29    PC30 
#> 0.01565 0.01516 0.01425 0.01403 0.01332 0.01274 0.01207 0.01174 0.01116 0.01065 
#>    PC31    PC32    PC33    PC34    PC35    PC36    PC37    PC38    PC39    PC40 
#> 0.00993 0.00982 0.00885 0.00849 0.00836 0.00809 0.00726 0.00714 0.00670 0.00667 
#>    PC41    PC42    PC43    PC44    PC45    PC46    PC47    PC48    PC49    PC50 
#> 0.00628 0.00569 0.00557 0.00522 0.00490 0.00476 0.00431 0.00407 0.00362 0.00356 
#>    PC51    PC52    PC53    PC54    PC55    PC56    PC57    PC58    PC59    PC60 
#> 0.00344 0.00332 0.00299 0.00281 0.00258 0.00252 0.00239 0.00217 0.00204 0.00177 
#>    PC61    PC62    PC63    PC64    PC65    PC66    PC67    PC68    PC69    PC70 
#> 0.00174 0.00160 0.00152 0.00130 0.00122 0.00115 0.00098 0.00093 0.00083 0.00065 
#>    PC71    PC72    PC73    PC74    PC75    PC76    PC77 
#> 0.00060 0.00053 0.00041 0.00038 0.00033 0.00025 0.00020